Summarize DAS effort

Chop DAS data into effort segments

das_effort(x, ...)

# S3 method for data.frame
das_effort(x, ...)

# S3 method for das_df
das_effort(
  x,
  method = c("condition", "equallength", "section"),
  conditions = NULL,
  distance.method = c("greatcircle", "lawofcosines", "haversine", "vincenty"),
  seg0.drop = FALSE,
  comment.drop = FALSE,
  event.touse = NULL,
  num.cores = NULL,
  ...
)

Arguments

x
an object of class das_df, or a data frame that can be coerced to class das_df

...
arguments passed to the specified chopping function, such as seg.km or seg.min.km

method
character; method to use to chop DAS data into effort segments Can be "condition", "equallength", "section", or any partial match thereof (case sensitive)

conditions
character vector of names of conditions to include in segdata output. These values must be column names from the output of das_process, e.g. 'Bft', 'SwellHght', etc. If method == "condition", then these also are the conditions which trigger segment chopping when they change.

distance.method
character; method to use to calculate distance between lat/lon coordinates. Can be "greatcircle", "lawofcosines", "haversine", "vincenty", or any partial match thereof (case sensitive). Default is "greatcircle"

seg0.drop
logical; flag indicating whether or not to drop segments of length 0 that contain no sighting (S, K, M, G, t) events. Default is FALSE

comment.drop
logical; flag indicating if comments ("C" events) should be ignored (i.e. position information should not be used) when segment chopping. Default is FALSE

event.touse
character vector of events to use to determine segment lengths; overrides comment.drop. If NULL (the default), then all on effort events are used. If used, this argument must include at least R, E, S, and A events, and cannot include ? or 1:8 events

num.cores
Number of CPUs to over which to distribute computations. Defaults to NULL, which uses one fewer than the number of cores reported by detectCores. Using 1 core likely will be faster for smaller datasets

Value

List of three data frames:

segdata: one row for every segment, and columns for information including unique segment number, start/end/midpoint coordinates, and conditions (e.g. Beaufort)

sightinfo: details for all sightings in x, including: the unique segment number it is associated with, segment mid points (lat/lon), the 'included' column described in the 'Details' section, and the output information described in das_sight

randpicks: see das_chop_equallength; NULL if using "condition" method

Details

This is the top-level function for chopping processed DAS data into modeling segments (henceforth 'segments'), and assigning sightings and related information (e.g., weather conditions) to each segment. This function returns data frames with all relevant information for the effort segments and associated sightings ('segdata' and 'sightinfo', respectively). Before chopping, the DAS data is filtered for events (rows) where either the 'OnEffort' column is TRUE or the 'Event' column "E". In other words, the data is filtered for continuous effort sections (henceforth 'effort sections'), where effort sections run from "R" to "E" events (inclusive), and then passed to the chopping function specified using method. Note that while B events immediately preceding an R are on effort, they are ignored during effort chopping. In addition, all on effort events (other than ? and numeric events) with NA DateTime, Lat, or Lon values are verbosely removed.

The following chopping methods are currently available: "condition", "equallength", and "section. When using the "condition" method, effort sections are chopped into segments every time a condition changes, thereby ensuring that the conditions are consistent across the entire segment. See das_chop_condition for more details about this method, including arguments that must be passed to it via the argument ...

The "equallength" method consists of chopping effort sections into equal-length segments of length seg.km, and doing a weighted average of the conditions for the length of that segment. See das_chop_equallength for more details about this method, including arguments that must be passed to it via the argument ...

The "section" method involves 'chopping' the effort into continuous effort sections, i.e. each continuous effort section is a single effort segment. See das_chop_section for more details about this method.

The distance between the lat/lon points of subsequent events is calculated using the method specified in distance.method. If "greatcircle", distance_greatcircle is used, while distance is used otherwise. See das_sight for how the sightings are processed.

The sightinfo data frame includes the column 'included', which is used in das_effort_sight when summarizing the number of sightings and animals for selected species. das_effort_sight is a separate function to allow users to personalize the included values as desired for their analysis. By default, i.e. in the output of this function, 'included' is TRUE if: the sighting was made when on effort, by a standard observer (see das_sight), and in a Beaufort sea state less than or equal to five.

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

# Using "condition" method
das_effort(
  y.proc, method = "condition", conditions = c("Bft", "SwellHght", "Vis"),
  seg.min.km = 0.05, num.cores = 1
)

#> $segdata
#>    segnum section_id section_sub_id           file stlin endlin     lat1
#> 1       1          1              1 das_sample.das     2     20 43.94750
#> 2       2          2              1 das_sample.das    23     33 44.00333
#> 3       3          2              2 das_sample.das    33     43 44.05667
#> 4       4          3              1 das_sample.das    59     69 44.19517
#> 5       5          3              2 das_sample.das    69     70 44.28798
#> 6       6          3              3 das_sample.das    70     75 44.28850
#> 7       7          3              4 das_sample.das    75     78 44.30067
#> 8       8          3              5 das_sample.das    78     84 44.31766
#> 9       9          3              6 das_sample.das    84     85 44.36631
#> 10     10          3              7 das_sample.das    85     90 44.37829
#> 11     11          4              1 das_sample.das    99    108 44.57233
#> 12     12          4              2 das_sample.das   108    121 44.59617
#> 13     13          5              1 das_sample.das   127    147 44.77933
#> 14     14          6              1 das_sample.das   150    160 44.89583
#> 15     15          6              2 das_sample.das   160    164 44.94750
#> 16     16          7              1 das_sample.das   167    174 45.00967
#> 17     17          7              2 das_sample.das   174    181 45.05681
#> 18     18          8              1 das_sample.das   188    199 45.14917
#> 19     19          9              1 das_sample.das   232    240 45.61433
#> 20     20         10              1 das_sample.das   242    259 45.65100
#>         lon1     lat2      lon2     mlat      mlon    dist           mDateTime
#> 1  -141.0217 43.99433 -140.9990 43.97093 -141.0104  5.5117 2014-03-05 06:37:02
#> 2  -141.0152 44.05667 -140.9888 44.03004 -141.0021  6.2888 2014-03-05 07:09:03
#> 3  -140.9888 44.14649 -140.9450 44.10152 -140.9667 10.5778 2014-03-05 07:38:33
#> 4  -140.8703 44.28798 -140.8305 44.24150 -140.8501 10.7925 2014-03-05 09:40:55
#> 5  -140.8305 44.28850 -140.8303 44.28825 -140.8304  0.0571 2014-03-05 09:59:44
#> 6  -140.8303 44.30067 -140.8253 44.29460 -140.8279  1.4100 2014-03-05 10:02:12
#> 7  -140.8253 44.31766 -140.8183 44.30917 -140.8218  1.9694 2014-03-05 10:07:52
#> 8  -140.8183 44.36631 -140.8205 44.34220 -140.8094  5.6458 2014-03-05 10:20:48
#> 9  -140.8205 44.37829 -140.8265 44.37233 -140.8235  1.4160 2014-03-05 10:32:51
#> 10 -140.8265 44.38150 -140.8280 44.37991 -140.8273  0.3716 2014-03-05 10:35:50
#> 11 -140.7865 44.59617 -140.7715 44.58426 -140.7790  2.9023 2014-03-05 11:57:10
#> 12 -140.7715 44.75460 -140.6661 44.67576 -140.7198 19.4871 2014-03-05 12:39:33
#> 13 -140.5910 44.89333 -140.5522 44.83613 -140.5704 13.0345 2014-03-05 14:14:10
#> 14 -140.5442 44.94750 -140.5282 44.92165 -140.5360  5.8777 2014-03-05 15:09:52
#> 15 -140.5282 45.00312 -140.5089 44.97547 -140.5192  6.3723 2014-03-05 15:31:47
#> 16 -140.5150 45.05681 -140.4919 45.03334 -140.5038  5.5485 2014-03-05 16:09:21
#> 17 -140.4919 45.07850 -140.4802 45.06769 -140.4861  2.5759 2014-03-05 16:24:56
#> 18 -140.4707 45.15250 -140.4688 45.15082 -140.4697  0.3974 2014-03-05 17:00:37
#> 19 -139.0153 45.64298 -138.9955 45.62866 -139.0054  3.5391 2014-03-06 11:30:58
#> 20 -138.9917 45.67316 -138.9768 45.66216 -138.9845  2.7200 2014-03-06 11:45:33
#>    year month day    mtime Cruise Mode EffType ESWsides maxdistBft
#> 1  2014     3   5 06:37:02   1000    C       S        2          3
#> 2  2014     3   5 07:09:03   1000    C       S        2          3
#> 3  2014     3   5 07:38:33   1000    C       S        2          3
#> 4  2014     3   5 09:40:55   1000    C       S        2          3
#> 5  2014     3   5 09:59:44   1000    C       S        2          3
#> 6  2014     3   5 10:02:12   1000    C       S        2          2
#> 7  2014     3   5 10:07:52   1000    C       S        2          2
#> 8  2014     3   5 10:20:48   1000    C       S        2          2
#> 9  2014     3   5 10:32:51   1000    C       S        2          2
#> 10 2014     3   5 10:35:50   1000    C       S        2          2
#> 11 2014     3   5 11:57:10   1000    C       S        2          3
#> 12 2014     3   5 12:39:33   1000    C       S        2          3
#> 13 2014     3   5 14:14:10   1000    C       S        2          3
#> 14 2014     3   5 15:09:52   1000    C       S        2          3
#> 15 2014     3   5 15:31:47   1000    C       S        2          2
#> 16 2014     3   5 16:09:21   1000    C       S        2          3
#> 17 2014     3   5 16:24:56   1000    C       S        2          2
#> 18 2014     3   5 17:00:37   1000    C       S        2          2
#> 19 2014     3   6 11:30:58   1000    C       S        2          2
#> 20 2014     3   6 11:45:33   1000    C       S        2          2
#>    maxdistSwellHght maxdistVis
#> 1                 3        6.0
#> 2                 3        6.0
#> 3                 3        5.5
#> 4                 3        5.5
#> 5                 3        6.0
#> 6                 3        6.0
#> 7                 3        5.5
#> 8                 3        4.5
#> 9                 3        3.5
#> 10                3        2.5
#> 11                3        5.8
#> 12                3        6.0
#> 13                3        6.0
#> 14                3        6.0
#> 15                3        6.0
#> 16                3        6.0
#> 17                3        6.0
#> 18                3        6.0
#> 19                1        4.0
#> 20                1        4.0
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime year      Lat       Lon
#> 1       1 43.97093 -141.0104     S 2014-03-05 06:46:02 2014 43.99333 -140.9993
#> 2       3 44.10152 -140.9667     S 2014-03-05 07:56:22 2014 44.14483 -140.9458
#> 3       4 44.24150 -140.8501     t 2014-03-05 09:34:27 2014 44.22450 -140.8573
#> 4      13 44.83613 -140.5704     t 2014-03-05 14:02:55 2014 44.81000 -140.5795
#> 5      13 44.83613 -140.5704     S 2014-03-05 14:37:56 2014 44.89283 -140.5523
#> 6      15 44.97547 -140.5192     t 2014-03-05 15:36:47 2014 44.98767 -140.5152
#> 7      17 45.06769 -140.4861     S 2014-03-05 16:29:50 2014 45.07850 -140.4802
#> 8      18 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 9      18 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 10     19 45.62866 -139.0054     F 2014-03-06 11:25:32 2014 45.61667 -139.0138
#> 11     20 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 12     20 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 13     20 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#> 14     20 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#>    OnEffort Cruise Mode EffType ESWsides Course SpdKt Bft SwellHght WindSpdKt
#> 1      TRUE   1000    C       S        2     25  10.2   3         3        10
#> 2      TRUE   1000    C       S        2     26   9.7   3         3        10
#> 3      TRUE   1000    C       S        2     27   9.0   3         3        10
#> 4      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 5      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 6      TRUE   1000    C       S        2     16   8.9   2         3         6
#> 7      TRUE   1000    C       S        2     25   8.9   2         3         6
#> 8      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 9      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 10     TRUE   1000    C       S        2     35   9.5   2         1         5
#> 11     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 12     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 13     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 14     TRUE   1000    C       S        2     23   9.6   2         1         5
#>    RainFog HorizSun VertSun Glare Vis ObsL Rec ObsR ObsInd EffortDot EventNum
#> 1        1       NA      NA    NA 6.0  208 280  001   <NA>      TRUE       15
#> 2        3        2       2 FALSE 5.5  125 208  280   <NA>      TRUE       35
#> 3        1        2       2 FALSE 5.5  001 126  149   <NA>      TRUE       59
#> 4        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      131
#> 5        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      136
#> 6        1        9       1 FALSE 6.0  125 208  280   <NA>      TRUE      153
#> 7        1        8       2 FALSE 6.0  149 125  208   <NA>      TRUE      167
#> 8        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 9        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 10       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE        8
#> 11       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 12       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 13       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#> 14       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#>          file_das line_num SightNo Subgroup SightNoDaily Obs ObsStd Bearing
#> 1  das_sample.das       15    1406     <NA>   20140305_1 208   TRUE     309
#> 2  das_sample.das       38    1407     <NA>   20140305_2 125   TRUE     326
#> 3  das_sample.das       65    <NA>     <NA>         <NA> 280  FALSE     120
#> 4  das_sample.das      137    <NA>     <NA>         <NA> 228  FALSE     300
#> 5  das_sample.das      142    1408     <NA>   20140305_3 280   TRUE     270
#> 6  das_sample.das      162    <NA>     <NA>         <NA> 231  FALSE      45
#> 7  das_sample.das      176    1409     <NA>   20140305_4 149   TRUE     344
#> 8  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 9  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 10 das_sample.das      238    <NA>     <NA>         <NA> 149   TRUE     309
#> 11 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 12 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 13 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#> 14 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#>    Reticle DistNm Cue Method Photos Birds CalibSchool PhotosAerial Biopsy  Prob
#> 1      2.8   1.06   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 2      0.4   2.97   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 3       NA   0.03  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 4       NA   0.02  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 5     14.0   0.28   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 6       NA   0.05  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 7      0.2   3.68   3      4      Y     Y        <NA>         <NA>   <NA> FALSE
#> 8      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 9      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 10     1.7   1.47  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 11     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 12     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 13     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#> 14     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#>    nSp Mixed SpCode SpCodeProb GsSchoolBest GsSchoolHigh GsSchoolLow   GsSpBest
#> 1    1 FALSE    018       <NA>    43.333333     57.00000   43.000000  43.333333
#> 2    1 FALSE    076       <NA>     8.333333     11.66667    7.666667   8.333333
#> 3   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 4   NA    NA     DC       <NA>     1.000000           NA          NA   1.000000
#> 5    1 FALSE    037       <NA>    10.333333     20.00000   10.000000  10.333333
#> 6   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 7    1 FALSE    016       <NA>    47.000000     80.66667   47.000000  47.000000
#> 8    2  TRUE    013       <NA>    43.000000     75.50000   43.000000  31.655000
#> 9    2  TRUE    016       <NA>    43.000000     75.50000   43.000000  11.345000
#> 10  NA    NA   <NA>       <NA>           NA           NA          NA         NA
#> 11   2  TRUE    018       <NA>   151.000000    250.00000  151.000000 126.800000
#> 12   2  TRUE    277       <NA>   151.000000    250.00000  151.000000  24.200000
#> 13   2  TRUE    016        016    21.250000     38.25000   21.250000  16.175000
#> 14   2  TRUE    277        016    21.250000     38.25000   21.250000   5.075000
#>     GsSpHigh    GsSpLow CourseSchool TurtleJFR TurtleAge TurtleCapt PerpDistKm
#> 1   57.00000  43.000000           NA      <NA>      <NA>       <NA> 1.52563078
#> 2   11.66667   7.666667           NA      <NA>      <NA>       <NA> 3.07580701
#> 3         NA         NA           NA      <NA>         A          N 0.04811637
#> 4         NA         NA           NA      <NA>         J          N 0.03207758
#> 5   20.00000  10.000000           NA      <NA>      <NA>       <NA> 0.51856000
#> 6         NA         NA           NA      <NA>         A       <NA> 0.06547809
#> 7   80.66667  47.000000           NA      <NA>      <NA>       <NA> 1.87856781
#> 8   55.09000  31.655000           NA      <NA>      <NA>       <NA> 2.88891582
#> 9   20.41000  11.345000           NA      <NA>      <NA>       <NA> 2.88891582
#> 10        NA         NA           NA      <NA>      <NA>       <NA> 2.11573325
#> 11 208.85000 126.800000           NA      <NA>      <NA>       <NA> 0.10601569
#> 12  41.15000  24.200000           NA      <NA>      <NA>       <NA> 0.10601569
#> 13  28.45500  16.175000           NA      <NA>      <NA>       <NA> 2.54265727
#> 14   9.79500   5.075000           NA      <NA>      <NA>       <NA> 2.54265727
#>    included
#> 1      TRUE
#> 2      TRUE
#> 3     FALSE
#> 4     FALSE
#> 5      TRUE
#> 6     FALSE
#> 7      TRUE
#> 8      TRUE
#> 9      TRUE
#> 10     TRUE
#> 11     TRUE
#> 12     TRUE
#> 13     TRUE
#> 14     TRUE
#> 
#> $randpicks
#> NULL
#> 

# \donttest{
# Using "equallength" method
y.rand <- system.file("das_sample_randpicks.csv", package = "swfscDAS")
das_effort(
  y.proc, method = "equallength", seg.km = 10, randpicks.load = y.rand,
  num.cores = 1
)

#> $segdata
#>    segnum section_id section_sub_id           file stlin endlin     lat1
#> 1       1          1              1 das_sample.das     2     20 43.94750
#> 2       2          2              1 das_sample.das    23     34 44.00333
#> 3       3          2              2 das_sample.das    34     43 44.06155
#> 4       4          3              1 das_sample.das    59     68 44.19517
#> 5       5          3              2 das_sample.das    68     90 44.28108
#> 6       6          4              1 das_sample.das    99    113 44.57233
#> 7       7          4              2 das_sample.das   113    121 44.65409
#> 8       8          5              1 das_sample.das   127    147 44.77933
#> 9       9          6              1 das_sample.das   150    164 44.89583
#> 10     10          7              1 das_sample.das   167    181 45.00967
#> 11     11          8              1 das_sample.das   188    199 45.14917
#> 12     12          9              1 das_sample.das   232    240 45.61433
#> 13     13         10              1 das_sample.das   242    259 45.65100
#>         lon1     lat2      lon2     mlat      mlon    dist           mDateTime
#> 1  -141.0217 43.99433 -140.9990 43.97093 -141.0104  5.5117 2014-03-05 06:37:02
#> 2  -141.0152 44.06155 -140.9864 44.03248 -141.0009  6.8667 2014-03-05 07:12:44
#> 3  -140.9864 44.14649 -140.9450 44.10397 -140.9655 10.0000 2014-03-05 07:42:15
#> 4  -140.8703 44.28108 -140.8334 44.23809 -140.8515 10.0000 2014-03-05 09:39:49
#> 5  -140.8334 44.38150 -140.8280 44.33148 -140.8126 11.6626 2014-03-05 10:16:56
#> 6  -140.7865 44.65409 -140.7338 44.61326 -140.7602 10.0000 2014-03-05 12:14:38
#> 7  -140.7338 44.75460 -140.6661 44.70464 -140.7006 12.3894 2014-03-05 12:57:01
#> 8  -140.5910 44.89333 -140.5522 44.83613 -140.5704 13.0345 2014-03-05 14:14:10
#> 9  -140.5442 45.00312 -140.5089 44.94967 -140.5275 12.2501 2014-03-05 15:21:13
#> 10 -140.5150 45.07850 -140.4802 45.04427 -140.4983  8.1244 2014-03-05 16:14:15
#> 11 -140.4707 45.15250 -140.4688 45.15082 -140.4697  0.3974 2014-03-05 17:00:37
#> 12 -139.0153 45.64298 -138.9955 45.62866 -139.0054  3.5391 2014-03-06 11:30:58
#> 13 -138.9917 45.67316 -138.9768 45.66216 -138.9845  2.7200 2014-03-06 11:45:33
#>    year month day    mtime Cruise Mode EffType ESWsides   avgBft avgSwellHght
#> 1  2014     3   5 06:37:02   1000    C       S        2 3.000000            3
#> 2  2014     3   5 07:12:44   1000    C       S        2 3.000000            3
#> 3  2014     3   5 07:42:15   1000    C       S        2 3.000000            3
#> 4  2014     3   5 09:39:49   1000    C       S        2 3.000000            3
#> 5  2014     3   5 10:16:56   1000    C       S        2 2.072851            3
#> 6  2014     3   5 12:14:38   1000    C       S        2 3.000000            3
#> 7  2014     3   5 12:57:01   1000    C       S        2 3.000000            3
#> 8  2014     3   5 14:14:10   1000    C       S        2 3.000000            3
#> 9  2014     3   5 15:21:13   1000    C       S        2 2.479812            3
#> 10 2014     3   5 16:14:15   1000    C       S        2 2.682941            3
#> 11 2014     3   5 17:00:37   1000    C       S        2 2.000000            3
#> 12 2014     3   6 11:30:58   1000    C       S        2 2.000000            1
#> 13 2014     3   6 11:45:33   1000    C       S        2 2.000000            1
#>    avgHorizSun avgVertSun avgGlare   avgVis
#> 1     2.000000          3        0 6.000000
#> 2           NA         NA       NA 5.957924
#> 3     2.000000          2        0 5.500000
#> 4     2.000000          2        0 5.500000
#> 5     2.000000          2        0 4.740370
#> 6    12.000000         12        0 5.941955
#> 7    12.000000         12        0 6.000000
#> 8     8.000000          1        0 6.000000
#> 9     8.520188          1        0 6.000000
#> 10    8.000000          2        0 6.000000
#> 11    8.000000          2        0 6.000000
#> 12          NA         NA       NA 4.000000
#> 13          NA         NA       NA 4.000000
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime year      Lat       Lon
#> 1       1 43.97093 -141.0104     S 2014-03-05 06:46:02 2014 43.99333 -140.9993
#> 2       3 44.10397 -140.9655     S 2014-03-05 07:56:22 2014 44.14483 -140.9458
#> 3       4 44.23809 -140.8515     t 2014-03-05 09:34:27 2014 44.22450 -140.8573
#> 4       8 44.83613 -140.5704     t 2014-03-05 14:02:55 2014 44.81000 -140.5795
#> 5       8 44.83613 -140.5704     S 2014-03-05 14:37:56 2014 44.89283 -140.5523
#> 6       9 44.94967 -140.5275     t 2014-03-05 15:36:47 2014 44.98767 -140.5152
#> 7      10 45.04427 -140.4983     S 2014-03-05 16:29:50 2014 45.07850 -140.4802
#> 8      11 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 9      11 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 10     12 45.62866 -139.0054     F 2014-03-06 11:25:32 2014 45.61667 -139.0138
#> 11     13 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 12     13 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 13     13 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#> 14     13 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#>    OnEffort Cruise Mode EffType ESWsides Course SpdKt Bft SwellHght WindSpdKt
#> 1      TRUE   1000    C       S        2     25  10.2   3         3        10
#> 2      TRUE   1000    C       S        2     26   9.7   3         3        10
#> 3      TRUE   1000    C       S        2     27   9.0   3         3        10
#> 4      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 5      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 6      TRUE   1000    C       S        2     16   8.9   2         3         6
#> 7      TRUE   1000    C       S        2     25   8.9   2         3         6
#> 8      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 9      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 10     TRUE   1000    C       S        2     35   9.5   2         1         5
#> 11     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 12     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 13     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 14     TRUE   1000    C       S        2     23   9.6   2         1         5
#>    RainFog HorizSun VertSun Glare Vis ObsL Rec ObsR ObsInd EffortDot EventNum
#> 1        1       NA      NA    NA 6.0  208 280  001   <NA>      TRUE       15
#> 2        3        2       2 FALSE 5.5  125 208  280   <NA>      TRUE       35
#> 3        1        2       2 FALSE 5.5  001 126  149   <NA>      TRUE       59
#> 4        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      131
#> 5        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      136
#> 6        1        9       1 FALSE 6.0  125 208  280   <NA>      TRUE      153
#> 7        1        8       2 FALSE 6.0  149 125  208   <NA>      TRUE      167
#> 8        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 9        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 10       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE        8
#> 11       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 12       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 13       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#> 14       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#>          file_das line_num SightNo Subgroup SightNoDaily Obs ObsStd Bearing
#> 1  das_sample.das       15    1406     <NA>   20140305_1 208   TRUE     309
#> 2  das_sample.das       38    1407     <NA>   20140305_2 125   TRUE     326
#> 3  das_sample.das       65    <NA>     <NA>         <NA> 280  FALSE     120
#> 4  das_sample.das      137    <NA>     <NA>         <NA> 228  FALSE     300
#> 5  das_sample.das      142    1408     <NA>   20140305_3 280   TRUE     270
#> 6  das_sample.das      162    <NA>     <NA>         <NA> 231  FALSE      45
#> 7  das_sample.das      176    1409     <NA>   20140305_4 149   TRUE     344
#> 8  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 9  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 10 das_sample.das      238    <NA>     <NA>         <NA> 149   TRUE     309
#> 11 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 12 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 13 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#> 14 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#>    Reticle DistNm Cue Method Photos Birds CalibSchool PhotosAerial Biopsy  Prob
#> 1      2.8   1.06   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 2      0.4   2.97   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 3       NA   0.03  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 4       NA   0.02  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 5     14.0   0.28   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 6       NA   0.05  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 7      0.2   3.68   3      4      Y     Y        <NA>         <NA>   <NA> FALSE
#> 8      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 9      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 10     1.7   1.47  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 11     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 12     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 13     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#> 14     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#>    nSp Mixed SpCode SpCodeProb GsSchoolBest GsSchoolHigh GsSchoolLow   GsSpBest
#> 1    1 FALSE    018       <NA>    43.333333     57.00000   43.000000  43.333333
#> 2    1 FALSE    076       <NA>     8.333333     11.66667    7.666667   8.333333
#> 3   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 4   NA    NA     DC       <NA>     1.000000           NA          NA   1.000000
#> 5    1 FALSE    037       <NA>    10.333333     20.00000   10.000000  10.333333
#> 6   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 7    1 FALSE    016       <NA>    47.000000     80.66667   47.000000  47.000000
#> 8    2  TRUE    013       <NA>    43.000000     75.50000   43.000000  31.655000
#> 9    2  TRUE    016       <NA>    43.000000     75.50000   43.000000  11.345000
#> 10  NA    NA   <NA>       <NA>           NA           NA          NA         NA
#> 11   2  TRUE    018       <NA>   151.000000    250.00000  151.000000 126.800000
#> 12   2  TRUE    277       <NA>   151.000000    250.00000  151.000000  24.200000
#> 13   2  TRUE    016        016    21.250000     38.25000   21.250000  16.175000
#> 14   2  TRUE    277        016    21.250000     38.25000   21.250000   5.075000
#>     GsSpHigh    GsSpLow CourseSchool TurtleJFR TurtleAge TurtleCapt PerpDistKm
#> 1   57.00000  43.000000           NA      <NA>      <NA>       <NA> 1.52563078
#> 2   11.66667   7.666667           NA      <NA>      <NA>       <NA> 3.07580701
#> 3         NA         NA           NA      <NA>         A          N 0.04811637
#> 4         NA         NA           NA      <NA>         J          N 0.03207758
#> 5   20.00000  10.000000           NA      <NA>      <NA>       <NA> 0.51856000
#> 6         NA         NA           NA      <NA>         A       <NA> 0.06547809
#> 7   80.66667  47.000000           NA      <NA>      <NA>       <NA> 1.87856781
#> 8   55.09000  31.655000           NA      <NA>      <NA>       <NA> 2.88891582
#> 9   20.41000  11.345000           NA      <NA>      <NA>       <NA> 2.88891582
#> 10        NA         NA           NA      <NA>      <NA>       <NA> 2.11573325
#> 11 208.85000 126.800000           NA      <NA>      <NA>       <NA> 0.10601569
#> 12  41.15000  24.200000           NA      <NA>      <NA>       <NA> 0.10601569
#> 13  28.45500  16.175000           NA      <NA>      <NA>       <NA> 2.54265727
#> 14   9.79500   5.075000           NA      <NA>      <NA>       <NA> 2.54265727
#>    included
#> 1      TRUE
#> 2      TRUE
#> 3     FALSE
#> 4     FALSE
#> 5      TRUE
#> 6     FALSE
#> 7      TRUE
#> 8      TRUE
#> 9      TRUE
#> 10     TRUE
#> 11     TRUE
#> 12     TRUE
#> 13     TRUE
#> 14     TRUE
#> 
#> $randpicks
#>    effort_section randpicks
#> 1               1        NA
#> 2               2         1
#> 3               3         2
#> 4               4         2
#> 5               5         1
#> 6               6         1
#> 7               7        NA
#> 8               8        NA
#> 9               9        NA
#> 10             10        NA
#> 

# Using "section" method
das_effort(y.proc, method = "section", num.cores = 1)

#> $segdata
#>    segnum section_id section_sub_id           file stlin endlin     lat1
#> 1       1          1              1 das_sample.das     2     20 43.94750
#> 2       2          2              1 das_sample.das    23     43 44.00333
#> 3       3          3              1 das_sample.das    59     90 44.19517
#> 4       4          4              1 das_sample.das    99    121 44.57233
#> 5       5          5              1 das_sample.das   127    147 44.77933
#> 6       6          6              1 das_sample.das   150    164 44.89583
#> 7       7          7              1 das_sample.das   167    181 45.00967
#> 8       8          8              1 das_sample.das   188    199 45.14917
#> 9       9          9              1 das_sample.das   232    240 45.61433
#> 10     10         10              1 das_sample.das   242    259 45.65100
#>         lon1     lat2      lon2     mlat      mlon    dist           mDateTime
#> 1  -141.0217 43.99433 -140.9990 43.97093 -141.0104  5.5117 2014-03-05 06:37:02
#> 2  -141.0152 44.14649 -140.9450 44.07485 -140.9798 16.8667 2014-03-05 07:27:34
#> 3  -140.8703 44.38150 -140.8280 44.28834 -140.8304 21.6626 2014-03-05 09:59:20
#> 4  -140.7865 44.75460 -140.6661 44.66389 -140.7275 22.3894 2014-03-05 12:34:14
#> 5  -140.5910 44.89333 -140.5522 44.83613 -140.5704 13.0345 2014-03-05 14:14:10
#> 6  -140.5442 45.00312 -140.5089 44.94967 -140.5275 12.2501 2014-03-05 15:21:13
#> 7  -140.5150 45.07850 -140.4802 45.04427 -140.4983  8.1244 2014-03-05 16:14:15
#> 8  -140.4707 45.15250 -140.4688 45.15082 -140.4697  0.3974 2014-03-05 17:00:37
#> 9  -139.0153 45.64298 -138.9955 45.62866 -139.0054  3.5391 2014-03-06 11:30:58
#> 10 -138.9917 45.67316 -138.9768 45.66216 -138.9845  2.7200 2014-03-06 11:45:33
#>    year month day    mtime Cruise Mode EffType ESWsides   avgBft avgSwellHght
#> 1  2014     3   5 06:37:02   1000    C       S        2 3.000000            3
#> 2  2014     3   5 07:27:34   1000    C       S        2 3.000000            3
#> 3  2014     3   5 09:59:20   1000    C       S        2 2.500846            3
#> 4  2014     3   5 12:34:14   1000    C       S        2 3.000000            3
#> 5  2014     3   5 14:14:10   1000    C       S        2 3.000000            3
#> 6  2014     3   5 15:21:13   1000    C       S        2 2.479812            3
#> 7  2014     3   5 16:14:15   1000    C       S        2 2.682941            3
#> 8  2014     3   5 17:00:37   1000    C       S        2 2.000000            3
#> 9  2014     3   6 11:30:58   1000    C       S        2 2.000000            1
#> 10 2014     3   6 11:45:33   1000    C       S        2 2.000000            1
#>    avgHorizSun avgVertSun avgGlare   avgVis
#> 1     2.000000          3        0 6.000000
#> 2     2.000000          2        0 5.686427
#> 3     2.000000          2        0 5.091034
#> 4    12.000000         12        0 5.974075
#> 5     8.000000          1        0 6.000000
#> 6     8.520188          1        0 6.000000
#> 7     8.000000          2        0 6.000000
#> 8     8.000000          2        0 6.000000
#> 9           NA         NA       NA 4.000000
#> 10          NA         NA       NA 4.000000
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime year      Lat       Lon
#> 1       1 43.97093 -141.0104     S 2014-03-05 06:46:02 2014 43.99333 -140.9993
#> 2       2 44.07485 -140.9798     S 2014-03-05 07:56:22 2014 44.14483 -140.9458
#> 3       3 44.28834 -140.8304     t 2014-03-05 09:34:27 2014 44.22450 -140.8573
#> 4       5 44.83613 -140.5704     t 2014-03-05 14:02:55 2014 44.81000 -140.5795
#> 5       5 44.83613 -140.5704     S 2014-03-05 14:37:56 2014 44.89283 -140.5523
#> 6       6 44.94967 -140.5275     t 2014-03-05 15:36:47 2014 44.98767 -140.5152
#> 7       7 45.04427 -140.4983     S 2014-03-05 16:29:50 2014 45.07850 -140.4802
#> 8       8 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 9       8 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 10      9 45.62866 -139.0054     F 2014-03-06 11:25:32 2014 45.61667 -139.0138
#> 11     10 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 12     10 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 13     10 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#> 14     10 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#>    OnEffort Cruise Mode EffType ESWsides Course SpdKt Bft SwellHght WindSpdKt
#> 1      TRUE   1000    C       S        2     25  10.2   3         3        10
#> 2      TRUE   1000    C       S        2     26   9.7   3         3        10
#> 3      TRUE   1000    C       S        2     27   9.0   3         3        10
#> 4      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 5      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 6      TRUE   1000    C       S        2     16   8.9   2         3         6
#> 7      TRUE   1000    C       S        2     25   8.9   2         3         6
#> 8      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 9      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 10     TRUE   1000    C       S        2     35   9.5   2         1         5
#> 11     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 12     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 13     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 14     TRUE   1000    C       S        2     23   9.6   2         1         5
#>    RainFog HorizSun VertSun Glare Vis ObsL Rec ObsR ObsInd EffortDot EventNum
#> 1        1       NA      NA    NA 6.0  208 280  001   <NA>      TRUE       15
#> 2        3        2       2 FALSE 5.5  125 208  280   <NA>      TRUE       35
#> 3        1        2       2 FALSE 5.5  001 126  149   <NA>      TRUE       59
#> 4        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      131
#> 5        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      136
#> 6        1        9       1 FALSE 6.0  125 208  280   <NA>      TRUE      153
#> 7        1        8       2 FALSE 6.0  149 125  208   <NA>      TRUE      167
#> 8        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 9        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 10       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE        8
#> 11       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 12       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 13       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#> 14       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#>          file_das line_num SightNo Subgroup SightNoDaily Obs ObsStd Bearing
#> 1  das_sample.das       15    1406     <NA>   20140305_1 208   TRUE     309
#> 2  das_sample.das       38    1407     <NA>   20140305_2 125   TRUE     326
#> 3  das_sample.das       65    <NA>     <NA>         <NA> 280  FALSE     120
#> 4  das_sample.das      137    <NA>     <NA>         <NA> 228  FALSE     300
#> 5  das_sample.das      142    1408     <NA>   20140305_3 280   TRUE     270
#> 6  das_sample.das      162    <NA>     <NA>         <NA> 231  FALSE      45
#> 7  das_sample.das      176    1409     <NA>   20140305_4 149   TRUE     344
#> 8  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 9  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 10 das_sample.das      238    <NA>     <NA>         <NA> 149   TRUE     309
#> 11 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 12 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 13 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#> 14 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#>    Reticle DistNm Cue Method Photos Birds CalibSchool PhotosAerial Biopsy  Prob
#> 1      2.8   1.06   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 2      0.4   2.97   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 3       NA   0.03  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 4       NA   0.02  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 5     14.0   0.28   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 6       NA   0.05  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 7      0.2   3.68   3      4      Y     Y        <NA>         <NA>   <NA> FALSE
#> 8      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 9      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 10     1.7   1.47  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 11     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 12     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 13     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#> 14     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#>    nSp Mixed SpCode SpCodeProb GsSchoolBest GsSchoolHigh GsSchoolLow   GsSpBest
#> 1    1 FALSE    018       <NA>    43.333333     57.00000   43.000000  43.333333
#> 2    1 FALSE    076       <NA>     8.333333     11.66667    7.666667   8.333333
#> 3   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 4   NA    NA     DC       <NA>     1.000000           NA          NA   1.000000
#> 5    1 FALSE    037       <NA>    10.333333     20.00000   10.000000  10.333333
#> 6   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 7    1 FALSE    016       <NA>    47.000000     80.66667   47.000000  47.000000
#> 8    2  TRUE    013       <NA>    43.000000     75.50000   43.000000  31.655000
#> 9    2  TRUE    016       <NA>    43.000000     75.50000   43.000000  11.345000
#> 10  NA    NA   <NA>       <NA>           NA           NA          NA         NA
#> 11   2  TRUE    018       <NA>   151.000000    250.00000  151.000000 126.800000
#> 12   2  TRUE    277       <NA>   151.000000    250.00000  151.000000  24.200000
#> 13   2  TRUE    016        016    21.250000     38.25000   21.250000  16.175000
#> 14   2  TRUE    277        016    21.250000     38.25000   21.250000   5.075000
#>     GsSpHigh    GsSpLow CourseSchool TurtleJFR TurtleAge TurtleCapt PerpDistKm
#> 1   57.00000  43.000000           NA      <NA>      <NA>       <NA> 1.52563078
#> 2   11.66667   7.666667           NA      <NA>      <NA>       <NA> 3.07580701
#> 3         NA         NA           NA      <NA>         A          N 0.04811637
#> 4         NA         NA           NA      <NA>         J          N 0.03207758
#> 5   20.00000  10.000000           NA      <NA>      <NA>       <NA> 0.51856000
#> 6         NA         NA           NA      <NA>         A       <NA> 0.06547809
#> 7   80.66667  47.000000           NA      <NA>      <NA>       <NA> 1.87856781
#> 8   55.09000  31.655000           NA      <NA>      <NA>       <NA> 2.88891582
#> 9   20.41000  11.345000           NA      <NA>      <NA>       <NA> 2.88891582
#> 10        NA         NA           NA      <NA>      <NA>       <NA> 2.11573325
#> 11 208.85000 126.800000           NA      <NA>      <NA>       <NA> 0.10601569
#> 12  41.15000  24.200000           NA      <NA>      <NA>       <NA> 0.10601569
#> 13  28.45500  16.175000           NA      <NA>      <NA>       <NA> 2.54265727
#> 14   9.79500   5.075000           NA      <NA>      <NA>       <NA> 2.54265727
#>    included
#> 1      TRUE
#> 2      TRUE
#> 3     FALSE
#> 4     FALSE
#> 5      TRUE
#> 6     FALSE
#> 7      TRUE
#> 8      TRUE
#> 9      TRUE
#> 10     TRUE
#> 11     TRUE
#> 12     TRUE
#> 13     TRUE
#> 14     TRUE
#> 
#> $randpicks
#>    effort_section randpicks
#> 1               1        NA
#> 2               2        NA
#> 3               3        NA
#> 4               4        NA
#> 5               5        NA
#> 6               6        NA
#> 7               7        NA
#> 8               8        NA
#> 9               9        NA
#> 10             10        NA
#> 

# }