Summarize DAS sightings by effort segment

Summarize number of sightings and animals for selected species by segment

das_effort_sight(
  x.list,
  sp.codes,
  sp.events = c("S", "G", "K", "M", "t", "p"),
  gs.low.use = TRUE
)

Arguments

x.list
output of das_effort; a list of three data frames named 'segdata', 'sightinfo', and 'randpicks', respectively

sp.codes
character; species code(s) to include in segdata output. These must exactly match the species codes in the data, such as including leading zeros

sp.events
character; event code(s) to include in the sightinfo output. This argument supersedes the 'included' value when determining whether a sighting is included in the segment summaries. Must be one or more of: "S", "K", "M", "G", "t", "p" (case-sensitive). The default is that all of these event codes are kept

gs.low.use
logical indicating if the low species group size estimate ('GsSpLow') should be used ONLY if the best estimate ('GsSpLow') is NA

Value

A list, identical to x.list except for 1) the nSI and ANI columns added to x.list$segdata, one each for each element of sp.codes, and 2) the 'included' column of x.list$sightinfo, which has been set as FALSE for sightings of species not listed in sp.codes. Thus, the 'included' column accurately reflects the sightings that were included in the effort segment summaries

Details

This function takes the output of das_effort and adds columns for the number of sightings (nSI) and number of animals (ANI) for selected species (selected via sp.codes) for each segment to the segdata element of x.list. However, only sightings with an included value of TRUE (included is a column in sightinfo) are included in the summaries. Having this step separate from das_effort allows users to personalize the included values as desired for their analysis.

The ANI columns are the sum of the 'GsSpBest' column output from das_sight. The only exception is if 'GsSpBest' is NA and gs.low.use is TRUE, in which case the value from 'GsSpLow' is used instead

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)
y.cond <- das_effort(
  y.proc, method = "condition",
  conditions = "Bft", seg.min.km = 0.05, num.cores = 1
)

das_effort_sight(y.cond, sp.codes = c("013", "076", "DC"), sp.events = c("S", "t"))

#> $segdata
#>    segnum section_id section_sub_id           file stlin endlin     lat1
#> 1       1          1              1 das_sample.das     2     20 43.94750
#> 2       2          2              1 das_sample.das    23     43 44.00333
#> 3       3          3              1 das_sample.das    59     70 44.19517
#> 4       4          3              2 das_sample.das    70     90 44.28850
#> 5       5          4              1 das_sample.das    99    121 44.57233
#> 6       6          5              1 das_sample.das   127    147 44.77933
#> 7       7          6              1 das_sample.das   150    160 44.89583
#> 8       8          6              2 das_sample.das   160    164 44.94750
#> 9       9          7              1 das_sample.das   167    174 45.00967
#> 10     10          7              2 das_sample.das   174    181 45.05681
#> 11     11          8              1 das_sample.das   188    199 45.14917
#> 12     12          9              1 das_sample.das   232    240 45.61433
#> 13     13         10              1 das_sample.das   242    259 45.65100
#>         lon1     lat2      lon2     mlat      mlon    dist           mDateTime
#> 1  -141.0217 43.99433 -140.9990 43.97093 -141.0104  5.5117 2014-03-05 06:37:02
#> 2  -141.0152 44.14649 -140.9450 44.07485 -140.9798 16.8667 2014-03-05 07:27:34
#> 3  -140.8703 44.28850 -140.8303 44.24175 -140.8500 10.8496 2014-03-05 09:41:01
#> 4  -140.8303 44.38150 -140.8280 44.33518 -140.8112 10.8130 2014-03-05 10:18:08
#> 5  -140.7865 44.75460 -140.6661 44.66389 -140.7275 22.3894 2014-03-05 12:34:14
#> 6  -140.5910 44.89333 -140.5522 44.83613 -140.5704 13.0345 2014-03-05 14:14:10
#> 7  -140.5442 44.94750 -140.5282 44.92165 -140.5360  5.8777 2014-03-05 15:09:52
#> 8  -140.5282 45.00312 -140.5089 44.97547 -140.5192  6.3723 2014-03-05 15:31:47
#> 9  -140.5150 45.05681 -140.4919 45.03334 -140.5038  5.5485 2014-03-05 16:09:21
#> 10 -140.4919 45.07850 -140.4802 45.06769 -140.4861  2.5759 2014-03-05 16:24:56
#> 11 -140.4707 45.15250 -140.4688 45.15082 -140.4697  0.3974 2014-03-05 17:00:37
#> 12 -139.0153 45.64298 -138.9955 45.62866 -139.0054  3.5391 2014-03-06 11:30:58
#> 13 -138.9917 45.67316 -138.9768 45.66216 -138.9845  2.7200 2014-03-06 11:45:33
#>    year month day    mtime Cruise Mode EffType ESWsides maxdistBft nSI_013
#> 1  2014     3   5 06:37:02   1000    C       S        2          3       0
#> 2  2014     3   5 07:27:34   1000    C       S        2          3       0
#> 3  2014     3   5 09:41:01   1000    C       S        2          3       0
#> 4  2014     3   5 10:18:08   1000    C       S        2          2       0
#> 5  2014     3   5 12:34:14   1000    C       S        2          3       0
#> 6  2014     3   5 14:14:10   1000    C       S        2          3       0
#> 7  2014     3   5 15:09:52   1000    C       S        2          3       0
#> 8  2014     3   5 15:31:47   1000    C       S        2          2       0
#> 9  2014     3   5 16:09:21   1000    C       S        2          3       0
#> 10 2014     3   5 16:24:56   1000    C       S        2          2       0
#> 11 2014     3   5 17:00:37   1000    C       S        2          2       1
#> 12 2014     3   6 11:30:58   1000    C       S        2          2       0
#> 13 2014     3   6 11:45:33   1000    C       S        2          2       0
#>    ANI_013 nSI_076  ANI_076 nSI_DC ANI_DC
#> 1    0.000       0 0.000000      0      0
#> 2    0.000       1 8.333333      0      0
#> 3    0.000       0 0.000000      0      0
#> 4    0.000       0 0.000000      0      0
#> 5    0.000       0 0.000000      0      0
#> 6    0.000       0 0.000000      0      0
#> 7    0.000       0 0.000000      0      0
#> 8    0.000       0 0.000000      0      0
#> 9    0.000       0 0.000000      0      0
#> 10   0.000       0 0.000000      0      0
#> 11  31.655       0 0.000000      0      0
#> 12   0.000       0 0.000000      0      0
#> 13   0.000       0 0.000000      0      0
#> 
#> $sightinfo
#>    segnum     mlat      mlon Event            DateTime year      Lat       Lon
#> 1       1 43.97093 -141.0104     S 2014-03-05 06:46:02 2014 43.99333 -140.9993
#> 2       2 44.07485 -140.9798     S 2014-03-05 07:56:22 2014 44.14483 -140.9458
#> 3       3 44.24175 -140.8500     t 2014-03-05 09:34:27 2014 44.22450 -140.8573
#> 4       6 44.83613 -140.5704     t 2014-03-05 14:02:55 2014 44.81000 -140.5795
#> 5       6 44.83613 -140.5704     S 2014-03-05 14:37:56 2014 44.89283 -140.5523
#> 6       8 44.97547 -140.5192     t 2014-03-05 15:36:47 2014 44.98767 -140.5152
#> 7      10 45.06769 -140.4861     S 2014-03-05 16:29:50 2014 45.07850 -140.4802
#> 8      11 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 9      11 45.15082 -140.4697     S 2014-03-05 17:00:45 2014 45.15117 -140.4695
#> 10     13 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 11     13 45.66216 -138.9845     S 2014-03-06 11:47:51 2014 45.66733 -138.9808
#> 12     13 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#> 13     13 45.66216 -138.9845     S 2014-03-06 11:49:14 2014 45.67050 -138.9788
#>    OnEffort Cruise Mode EffType ESWsides Course SpdKt Bft SwellHght WindSpdKt
#> 1      TRUE   1000    C       S        2     25  10.2   3         3        10
#> 2      TRUE   1000    C       S        2     26   9.7   3         3        10
#> 3      TRUE   1000    C       S        2     27   9.0   3         3        10
#> 4      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 5      TRUE   1000    C       S        2     20   9.3   3         3         6
#> 6      TRUE   1000    C       S        2     16   8.9   2         3         6
#> 7      TRUE   1000    C       S        2     25   8.9   2         3         6
#> 8      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 9      TRUE   1000    C       S        2     30   9.5   2         3         6
#> 10     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 11     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 12     TRUE   1000    C       S        2     23   9.6   2         1         5
#> 13     TRUE   1000    C       S        2     23   9.6   2         1         5
#>    RainFog HorizSun VertSun Glare Vis ObsL Rec ObsR ObsInd EffortDot EventNum
#> 1        1       NA      NA    NA 6.0  208 280  001   <NA>      TRUE       15
#> 2        3        2       2 FALSE 5.5  125 208  280   <NA>      TRUE       35
#> 3        1        2       2 FALSE 5.5  001 126  149   <NA>      TRUE       59
#> 4        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      131
#> 5        1        8       1 FALSE 6.0  280 001  126   <NA>      TRUE      136
#> 6        1        9       1 FALSE 6.0  125 208  280   <NA>      TRUE      153
#> 7        1        8       2 FALSE 6.0  149 125  208   <NA>      TRUE      167
#> 8        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 9        1        8       2 FALSE 6.0  126 149  125   <NA>      TRUE      181
#> 10       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 11       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       18
#> 12       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#> 13       3       NA      NA    NA 4.0  149 125  208   <NA>      TRUE       20
#>          file_das line_num SightNo Subgroup SightNoDaily Obs ObsStd Bearing
#> 1  das_sample.das       15    1406     <NA>   20140305_1 208   TRUE     309
#> 2  das_sample.das       38    1407     <NA>   20140305_2 125   TRUE     326
#> 3  das_sample.das       65    <NA>     <NA>         <NA> 280  FALSE     120
#> 4  das_sample.das      137    <NA>     <NA>         <NA> 228  FALSE     300
#> 5  das_sample.das      142    1408     <NA>   20140305_3 280   TRUE     270
#> 6  das_sample.das      162    <NA>     <NA>         <NA> 231  FALSE      45
#> 7  das_sample.das      176    1409     <NA>   20140305_4 149   TRUE     344
#> 8  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 9  das_sample.das      193    1410     <NA>   20140305_5 125   TRUE      70
#> 10 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 11 das_sample.das      248    1412     <NA>   20140306_1 149   TRUE     359
#> 12 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#> 13 das_sample.das      252    1413     <NA>   20140306_2 208   TRUE      38
#>    Reticle DistNm Cue Method Photos Birds CalibSchool PhotosAerial Biopsy  Prob
#> 1      2.8   1.06   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 2      0.4   2.97   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 3       NA   0.03  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 4       NA   0.02  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 5     14.0   0.28   3      4      N     N        <NA>         <NA>   <NA> FALSE
#> 6       NA   0.05  NA     NA   <NA>  <NA>        <NA>         <NA>   <NA>    NA
#> 7      0.2   3.68   3      4      Y     Y        <NA>         <NA>   <NA> FALSE
#> 8      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 9      1.4   1.66   3      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 10     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 11     0.3   3.28   2      4      Y     N        <NA>         <NA>   <NA> FALSE
#> 12     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#> 13     0.8   2.23   3      4      Y     N        <NA>         <NA>   <NA>  TRUE
#>    nSp Mixed SpCode SpCodeProb GsSchoolBest GsSchoolHigh GsSchoolLow   GsSpBest
#> 1    1 FALSE    018       <NA>    43.333333     57.00000   43.000000  43.333333
#> 2    1 FALSE    076       <NA>     8.333333     11.66667    7.666667   8.333333
#> 3   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 4   NA    NA     DC       <NA>     1.000000           NA          NA   1.000000
#> 5    1 FALSE    037       <NA>    10.333333     20.00000   10.000000  10.333333
#> 6   NA    NA     LV       <NA>     1.000000           NA          NA   1.000000
#> 7    1 FALSE    016       <NA>    47.000000     80.66667   47.000000  47.000000
#> 8    2  TRUE    013       <NA>    43.000000     75.50000   43.000000  31.655000
#> 9    2  TRUE    016       <NA>    43.000000     75.50000   43.000000  11.345000
#> 10   2  TRUE    018       <NA>   151.000000    250.00000  151.000000 126.800000
#> 11   2  TRUE    277       <NA>   151.000000    250.00000  151.000000  24.200000
#> 12   2  TRUE    016        016    21.250000     38.25000   21.250000  16.175000
#> 13   2  TRUE    277        016    21.250000     38.25000   21.250000   5.075000
#>     GsSpHigh    GsSpLow CourseSchool TurtleJFR TurtleAge TurtleCapt PerpDistKm
#> 1   57.00000  43.000000           NA      <NA>      <NA>       <NA> 1.52563078
#> 2   11.66667   7.666667           NA      <NA>      <NA>       <NA> 3.07580701
#> 3         NA         NA           NA      <NA>         A          N 0.04811637
#> 4         NA         NA           NA      <NA>         J          N 0.03207758
#> 5   20.00000  10.000000           NA      <NA>      <NA>       <NA> 0.51856000
#> 6         NA         NA           NA      <NA>         A       <NA> 0.06547809
#> 7   80.66667  47.000000           NA      <NA>      <NA>       <NA> 1.87856781
#> 8   55.09000  31.655000           NA      <NA>      <NA>       <NA> 2.88891582
#> 9   20.41000  11.345000           NA      <NA>      <NA>       <NA> 2.88891582
#> 10 208.85000 126.800000           NA      <NA>      <NA>       <NA> 0.10601569
#> 11  41.15000  24.200000           NA      <NA>      <NA>       <NA> 0.10601569
#> 12  28.45500  16.175000           NA      <NA>      <NA>       <NA> 2.54265727
#> 13   9.79500   5.075000           NA      <NA>      <NA>       <NA> 2.54265727
#>    included
#> 1     FALSE
#> 2      TRUE
#> 3     FALSE
#> 4     FALSE
#> 5     FALSE
#> 6     FALSE
#> 7     FALSE
#> 8      TRUE
#> 9     FALSE
#> 10    FALSE
#> 11    FALSE
#> 12    FALSE
#> 13    FALSE
#> 
#> $randpicks
#> NULL
#>