Process DAS data

Process DAS data (the output of das_read), including extracting state and condition information for each DAS event

das_process(x, ...)

# S3 method for character
das_process(x, ...)

# S3 method for data.frame
das_process(x, days.gap = 10, reset.event = TRUE, reset.day = TRUE, ...)

Arguments

x
either a data frame (the output of das_read), or a character (filepath) which is first passed to das_read

...
ignored

days.gap
numeric of length 1; time gap (in days) used to identify a new cruise in concatenated DAS files, and thus also when state/condition information (weather, Bft, Mode, etc) is reset. Default is 10 days

reset.event
logical; indicates if state/condition information (weather, Bft, Mode, etc) should be reset to NA if there is an applicable event with an NA for that state/condition

reset.day
logical; indicates if state/condition information (weather, Bft, Mode, etc) should be reset to NA at the beginning of each day. This argument should only be set to FALSE for comparison with older methods, such as Report

Value

Data frame; the input data frame, i.e. the output of das_read, with the following columns added:

State/condition	Column name

On/off effort	OnEffort

Cruise number	Cruise

Effort mode	Mode

Effort type	EffType

Course (ship direction)	Course

Beaufort sea state	Bft

Swell height (ft)	SwellHght

Rain/fog/haze code	RainFog

Horizontal sun (clock system)	HorizSun

Vertical sun (clock system)	VertSun

Glare	Glare

Visibility (nm)	Vis

Details

If x is a character, it is assumed to be a filepath and first passed to das_read. This output is then passed to das_process.

DAS data is event-based, meaning most events indicate when a state or weather condition changes. For instance, a 'V' event indicates when the Beaufort sea state changes, and the Beaufort is the same for subsequent events until the next 'V' event. For each state/condition: a new column is created, the state/condition information is extracted from relevant events, and extracted information is propagated to appropriate subsequent rows (events). Thus, each row in the output data frame contains all pertinent state/condition information for that row.

The following assumptions/decisions are made during processing:

Event codes are expected to be one of the following: #, *, ?, 1, 2, 3, 4, 5, 6, 7, 8, A, B, C, E, F, k, K, N, P, Q, R, s, S, t, V, W

All '#' events (deleted events) are removed

An event is considered 'on effort' if it is 1) an R event, 2) a B event immediately preceding an R event, or 3) between corresponding R and E events. The 'EffortDot' column is ignored

All state/condition information is reset at the beginning of each cruise. New cruises are identifed using days.gap.

'Mode' is capitalized, and 'Mode' values of NA are assigned a value of "C"

'EffType' is capitalized, and values of NA are assigned a value of "S"

'Glare': TRUE if 'HzSun' is 11, 12 or 1 and 'VtSun' is 2 or 3, or if 'HzSun' is 12 and 'VtSun' is 1; NA if 'HzSun' or 'VtSun' is NA; otherwise FALSE

Missing values are NA rather than -1

This function was inspired by das.read

See also

Examples

# TODO