Read DAS file(s)

Read one or more fixed-width DAS text file(s) generated by WinCruz into a data frame, where each line is data for a specific event

das_read(file, tz = "UTC")

Arguments

file
filename(s) of one or more DAS files

tz
character; see strptime. Default is UTC

Value

A das_dfr object, which is also a data frame, with DAS data read into columns. The data are read into the data frame as characters as described in 'Details', with the following exceptions:

Name	Class	Details

EffortDot	logical	TRUE if "." was present, and FALSE otherwise

DateTime	POSIXct	combination of 'Date' and 'Time' columns with time zone tz

Lat	numeric	'Latitude' column converted to decimal degrees in range [-90, 90]

Lon	numeric	'Longitude' column converted to decimal degrees in range [-180, 180]

Data#	character	leading/trailing whitespace trimmed for non-comment events (i.e. rows where 'Event' is not "C" or "c")

file_das	character	filename, extracted from the file argument using basename

event_num	integer	'Event number' coerced to an integer using as.integer

line_num	integer	line number of each data row

Details

Reads/parses DAS data into columns of a data frame. If file contains multiple filenames, then the individual data frames will be concatenated. The provided DAS file must adhere to the following column number and format specifications; note that 'Data#' columns may be referred to as 'Field#' columns in other documentation:

Item	Columns	Format

Sequence (event number)	1-3	Event

4	Effort dot	5

Time	6-11	HHMMSS or HHMM

Date	13-18	MMDDYY

Latitude	20-28	NDD:MM.MM

Longitude	30-39	WDDD:MM.MM

Data1	40-44	Data2

45-49	Data3	50-54

Data4	55-59	Data5

60-64	Data6	65-69

Data7	70-74	Data8

75-79	Data9	80+

See das_format_pdf for more information about DAS format requirements. This function was inspired by das.read

See also

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
das_read(y)

#>     Event EffortDot            DateTime      Lat       Lon Data1 Data2 Data3
#> 1       B      TRUE 2014-07-02 11:49:09 27.53600 -125.5173  1000     C     7
#> 2       R      TRUE 2014-07-02 11:49:09 27.53600 -125.5173     S  <NA>  <NA>
#> 3       P      TRUE 2014-07-02 11:49:09 27.53600 -125.5173   153   073   126
#> 4       V      TRUE 2014-07-02 11:49:09 27.53600 -125.5173     3    04   120
#> 5       N      TRUE 2014-07-02 11:49:09 27.53600 -125.5173   259  10.0  <NA>
#> 6       W      TRUE 2014-07-02 11:49:09 27.53600 -125.5173     3    04    01
#> 7       *      TRUE 2014-07-02 11:55:54 27.54100 -125.5352  <NA>  <NA>  <NA>
#> 8       P      TRUE 2014-07-02 12:00:34 27.54033 -125.5490   126   185   073
#> 9       V      TRUE 2014-07-02 12:00:34 27.54033 -125.5490     3    04   120
#> 10      N      TRUE 2014-07-02 12:00:34 27.54033 -125.5490   259  10.0  <NA>
#> 11      W      TRUE 2014-07-02 12:00:34 27.54033 -125.5490     1    04    01
#> 12      *      TRUE 2014-07-02 12:05:54 27.53900 -125.5647  <NA>  <NA>  <NA>
#> 13      *      TRUE 2014-07-02 12:15:54 27.53650 -125.5955  <NA>  <NA>  <NA>
#> 14      S      TRUE 2014-07-02 12:17:00 27.53617 -125.5988   079   073     3
#> 15      A      TRUE 2014-07-02 12:17:00 27.53617 -125.5988   079  <NA>     N
#> 16      ?     FALSE                <NA>       NA        NA   079  <NA>  <NA>
#> 17      1     FALSE                <NA>       NA        NA   073     3     4
#> 18      *      TRUE 2014-07-02 12:25:54 27.53400 -125.6262  <NA>  <NA>  <NA>
#> 19      P      TRUE 2014-07-02 12:29:48 27.53283 -125.6383   184   126   185
#> 20      V      TRUE 2014-07-02 12:29:48 27.53283 -125.6383     3    04   120
#> 21      N      TRUE 2014-07-02 12:29:48 27.53283 -125.6383   259  10.0  <NA>
#> 22      W      TRUE 2014-07-02 12:29:48 27.53283 -125.6383     1    04    01
#> 23      *      TRUE 2014-07-02 12:35:54 27.53117 -125.6572  <NA>  <NA>  <NA>
#> 24      *      TRUE 2014-07-02 12:45:54 27.52867 -125.6875  <NA>  <NA>  <NA>
#> 25      *      TRUE 2014-07-02 12:55:54 27.52483 -125.7170  <NA>  <NA>  <NA>
#> 26      P      TRUE 2014-07-02 13:00:20 27.52233 -125.7302   001   184   126
#> 27      V      TRUE 2014-07-02 13:00:20 27.52233 -125.7302     3    04   180
#> 28      N      TRUE 2014-07-02 13:00:20 27.52233 -125.7302   259  10.0  <NA>
#> 29      W      TRUE 2014-07-02 13:00:20 27.52233 -125.7302     1    03    01
#> 30      *      TRUE 2014-07-02 13:05:54 27.51833 -125.7462  <NA>  <NA>  <NA>
#> 31      S      TRUE 2014-07-02 13:07:16 27.51750 -125.7502   080   001     3
#> 32      A      TRUE 2014-07-02 13:07:16 27.51750 -125.7502   080  <NA>     N
#> 33      1     FALSE                <NA>       NA        NA   001    20    30
#> 34      E     FALSE 2014-07-02 13:07:53 27.51717 -125.7520     C  <NA>  <NA>
#> 35      *     FALSE 2014-07-02 13:15:54 27.50750 -125.7725  <NA>  <NA>  <NA>
#> 36      N     FALSE 2014-07-02 13:19:07 27.50167 -125.7795   220  10.0  <NA>
#> 37      *     FALSE 2014-07-02 13:25:54 27.49467 -125.7918  <NA>  <NA>  <NA>
#> 38      R      TRUE 2014-07-02 13:34:27 27.48833 -125.8168     S  <NA>  <NA>
#> 39      P      TRUE 2014-07-02 13:34:27 27.48833 -125.8168   182   001   186
#> 40      V      TRUE 2014-07-02 13:34:27 27.48833 -125.8168     3    04   180
#> 41      N      TRUE 2014-07-02 13:34:27 27.48833 -125.8168   252  10.0  <NA>
#> 42      W      TRUE 2014-07-02 13:34:27 27.48833 -125.8168     1  <NA>  <NA>
#> 43      *      TRUE 2014-07-02 13:35:54 27.48767 -125.8212  <NA>  <NA>  <NA>
#> 44      *      TRUE 2014-07-02 13:45:54 27.48100 -125.8505  <NA>  <NA>  <NA>
#> 45      *      TRUE 2014-07-02 13:55:54 27.47400 -125.8800  <NA>  <NA>  <NA>
#> 46      V      TRUE 2014-07-02 13:56:10 27.47383 -125.8808     3    04   180
#> 47      W      TRUE 2014-07-02 13:56:25 27.47367 -125.8815     1  <NA>  <NA>
#> 48      P      TRUE 2014-07-02 14:01:49 27.47100 -125.8973   153   182   001
#> 49      V      TRUE 2014-07-02 14:01:49 27.47100 -125.8973     3    04   180
#> 50      N      TRUE 2014-07-02 14:01:49 27.47100 -125.8973   252  10.0  <NA>
#> 51      W      TRUE 2014-07-02 14:01:49 27.47100 -125.8973     3  <NA>  <NA>
#> 52      *      TRUE 2014-07-02 14:05:54 27.46783 -125.9090  <NA>  <NA>  <NA>
#> 53      *      TRUE 2014-07-02 14:15:54 27.46067 -125.9383  <NA>  <NA>  <NA>
#> 54      *      TRUE 2014-07-02 14:25:54 27.45417 -125.9675  <NA>  <NA>  <NA>
#> 55      W      TRUE 2014-07-02 14:29:31 27.45183 -125.9780     3  <NA>  <NA>
#> 56      P      TRUE 2014-07-02 14:31:21 27.45133 -125.9833   186   153   182
#> 57      V      TRUE 2014-07-02 14:31:21 27.45133 -125.9833     3    04   180
#> 58      N      TRUE 2014-07-02 14:31:21 27.45133 -125.9833   252  10.0  <NA>
#> 59      W      TRUE 2014-07-02 14:31:21 27.45133 -125.9833     3  <NA>  <NA>
#> 60      *      TRUE 2014-07-02 14:35:54 27.44767 -125.9965  <NA>  <NA>  <NA>
#> 61      *      TRUE 2014-07-02 14:45:54 27.44067 -126.0263  <NA>  <NA>  <NA>
#> 62      *      TRUE 2014-07-02 14:55:54 27.43483 -126.0552  <NA>  <NA>  <NA>
#> 63      P      TRUE 2014-07-02 14:59:57 27.43267 -126.0662   073   186   153
#> 64      V      TRUE 2014-07-02 14:59:57 27.43267 -126.0662     3    04   180
#> 65      N      TRUE 2014-07-02 14:59:57 27.43267 -126.0662   252  10.0  <NA>
#> 66      W      TRUE 2014-07-02 14:59:57 27.43267 -126.0662     3  <NA>  <NA>
#> 67      *      TRUE 2014-07-02 15:05:54 27.42867 -126.0838  <NA>  <NA>  <NA>
#> 68      *      TRUE 2014-07-02 15:15:54 27.42117 -126.1125  <NA>  <NA>  <NA>
#> 69      *      TRUE 2014-07-02 15:25:54 27.41467 -126.1410  <NA>  <NA>  <NA>
#> 70      P      TRUE 2014-07-02 15:30:31 27.41117 -126.1543   185   073   186
#> 71      V      TRUE 2014-07-02 15:30:31 27.41117 -126.1543     3    04   180
#> 72      N      TRUE 2014-07-02 15:30:31 27.41117 -126.1543   252  10.0  <NA>
#> 73      W      TRUE 2014-07-02 15:30:31 27.41117 -126.1543     3  <NA>  <NA>
#> 74      *      TRUE 2014-07-02 15:35:54 27.40700 -126.1702  <NA>  <NA>  <NA>
#> 75      *      TRUE 2014-07-02 15:45:54 27.40017 -126.1993  <NA>  <NA>  <NA>
#> 76      *      TRUE 2014-07-02 15:55:54 27.39250 -126.2277  <NA>  <NA>  <NA>
#> 77      P      TRUE 2014-07-02 16:00:09 27.39000 -126.2408   126   185   073
#> 78      V      TRUE 2014-07-02 16:00:09 27.39000 -126.2408     3    04   180
#> 79      N      TRUE 2014-07-02 16:00:09 27.39000 -126.2408   252  10.0  <NA>
#> 80      W      TRUE 2014-07-02 16:00:09 27.39000 -126.2408     3  <NA>  <NA>
#> 81      *      TRUE 2014-07-02 16:05:54 27.38583 -126.2572  <NA>  <NA>  <NA>
#> 82      *      TRUE 2014-07-02 16:15:54 27.37983 -126.2863  <NA>  <NA>  <NA>
#> 83      W      TRUE 2014-07-02 16:17:31 27.37867 -126.2910     1  <NA>  <NA>
#> 84      *      TRUE 2014-07-02 16:25:54 27.37400 -126.3168  <NA>  <NA>  <NA>
#> 85      K      TRUE 2014-07-02 16:28:37 27.37233 -126.3248  1014   186     6
#> 86      A      TRUE 2014-07-02 16:28:37 27.37233 -126.3248  1014  <NA>     N
#> 87      1     FALSE                <NA>       NA        NA   186     3     3
#> 88      P      TRUE 2014-07-02 16:30:15 27.37150 -126.3295   184   126   185
#> 89      V      TRUE 2014-07-02 16:30:15 27.37150 -126.3295     3    04   180
#> 90      N      TRUE 2014-07-02 16:30:15 27.37150 -126.3295   252  10.0  <NA>
#> 91      W      TRUE 2014-07-02 16:30:15 27.37150 -126.3295     1  <NA>  <NA>
#> 92      #      TRUE 2014-07-02 16:33:57 27.36967 -126.3405   081   185     2
#> 93      #      TRUE 2014-07-02 16:33:57 27.36967 -126.3405   081  <NA>     N
#> 94      *      TRUE 2014-07-02 16:35:54 27.36800 -126.3458  <NA>  <NA>  <NA>
#> 95      E     FALSE 2014-07-02 16:36:34 27.36750 -126.3477     U  <NA>  <NA>
#> 96      k     FALSE 2014-07-02 16:37:44 27.36700 -126.3510  1014   315   1.0
#> 97      R      TRUE 2014-07-02 16:43:46 27.36517 -126.3620     S  <NA>  <NA>
#> 98      P      TRUE 2014-07-02 16:43:46 27.36517 -126.3620   184   126   185
#> 99      V      TRUE 2014-07-02 16:43:46 27.36517 -126.3620     3    04   180
#> 100     N      TRUE 2014-07-02 16:43:46 27.36517 -126.3620   252  10.0  <NA>
#> 101     W      TRUE 2014-07-02 16:43:46 27.36517 -126.3620     3  <NA>  <NA>
#> 102     *      TRUE 2014-07-02 16:45:54 27.36400 -126.3673  <NA>  <NA>  <NA>
#> 103     S      TRUE 2014-07-02 16:47:22 27.36317 -126.3718   081   126     6
#> 104     A      TRUE 2014-07-02 16:47:22 27.36317 -126.3718   081  <NA>     N
#> 105     1     FALSE                <NA>       NA        NA   001     6     6
#> 106     2     FALSE                <NA>       NA        NA   073     4     5
#> 107     3     FALSE                <NA>       NA        NA   126     3     5
#> 108     4     FALSE                <NA>       NA        NA   153     4     5
#> 109     5     FALSE                <NA>       NA        NA   184     5     8
#> 110     6     FALSE                <NA>       NA        NA   185     7     9
#> 111     E     FALSE 2014-07-02 16:48:44 27.36217 -126.3758     C  <NA>  <NA>
#> 112     C     FALSE 2014-07-02 16:51:19 27.35967 -126.3827  The  sight ing 8
#> 113     *     FALSE 2014-07-02 16:55:54 27.35117 -126.3835  <NA>  <NA>  <NA>
#> 114     s     FALSE 2014-07-02 16:58:43 27.34817 -126.3832   081   330   2.0
#> 115     N     FALSE 2014-07-02 17:04:50 27.34133 -126.3792   125  10.0  <NA>
#> 116     *     FALSE 2014-07-02 17:05:54 27.34033 -126.3777  <NA>  <NA>  <NA>
#> 117     *     FALSE 2014-07-02 17:15:54 27.33267 -126.3687  <NA>  <NA>  <NA>
#> 118     s     FALSE 2014-07-02 17:21:01 27.32683 -126.3672   081   329   6.0
#> 119     *     FALSE 2014-07-02 17:25:54 27.32083 -126.3662  <NA>  <NA>  <NA>
#> 120     s     FALSE 2014-07-02 17:29:44 27.31633 -126.3655   081   320   9.0
#> 121     *     FALSE 2014-07-02 17:35:54 27.31233 -126.3597  <NA>  <NA>  <NA>
#> 122     *     FALSE 2014-07-02 17:45:54 27.31617 -126.3535  <NA>  <NA>  <NA>
#> 123     s     FALSE 2014-07-02 17:50:15 27.32083 -126.3608   081   280  15.0
#> 124     N     FALSE 2014-07-02 17:51:13 27.32200 -126.3625   290  10.0  <NA>
#> 125     F     FALSE 2014-07-02 17:51:13 27.32200 -126.3625   126   000   2.7
#> 126     *     FALSE 2014-07-02 17:55:54 27.32650 -126.3703  <NA>  <NA>  <NA>
#> 127     *     FALSE 2014-07-02 18:05:54 27.33050 -126.3833  <NA>  <NA>  <NA>
#> 128     N     FALSE 2014-07-02 18:06:51 27.33000 -126.3822   175  10.0  <NA>
#> 129     s     FALSE 2014-07-02 18:10:25 27.32800 -126.3790   081   341   2.5
#> 130     *     FALSE 2014-07-02 18:15:54 27.32700 -126.3727  <NA>  <NA>  <NA>
#> 131     N     FALSE 2014-07-02 18:18:49 27.32683 -126.3692   105  10.0  <NA>
#> 132     s     FALSE 2014-07-02 18:22:24 27.32600 -126.3650   081   000  <NA>
#> 133     *     FALSE 2014-07-02 18:25:54 27.32317 -126.3627  <NA>  <NA>  <NA>
#> 134     *     FALSE 2014-07-02 18:35:54 27.31783 -126.3530  <NA>  <NA>  <NA>
#> 135     C     FALSE 2014-07-02 18:39:42 27.31750 -126.3523  clos ing d own o
#> 136     B      TRUE 2014-07-03 06:31:44 27.10233 -127.7068  1000     C     7
#> 137     R      TRUE 2014-07-03 06:31:44 27.10233 -127.7068     S  <NA>  <NA>
#> 138     P      TRUE 2014-07-03 06:31:44 27.10233 -127.7068   186   153   182
#> 139     V      TRUE 2014-07-03 06:31:44 27.10233 -127.7068     3    04   125
#> 140     N      TRUE 2014-07-03 06:31:44 27.10233 -127.7068   259  10.0  <NA>
#> 141     W      TRUE 2014-07-03 06:31:44 27.10233 -127.7068     1  <NA>  <NA>
#> 142     *      TRUE 2014-07-03 06:41:41 27.09617 -127.7353  <NA>  <NA>  <NA>
#> 143     V      TRUE 2014-07-03 06:47:57 27.09283 -127.7537     3    06   175
#> 144     *      TRUE 2014-07-03 06:51:41 27.09050 -127.7640  <NA>  <NA>  <NA>
#> 145     S      TRUE 2014-07-03 06:53:54 27.08883 -127.7702   082   153     6
#> 146     A      TRUE 2014-07-03 06:53:54 27.08883 -127.7702   082  <NA>     N
#> 147     1     FALSE                <NA>       NA        NA   001     3     4
#> 148     2     FALSE                <NA>       NA        NA   073     6     6
#> 149     3     FALSE                <NA>       NA        NA   126     5     6
#> 150     4     FALSE                <NA>       NA        NA   153     6     6
#> 151     5     FALSE                <NA>       NA        NA   182     4     6
#> 152     6     FALSE                <NA>       NA        NA   185     6     9
#> 153     7     FALSE                <NA>       NA        NA   186     4     7
#> 154     E     FALSE 2014-07-03 06:53:54 27.08883 -127.7702     C  <NA>  <NA>
#>     Data4 Data5 Data6 Data7 Data8
#> 1       Y  <NA>  <NA>  <NA>  <NA>
#> 2    <NA>  <NA>  <NA>  <NA>  <NA>
#> 3    <NA>  <NA>  <NA>  <NA>  <NA>
#> 4    <NA>  09.0  <NA>  <NA>  <NA>
#> 5    <NA>  <NA>  <NA>  <NA>  <NA>
#> 6     140   6.0  <NA>  <NA>  <NA>
#> 7    <NA>  <NA>  <NA>  <NA>  <NA>
#> 8    <NA>  <NA>  <NA>  <NA>  <NA>
#> 9    <NA>  09.0  <NA>  <NA>  <NA>
#> 10   <NA>  <NA>  <NA>  <NA>  <NA>
#> 11    140   6.0  <NA>  <NA>  <NA>
#> 12   <NA>  <NA>  <NA>  <NA>  <NA>
#> 13   <NA>  <NA>  <NA>  <NA>  <NA>
#> 14      4   012   0.2   3.6   210
#> 15      N   075  <NA>  <NA>  <NA>
#> 16   <NA>   074  <NA>  <NA>  <NA>
#> 17      2   100   000   000   000
#> 18   <NA>  <NA>  <NA>  <NA>  <NA>
#> 19   <NA>  <NA>  <NA>  <NA>  <NA>
#> 20   <NA>  09.0  <NA>  <NA>  <NA>
#> 21   <NA>  <NA>  <NA>  <NA>  <NA>
#> 22    140   6.0  <NA>  <NA>  <NA>
#> 23   <NA>  <NA>  <NA>  <NA>  <NA>
#> 24   <NA>  <NA>  <NA>  <NA>  <NA>
#> 25   <NA>  <NA>  <NA>  <NA>  <NA>
#> 26   <NA>  <NA>  <NA>  <NA>  <NA>
#> 27   <NA>  09.0  <NA>  <NA>  <NA>
#> 28   <NA>  <NA>  <NA>  <NA>  <NA>
#> 29    140   7.0  <NA>  <NA>  <NA>
#> 30   <NA>  <NA>  <NA>  <NA>  <NA>
#> 31      4   025   0.6   2.5   330
#> 32      N   018   016  <NA>  <NA>
#> 33     10    60    40   000   000
#> 34   <NA>  <NA>  <NA>  <NA>  <NA>
#> 35   <NA>  <NA>  <NA>  <NA>  <NA>
#> 36   <NA>  <NA>  <NA>  <NA>  <NA>
#> 37   <NA>  <NA>  <NA>  <NA>  <NA>
#> 38   <NA>  <NA>  <NA>  <NA>  <NA>
#> 39   <NA>  <NA>  <NA>  <NA>  <NA>
#> 40   <NA>  09.0  <NA>  <NA>  <NA>
#> 41   <NA>  <NA>  <NA>  <NA>  <NA>
#> 42    140   6.0  <NA>  <NA>  <NA>
#> 43   <NA>  <NA>  <NA>  <NA>  <NA>
#> 44   <NA>  <NA>  <NA>  <NA>  <NA>
#> 45   <NA>  <NA>  <NA>  <NA>  <NA>
#> 46   <NA>  09.0  <NA>  <NA>  <NA>
#> 47    140   5.0  <NA>  <NA>  <NA>
#> 48   <NA>  <NA>  <NA>  <NA>  <NA>
#> 49   <NA>  09.0  <NA>  <NA>  <NA>
#> 50   <NA>  <NA>  <NA>  <NA>  <NA>
#> 51    140   4.0  <NA>  <NA>  <NA>
#> 52   <NA>  <NA>  <NA>  <NA>  <NA>
#> 53   <NA>  <NA>  <NA>  <NA>  <NA>
#> 54   <NA>  <NA>  <NA>  <NA>  <NA>
#> 55    140   5.0  <NA>  <NA>  <NA>
#> 56   <NA>  <NA>  <NA>  <NA>  <NA>
#> 57   <NA>  09.0  <NA>  <NA>  <NA>
#> 58   <NA>  <NA>  <NA>  <NA>  <NA>
#> 59    140   5.0  <NA>  <NA>  <NA>
#> 60   <NA>  <NA>  <NA>  <NA>  <NA>
#> 61   <NA>  <NA>  <NA>  <NA>  <NA>
#> 62   <NA>  <NA>  <NA>  <NA>  <NA>
#> 63   <NA>  <NA>  <NA>  <NA>  <NA>
#> 64   <NA>  09.0  <NA>  <NA>  <NA>
#> 65   <NA>  <NA>  <NA>  <NA>  <NA>
#> 66    140   5.0  <NA>  <NA>  <NA>
#> 67   <NA>  <NA>  <NA>  <NA>  <NA>
#> 68   <NA>  <NA>  <NA>  <NA>  <NA>
#> 69   <NA>  <NA>  <NA>  <NA>  <NA>
#> 70   <NA>  <NA>  <NA>  <NA>  <NA>
#> 71   <NA>  09.0  <NA>  <NA>  <NA>
#> 72   <NA>  <NA>  <NA>  <NA>  <NA>
#> 73    140   5.0  <NA>  <NA>  <NA>
#> 74   <NA>  <NA>  <NA>  <NA>  <NA>
#> 75   <NA>  <NA>  <NA>  <NA>  <NA>
#> 76   <NA>  <NA>  <NA>  <NA>  <NA>
#> 77   <NA>  <NA>  <NA>  <NA>  <NA>
#> 78   <NA>  09.0  <NA>  <NA>  <NA>
#> 79   <NA>  <NA>  <NA>  <NA>  <NA>
#> 80    140   5.0  <NA>  <NA>  <NA>
#> 81   <NA>  <NA>  <NA>  <NA>  <NA>
#> 82   <NA>  <NA>  <NA>  <NA>  <NA>
#> 83    140   6.0  <NA>  <NA>  <NA>
#> 84   <NA>  <NA>  <NA>  <NA>  <NA>
#> 85      4   340   1.5   2.2  <NA>
#> 86      N   079  <NA>  <NA>  <NA>
#> 87      2   100  <NA>  <NA>  <NA>
#> 88   <NA>  <NA>  <NA>  <NA>  <NA>
#> 89   <NA>  09.0  <NA>  <NA>  <NA>
#> 90   <NA>  <NA>  <NA>  <NA>  <NA>
#> 91    140   6.0  <NA>  <NA>  <NA>
#> 92      4   020   3.0   1.0  <NA>
#> 93      n  <NA>  <NA>  <NA>  <NA>
#> 94   <NA>  <NA>  <NA>  <NA>  <NA>
#> 95   <NA>  <NA>  <NA>  <NA>  <NA>
#> 96    2.7   257  <NA>  <NA>  <NA>
#> 97   <NA>  <NA>  <NA>  <NA>  <NA>
#> 98   <NA>  <NA>  <NA>  <NA>  <NA>
#> 99   <NA>  09.0  <NA>  <NA>  <NA>
#> 100  <NA>  <NA>  <NA>  <NA>  <NA>
#> 101   140   5.0  <NA>  <NA>  <NA>
#> 102  <NA>  <NA>  <NA>  <NA>  <NA>
#> 103     1   080   2.0   1.3   230
#> 104     N   079  <NA>  <NA>  <NA>
#> 105     5   100   000   000   000
#> 106     4   100   000   000   000
#> 107     1   100   000   000   000
#> 108     3   100   000   000   000
#> 109     5   100   000   000   000
#> 110     6   100   000   000   000
#> 111  <NA>  <NA>  <NA>  <NA>  <NA>
#> 112 1 is  a mat ch wi th th e sig
#> 113  <NA>  <NA>  <NA>  <NA>  <NA>
#> 114   1.3   252  <NA>  <NA>  <NA>
#> 115  <NA>  <NA>  <NA>  <NA>  <NA>
#> 116  <NA>  <NA>  <NA>  <NA>  <NA>
#> 117  <NA>  <NA>  <NA>  <NA>  <NA>
#> 118   0.6   125  <NA>  <NA>  <NA>
#> 119  <NA>  <NA>  <NA>  <NA>  <NA>
#> 120   0.4   125  <NA>  <NA>  <NA>
#> 121  <NA>  <NA>  <NA>  <NA>  <NA>
#> 122  <NA>  <NA>  <NA>  <NA>  <NA>
#> 123   0.3   125  <NA>  <NA>  <NA>
#> 124  <NA>  <NA>  <NA>  <NA>  <NA>
#> 125   0.5  <NA>  <NA>  <NA>  <NA>
#> 126  <NA>  <NA>  <NA>  <NA>  <NA>
#> 127  <NA>  <NA>  <NA>  <NA>  <NA>
#> 128  <NA>  <NA>  <NA>  <NA>  <NA>
#> 129   1.1   175  <NA>  <NA>  <NA>
#> 130  <NA>  <NA>  <NA>  <NA>  <NA>
#> 131  <NA>  <NA>  <NA>  <NA>  <NA>
#> 132   0.5   105  <NA>  <NA>  <NA>
#> 133  <NA>  <NA>  <NA>  <NA>  <NA>
#> 134  <NA>  <NA>  <NA>  <NA>  <NA>
#> 135 n fly ing b ridge .  be en tr
#> 136     Y  <NA>  <NA>  <NA>  <NA>
#> 137  <NA>  <NA>  <NA>  <NA>  <NA>
#> 138  <NA>  <NA>  <NA>  <NA>  <NA>
#> 139  <NA>  10.0  <NA>  <NA>  <NA>
#> 140  <NA>  <NA>  <NA>  <NA>  <NA>
#> 141   138   5.0  <NA>  <NA>  <NA>
#> 142  <NA>  <NA>  <NA>  <NA>  <NA>
#> 143  <NA>  10.0  <NA>  <NA>  <NA>
#> 144  <NA>  <NA>  <NA>  <NA>  <NA>
#> 145     1   070   2.2   1.2   090
#> 146     Y   096  <NA>  <NA>  <NA>
#> 147     3   100   000   000   000
#> 148     5   100   000   000   000
#> 149     3   100   000   000   000
#> 150     4   100   000   000   000
#> 151     1   100   000   000   000
#> 152     5   100   000   000   000
#> 153     3   100   000   000   000
#> 154  <NA>  <NA>  <NA>  <NA>  <NA>
#>                                               Data9       file_das event_num
#> 1                                              <NA> das_sample.das         1
#> 2                                              <NA> das_sample.das         2
#> 3                                              <NA> das_sample.das         3
#> 4                                              <NA> das_sample.das         4
#> 5                                              <NA> das_sample.das         5
#> 6                                              <NA> das_sample.das         6
#> 7                                              <NA> das_sample.das         7
#> 8                                              <NA> das_sample.das         8
#> 9                                              <NA> das_sample.das         9
#> 10                                             <NA> das_sample.das        10
#> 11                                             <NA> das_sample.das        11
#> 12                                             <NA> das_sample.das        12
#> 13                                             <NA> das_sample.das        13
#> 14                                              4.0 das_sample.das        14
#> 15                                             <NA> das_sample.das        15
#> 16                                             <NA> das_sample.das        NA
#> 17                                             <NA> das_sample.das        NA
#> 18                                             <NA> das_sample.das        16
#> 19                                             <NA> das_sample.das        17
#> 20                                             <NA> das_sample.das        18
#> 21                                             <NA> das_sample.das        19
#> 22                                             <NA> das_sample.das        20
#> 23                                             <NA> das_sample.das        21
#> 24                                             <NA> das_sample.das        22
#> 25                                             <NA> das_sample.das        23
#> 26                                             <NA> das_sample.das        24
#> 27                                             <NA> das_sample.das        25
#> 28                                             <NA> das_sample.das        26
#> 29                                             <NA> das_sample.das        27
#> 30                                             <NA> das_sample.das        28
#> 31                                              2.0 das_sample.das        29
#> 32                                             <NA> das_sample.das        30
#> 33                                             <NA> das_sample.das        NA
#> 34                                             <NA> das_sample.das        31
#> 35                                             <NA> das_sample.das        32
#> 36                                             <NA> das_sample.das        33
#> 37                                             <NA> das_sample.das        34
#> 38                                             <NA> das_sample.das        35
#> 39                                             <NA> das_sample.das        36
#> 40                                             <NA> das_sample.das        37
#> 41                                             <NA> das_sample.das        38
#> 42                                             <NA> das_sample.das        39
#> 43                                             <NA> das_sample.das        40
#> 44                                             <NA> das_sample.das        41
#> 45                                             <NA> das_sample.das        42
#> 46                                             <NA> das_sample.das        43
#> 47                                             <NA> das_sample.das        44
#> 48                                             <NA> das_sample.das        45
#> 49                                             <NA> das_sample.das        46
#> 50                                             <NA> das_sample.das        47
#> 51                                             <NA> das_sample.das        48
#> 52                                             <NA> das_sample.das        49
#> 53                                             <NA> das_sample.das        50
#> 54                                             <NA> das_sample.das        51
#> 55                                             <NA> das_sample.das        52
#> 56                                             <NA> das_sample.das        53
#> 57                                             <NA> das_sample.das        54
#> 58                                             <NA> das_sample.das        55
#> 59                                             <NA> das_sample.das        56
#> 60                                             <NA> das_sample.das        57
#> 61                                             <NA> das_sample.das        58
#> 62                                             <NA> das_sample.das        59
#> 63                                             <NA> das_sample.das        60
#> 64                                             <NA> das_sample.das        61
#> 65                                             <NA> das_sample.das        62
#> 66                                             <NA> das_sample.das        63
#> 67                                             <NA> das_sample.das        64
#> 68                                             <NA> das_sample.das        65
#> 69                                             <NA> das_sample.das        66
#> 70                                             <NA> das_sample.das        67
#> 71                                             <NA> das_sample.das        68
#> 72                                             <NA> das_sample.das        69
#> 73                                             <NA> das_sample.das        70
#> 74                                             <NA> das_sample.das        71
#> 75                                             <NA> das_sample.das        72
#> 76                                             <NA> das_sample.das        73
#> 77                                             <NA> das_sample.das        74
#> 78                                             <NA> das_sample.das        75
#> 79                                             <NA> das_sample.das        76
#> 80                                             <NA> das_sample.das        77
#> 81                                             <NA> das_sample.das        78
#> 82                                             <NA> das_sample.das        79
#> 83                                             <NA> das_sample.das        80
#> 84                                             <NA> das_sample.das        81
#> 85                                             <NA> das_sample.das        82
#> 86                                             <NA> das_sample.das        83
#> 87                                             <NA> das_sample.das        NA
#> 88                                             <NA> das_sample.das        84
#> 89                                             <NA> das_sample.das        85
#> 90                                             <NA> das_sample.das        86
#> 91                                             <NA> das_sample.das        87
#> 92                                             <NA> das_sample.das        88
#> 93                                             <NA> das_sample.das        89
#> 94                                             <NA> das_sample.das        90
#> 95                                             <NA> das_sample.das        91
#> 96                                             <NA> das_sample.das        92
#> 97                                             <NA> das_sample.das        93
#> 98                                             <NA> das_sample.das        94
#> 99                                             <NA> das_sample.das        95
#> 100                                            <NA> das_sample.das        96
#> 101                                            <NA> das_sample.das        97
#> 102                                            <NA> das_sample.das        98
#> 103                                        2.0 1014 das_sample.das        99
#> 104                                            <NA> das_sample.das       100
#> 105                                            <NA> das_sample.das        NA
#> 106                                            <NA> das_sample.das        NA
#> 107                                            <NA> das_sample.das        NA
#> 108                                            <NA> das_sample.das        NA
#> 109                                            <NA> das_sample.das        NA
#> 110                                            <NA> das_sample.das        NA
#> 111                                            <NA> das_sample.das       101
#> 112 hting 1014                                      das_sample.das       102
#> 113                                            <NA> das_sample.das       103
#> 114                                            <NA> das_sample.das       104
#> 115                                            <NA> das_sample.das       105
#> 116                                            <NA> das_sample.das       106
#> 117                                            <NA> das_sample.das       107
#> 118                                            <NA> das_sample.das       108
#> 119                                            <NA> das_sample.das       109
#> 120                                            <NA> das_sample.das       110
#> 121                                            <NA> das_sample.das       111
#> 122                                            <NA> das_sample.das       112
#> 123                                            <NA> das_sample.das       113
#> 124                                            <NA> das_sample.das       114
#> 125                                            <NA> das_sample.das       115
#> 126                                            <NA> das_sample.das       116
#> 127                                            <NA> das_sample.das       117
#> 128                                            <NA> das_sample.das       118
#> 129                                            <NA> das_sample.das       119
#> 130                                            <NA> das_sample.das       120
#> 131                                            <NA> das_sample.das       121
#> 132                                            <NA> das_sample.das       122
#> 133                                            <NA> das_sample.das       123
#> 134                                            <NA> das_sample.das       124
#> 135 acking sighting 081 and prey patches until dark das_sample.das       125
#> 136                                            <NA> das_sample.das         1
#> 137                                            <NA> das_sample.das         2
#> 138                                            <NA> das_sample.das         3
#> 139                                            <NA> das_sample.das         4
#> 140                                            <NA> das_sample.das         5
#> 141                                            <NA> das_sample.das         6
#> 142                                            <NA> das_sample.das         7
#> 143                                            <NA> das_sample.das         8
#> 144                                            <NA> das_sample.das         9
#> 145                                             4.0 das_sample.das        10
#> 146                                            <NA> das_sample.das        11
#> 147                                            <NA> das_sample.das        NA
#> 148                                            <NA> das_sample.das        NA
#> 149                                            <NA> das_sample.das        NA
#> 150                                            <NA> das_sample.das        NA
#> 151                                            <NA> das_sample.das        NA
#> 152                                            <NA> das_sample.das        NA
#> 153                                            <NA> das_sample.das        NA
#> 154                                            <NA> das_sample.das        12
#>     line_num
#> 1          1
#> 2          2
#> 3          3
#> 4          4
#> 5          5
#> 6          6
#> 7          7
#> 8          8
#> 9          9
#> 10        10
#> 11        11
#> 12        12
#> 13        13
#> 14        14
#> 15        15
#> 16        16
#> 17        17
#> 18        18
#> 19        19
#> 20        20
#> 21        21
#> 22        22
#> 23        23
#> 24        24
#> 25        25
#> 26        26
#> 27        27
#> 28        28
#> 29        29
#> 30        30
#> 31        31
#> 32        32
#> 33        33
#> 34        34
#> 35        35
#> 36        36
#> 37        37
#> 38        38
#> 39        39
#> 40        40
#> 41        41
#> 42        42
#> 43        43
#> 44        44
#> 45        45
#> 46        46
#> 47        47
#> 48        48
#> 49        49
#> 50        50
#> 51        51
#> 52        52
#> 53        53
#> 54        54
#> 55        55
#> 56        56
#> 57        57
#> 58        58
#> 59        59
#> 60        60
#> 61        61
#> 62        62
#> 63        63
#> 64        64
#> 65        65
#> 66        66
#> 67        67
#> 68        68
#> 69        69
#> 70        70
#> 71        71
#> 72        72
#> 73        73
#> 74        74
#> 75        75
#> 76        76
#> 77        77
#> 78        78
#> 79        79
#> 80        80
#> 81        81
#> 82        82
#> 83        83
#> 84        84
#> 85        85
#> 86        86
#> 87        87
#> 88        88
#> 89        89
#> 90        90
#> 91        91
#> 92        92
#> 93        93
#> 94        94
#> 95        95
#> 96        96
#> 97        97
#> 98        98
#> 99        99
#> 100      100
#> 101      101
#> 102      102
#> 103      103
#> 104      104
#> 105      105
#> 106      106
#> 107      107
#> 108      108
#> 109      109
#> 110      110
#> 111      111
#> 112      112
#> 113      113
#> 114      114
#> 115      115
#> 116      116
#> 117      117
#> 118      118
#> 119      119
#> 120      120
#> 121      121
#> 122      122
#> 123      123
#> 124      124
#> 125      125
#> 126      126
#> 127      127
#> 128      128
#> 129      129
#> 130      130
#> 131      131
#> 132      132
#> 133      133
#> 134      134
#> 135      135
#> 136      136
#> 137      137
#> 138      138
#> 139      139
#> 140      140
#> 141      141
#> 142      142
#> 143      143
#> 144      144
#> 145      145
#> 146      146
#> 147      147
#> 148      148
#> 149      149
#> 150      150
#> 151      151
#> 152      152
#> 153      153
#> 154      154