DAS sightings

Extract sightings and associated information from processed DAS data

das_sight(x, mixed.multi)

# S3 method for data.frame
das_sight(x, mixed.multi = FALSE)

# S3 method for das_df
das_sight(x, mixed.multi = FALSE)

Arguments

x
das_df object; output from das_process. Can also be a data frame that can be coerced to a das_df object

mixed.multi
logical; indicates if mixed-species sightings should be output in multiple rows

Value

Data frame with 1) the columns from x and 2) columns with sighting information (observer, species, etc.) extracted from 'Data#' columns as specified in Details. The data frame has one row for each sighting, or one row for each species of each sighting if mixed.multi is TRUE.

Details

DAS events contain specific information in the 'Data#' columns, with the information depending on the event code for that row. The output data frame contains columns with this specific information extracted to dedicated columns as described below. This function recognizes the following types of sightings: marine mammal sightings (event codes "S", "K", or "M"), marine mammal resights (codes "s" or "K"), turtle sightings (code "t"), and fishing vessel sightings (code "F"). See the format PDF, das_format_pdf, for the required format for each of these events.

Abbreviations used in column names: Gs = group size, Sp = species, Nm = nautical mile, Perc = percentage

This function makes the following assumptions, and alterations to the raw data:

"S", "K", and "M" events, and only these events, are immediately followed by an "A" event

The 'Mixed' column is TRUE if two or more of the 'Data5', 'Data6', 'Data7', and 'Data8' values for the corresponding "A" event are not NA, and FALSE otherwise. This column has a non-NA value for only "S", "K", and "M" events

The 'Prob' column is TRUE is the sighting has an associated "?" event, and FALSE otherwise. This column has a non-NA value for only "S", "K", and "M" events

The 'GsTotal' column is the mean of the 'Data2' columns (with NAs removed) for the associated "1"-"8" events

The 'Sp1Perc', 'Sp2Perc', 'Sp3Perc', and 'Sp4Perc' columns are the means of the 'Data5', 'Data6', 'Data7', and 'Data8' columns (with NAs removed), respectively, for the associated "1"-"8" events

The values for the following columns were capitalized using toupper: 'Birds', 'Photos', 'turtle_age', and 'turtle_captured'

Outstanding questions:

Should any columns be converted to logicals, e.g. 'Birds', 'Photos', 'turtle_age', and 'turtle_captured'?

Should NA values for columns 'Sp1', 'Sp1Perc', 'GsSp1', etc., be changed from NA to 0?

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_sight(y.proc)

#>    Event            DateTime      Lat       Lon OnEffort Cruise Mode EffType
#> 1      S 2009-11-29 12:17:00 24.82967 -124.2333     TRUE   1000    C       S
#> 2      S 2009-11-29 13:07:16 24.81100 -124.3847     TRUE   1000    C       S
#> 3      t 2009-11-29 13:55:54 24.76750 -124.5145     TRUE   1000    C       S
#> 4      K 2009-11-29 16:28:37 24.66583 -124.9593     TRUE   1000    C       S
#> 5      k 2009-11-29 16:37:44 24.66050 -124.9855    FALSE   1000    C       S
#> 6      S 2009-11-29 16:47:22 24.65667 -125.0063     TRUE   1000    C       S
#> 7      s 2009-11-29 16:58:43 24.64167 -125.0177    FALSE   1000    C       S
#> 8      s 2009-11-29 17:21:01 24.62033 -125.0017    FALSE   1000    C       S
#> 9      s 2009-11-29 17:29:44 24.60983 -125.0000    FALSE   1000    C       S
#> 10     s 2009-11-29 17:50:15 24.61433 -124.9953    FALSE   1000    C       S
#> 11     F 2009-11-29 17:51:13 24.61550 -124.9970    FALSE   1000    C       S
#> 12     s 2009-11-29 18:10:25 24.62150 -125.0135    FALSE   1000    C       S
#> 13     s 2009-11-29 18:22:24 24.61950 -124.9995    FALSE   1000    C       S
#> 14     S 2009-11-30 06:53:54 24.38233 -126.4047     TRUE   1000    C       S
#>    Course Bft SwellHght RainFog HorizSun VertSun Glare Vis Data1 Data2 Data3
#> 1     259   3         4       1        4       1 FALSE   6   079   073     3
#> 2     259   3         4       1        3       1 FALSE   7   080   001     3
#> 3     252   3         4       1       NA      NA    NA   6   001    DC   180
#> 4     252   3         4       1       NA      NA    NA   6  1014   186     6
#> 5     252   3         4       1       NA      NA    NA   6  1014   315   1.0
#> 6     252   3         4       3       NA      NA    NA   5   081   126     6
#> 7     252   3         4       3       NA      NA    NA   5   081   330   2.0
#> 8     125   3         4       3       NA      NA    NA   5   081   329   6.0
#> 9     125   3         4       3       NA      NA    NA   5   081   320   9.0
#> 10    125   3         4       3       NA      NA    NA   5   081   280  15.0
#> 11    290   3         4       3       NA      NA    NA   5   126   000   2.7
#> 12    175   3         4       3       NA      NA    NA   5   081   341   2.5
#> 13    105   3         4       3       NA      NA    NA   5   081   000  <NA>
#> 14    259   3         6       1       NA      NA    NA   5   082   153     6
#>    Data4 Data5 Data6 Data7 Data8 Data9 EventNum       file_das line_num  Obs
#> 1      4   012   0.2   3.6   210   4.0       14 das_sample.das       14  073
#> 2      4   025   0.6   2.5   330   2.0       29 das_sample.das       31  001
#> 3    0.2     1     J  12.2     A     N       43 das_sample.das       46  001
#> 4      4   340   1.5   2.2  <NA>  <NA>       83 das_sample.das       86  186
#> 5    2.7   257  <NA>  <NA>  <NA>  <NA>       93 das_sample.das       97 <NA>
#> 6      1   080   2.0   1.3   230   2.0      100 das_sample.das      104  126
#> 7    1.3   252  <NA>  <NA>  <NA>  <NA>      105 das_sample.das      115 <NA>
#> 8    0.6   125  <NA>  <NA>  <NA>  <NA>      109 das_sample.das      119 <NA>
#> 9    0.4   125  <NA>  <NA>  <NA>  <NA>      111 das_sample.das      121 <NA>
#> 10   0.3   125  <NA>  <NA>  <NA>  <NA>      114 das_sample.das      124 <NA>
#> 11   0.5  <NA>  <NA>  <NA>  <NA>  <NA>      116 das_sample.das      126  126
#> 12   1.1   175  <NA>  <NA>  <NA>  <NA>      120 das_sample.das      130 <NA>
#> 13   0.5   105  <NA>  <NA>  <NA>  <NA>      123 das_sample.das      133 <NA>
#> 14     1   070   2.2   1.2   090   4.0       10 das_sample.das      146  153
#>    Bearing Reticle DistNm SightNo Cue Method Photos Birds Mixed  Prob   GsTotal
#> 1      012     0.2    3.6     079   3      4      N     N FALSE  TRUE  3.000000
#> 2      025     0.6    2.5     080   3      4      N     N  TRUE FALSE 20.000000
#> 3      180    12.2    0.2    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 4      340     1.5    2.2    1014   6      4      N     N FALSE FALSE  4.000000
#> 5      315     1.0    2.7    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 6      080     2.0    1.3     081   6      1      N     N FALSE FALSE  4.500000
#> 7      330     2.0    1.3    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 8      329     6.0    0.6    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 9      320     9.0    0.4    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 10     280    15.0    0.3    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 11     000     0.5    2.7    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 12     341     2.5    1.1    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 13     000    <NA>    0.5    <NA>  NA     NA   <NA>  <NA>    NA    NA        NA
#> 14     070     2.2    1.2     082   6      1      N     Y FALSE FALSE  4.714286
#>     Sp1  Sp2  Sp3  Sp4 Sp1Perc Sp2Perc Sp3Perc Sp4Perc     GsSp1 GsSp2 GsSp3
#> 1   075 <NA> <NA> <NA>     100       0       0       0  3.000000     0     0
#> 2   018  016 <NA> <NA>      60      40       0       0 12.000000     8     0
#> 3  <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 4   079 <NA> <NA> <NA>     100     NaN     NaN     NaN  4.000000   NaN   NaN
#> 5  <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 6   079 <NA> <NA> <NA>     100       0       0       0  4.500000     0     0
#> 7  <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 8  <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 9  <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 10 <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 11 <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 12 <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 13 <NA> <NA> <NA> <NA>      NA      NA      NA      NA        NA    NA    NA
#> 14  096 <NA> <NA> <NA>     100       0       0       0  4.714286     0     0
#>    GsSp4 ResightCourse TurtleSp TurtleNum TurtleJFR TurtleAge TurtleCapt
#> 1      0            NA     <NA>        NA      <NA>      <NA>       <NA>
#> 2      0            NA     <NA>        NA      <NA>      <NA>       <NA>
#> 3     NA            NA       DC         1         J         A          N
#> 4    NaN            NA     <NA>        NA      <NA>      <NA>       <NA>
#> 5     NA           257     <NA>        NA      <NA>      <NA>       <NA>
#> 6      0            NA     <NA>        NA      <NA>      <NA>       <NA>
#> 7     NA           252     <NA>        NA      <NA>      <NA>       <NA>
#> 8     NA           125     <NA>        NA      <NA>      <NA>       <NA>
#> 9     NA           125     <NA>        NA      <NA>      <NA>       <NA>
#> 10    NA           125     <NA>        NA      <NA>      <NA>       <NA>
#> 11    NA            NA     <NA>        NA      <NA>      <NA>       <NA>
#> 12    NA           175     <NA>        NA      <NA>      <NA>       <NA>
#> 13    NA           105     <NA>        NA      <NA>      <NA>       <NA>
#> 14     0            NA     <NA>        NA      <NA>      <NA>       <NA>
#>    BoatType BoatNum
#> 1      <NA>      NA
#> 2      <NA>      NA
#> 3      <NA>      NA
#> 4      <NA>      NA
#> 5      <NA>      NA
#> 6      <NA>      NA
#> 7      <NA>      NA
#> 8      <NA>      NA
#> 9      <NA>      NA
#> 10     <NA>      NA
#> 11     <NA>      NA
#> 12     <NA>      NA
#> 13     <NA>      NA
#> 14     <NA>      NA