Read DAS file(s)

Read one or more fixed-width DAS text file(s) generated by WinCruz into a data frame, where each line is data for a specific event

das_read(file, tz = "UTC")

Arguments

file
filename(s) of one or more DAS files

tz
character; see strptime. Default is UTC

Value

A das_dfr object, which is also a data frame, with DAS data read into columns. The data are read into the data frame as characters as described in 'Details', with the following exceptions:

Name	Class	Details

EffortDot	logical	TRUE if "." was present, and FALSE otherwise

DateTime	POSIXct	combination of 'Date' and 'Time' columns with time zone tz

Lat	numeric	'Latitude' column converted to decimal degrees in range [-90, 90]

Lon	numeric	'Longitude' column converted to decimal degrees in range [-180, 180]

Data#	character	leading/trailing whitespace trimmed for non-comment events (i.e. rows where 'Event' is not "C" or "c")

EventNum	integer	'Event number' coerced to an integer using as.integer

file_das	character	filename, extracted from the file argument using basename

line_num	integer	line number of each data row

Details

Reads/parses DAS data into columns of a data frame. If file contains multiple filenames, then the individual data frames will be concatenated.

The provided DAS file must adhere to the following column number and format specifications; note that 'Data#' columns may be referred to as 'Field#' columns in other documentation:

Item	Columns	Format

Event number	1-3	Event

4	Effort dot	5

Time	6-11	HHMMSS or HHMM

Date	13-18	MMDDYY

Latitude	20-28	NDD:MM.MM

Longitude	30-39	WDDD:MM.MM

Data1	40-44	Data2

45-49	Data3	50-54

Data4	55-59	Data5

60-64	Data6	65-69

Data7	70-74	Data8

75-79	Data9	80+

See das_format_pdf for more information about DAS format requirements. This function was inspired by das.read

See also

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
das_read(y)

#>     Event EffortDot            DateTime      Lat       Lon Data1 Data2 Data3
#> 1       B      TRUE 2009-11-29 11:49:09 24.82950 -124.1518  1000     C     7
#> 2       R      TRUE 2009-11-29 11:49:09 24.82950 -124.1518     S  <NA>  <NA>
#> 3       P      TRUE 2009-11-29 11:49:09 24.82950 -124.1518   153   073   126
#> 4       V      TRUE 2009-11-29 11:49:09 24.82950 -124.1518     3    04   120
#> 5       N      TRUE 2009-11-29 11:49:09 24.82950 -124.1518   259  10.0  <NA>
#> 6       W      TRUE 2009-11-29 11:49:09 24.82950 -124.1518     3    04    01
#> 7       *      TRUE 2009-11-29 11:55:54 24.83450 -124.1697  <NA>  <NA>  <NA>
#> 8       P      TRUE 2009-11-29 12:00:34 24.83383 -124.1835   126   185   073
#> 9       V      TRUE 2009-11-29 12:00:34 24.83383 -124.1835     3    04   120
#> 10      N      TRUE 2009-11-29 12:00:34 24.83383 -124.1835   259  10.0  <NA>
#> 11      W      TRUE 2009-11-29 12:00:34 24.83383 -124.1835     1    04    01
#> 12      *      TRUE 2009-11-29 12:05:54 24.83250 -124.1992  <NA>  <NA>  <NA>
#> 13      *      TRUE 2009-11-29 12:15:54 24.83000 -124.2300  <NA>  <NA>  <NA>
#> 14      S      TRUE 2009-11-29 12:17:00 24.82967 -124.2333   079   073     3
#> 15      A      TRUE 2009-11-29 12:17:00 24.82967 -124.2333   079  <NA>     N
#> 16      ?     FALSE                <NA>       NA        NA   079  <NA>  <NA>
#> 17      1     FALSE                <NA>       NA        NA   073     3     3
#> 18      *      TRUE 2009-11-29 12:25:54 24.82750 -124.2607  <NA>  <NA>  <NA>
#> 19      P      TRUE 2009-11-29 12:29:48 24.82633 -124.2728   184   126   185
#> 20      V      TRUE 2009-11-29 12:29:48 24.82633 -124.2728     3    04   120
#> 21      N      TRUE 2009-11-29 12:29:48 24.82633 -124.2728   259  10.0  <NA>
#> 22      W      TRUE 2009-11-29 12:29:48 24.82633 -124.2728     1    04    01
#> 23      *      TRUE 2009-11-29 12:35:54 24.82467 -124.2917  <NA>  <NA>  <NA>
#> 24      *      TRUE 2009-11-29 12:45:54 24.82217 -124.3220  <NA>  <NA>  <NA>
#> 25      *      TRUE 2009-11-29 12:55:54 24.81833 -124.3515  <NA>  <NA>  <NA>
#> 26      P      TRUE 2009-11-29 13:00:20 24.81583 -124.3647   001   184   126
#> 27      V      TRUE 2009-11-29 13:00:20 24.81583 -124.3647     3    04   180
#> 28      N      TRUE 2009-11-29 13:00:20 24.81583 -124.3647   259  10.0  <NA>
#> 29      W      TRUE 2009-11-29 13:00:20 24.81583 -124.3647     1    03    01
#> 30      *      TRUE 2009-11-29 13:05:54 24.81183 -124.3807  <NA>  <NA>  <NA>
#> 31      S      TRUE 2009-11-29 13:07:16 24.81100 -124.3847   080   001     3
#> 32      A      TRUE 2009-11-29 13:07:16 24.81100 -124.3847   080  <NA>     N
#> 33      1     FALSE                <NA>       NA        NA   001    20    30
#> 34      E     FALSE 2009-11-29 13:07:53 24.81067 -124.3865     C  <NA>  <NA>
#> 35      *     FALSE 2009-11-29 13:15:54 24.80100 -124.4070  <NA>  <NA>  <NA>
#> 36      N     FALSE 2009-11-29 13:19:07 24.79517 -124.4140   220  10.0  <NA>
#> 37      *     FALSE 2009-11-29 13:25:54 24.78817 -124.4263  <NA>  <NA>  <NA>
#> 38      R      TRUE 2009-11-29 13:34:27 24.78183 -124.4513     S  <NA>  <NA>
#> 39      P      TRUE 2009-11-29 13:34:27 24.78183 -124.4513   182   001   186
#> 40      V      TRUE 2009-11-29 13:34:27 24.78183 -124.4513     3    04   180
#> 41      N      TRUE 2009-11-29 13:34:27 24.78183 -124.4513   252  10.0  <NA>
#> 42      W      TRUE 2009-11-29 13:34:27 24.78183 -124.4513     1  <NA>  <NA>
#> 43      *      TRUE 2009-11-29 13:35:54 24.78117 -124.4557  <NA>  <NA>  <NA>
#> 44      *      TRUE 2009-11-29 13:45:54 24.77450 -124.4850  <NA>  <NA>  <NA>
#> 45      *      TRUE 2009-11-29 13:55:54 24.76750 -124.5145  <NA>  <NA>  <NA>
#> 46      t      TRUE 2009-11-29 13:55:54 24.76750 -124.5145   001    DC   180
#> 47      V      TRUE 2009-11-29 13:56:10 24.76733 -124.5153     3    04   180
#> 48      W      TRUE 2009-11-29 13:56:25 24.76717 -124.5160     1  <NA>  <NA>
#> 49      P      TRUE 2009-11-29 14:01:49 24.76450 -124.5318   153   182   001
#> 50      V      TRUE 2009-11-29 14:01:49 24.76450 -124.5318     3    04   180
#> 51      N      TRUE 2009-11-29 14:01:49 24.76450 -124.5318   252  10.0  <NA>
#> 52      W      TRUE 2009-11-29 14:01:49 24.76450 -124.5318     3  <NA>  <NA>
#> 53      *      TRUE 2009-11-29 14:05:54 24.76133 -124.5435  <NA>  <NA>  <NA>
#> 54      *      TRUE 2009-11-29 14:15:54 24.75417 -124.5728  <NA>  <NA>  <NA>
#> 55      *      TRUE 2009-11-29 14:25:54 24.74767 -124.6020  <NA>  <NA>  <NA>
#> 56      W      TRUE 2009-11-29 14:29:31 24.74533 -124.6125     3  <NA>  <NA>
#> 57      P      TRUE 2009-11-29 14:31:21 24.74483 -124.6178   186   153   182
#> 58      V      TRUE 2009-11-29 14:31:21 24.74483 -124.6178     3    04   180
#> 59      N      TRUE 2009-11-29 14:31:21 24.74483 -124.6178   252  10.0  <NA>
#> 60      W      TRUE 2009-11-29 14:31:21 24.74483 -124.6178     3  <NA>  <NA>
#> 61      *      TRUE 2009-11-29 14:35:54 24.74117 -124.6310  <NA>  <NA>  <NA>
#> 62      *      TRUE 2009-11-29 14:45:54 24.73417 -124.6608  <NA>  <NA>  <NA>
#> 63      *      TRUE 2009-11-29 14:55:54 24.72833 -124.6897  <NA>  <NA>  <NA>
#> 64      P      TRUE 2009-11-29 14:59:57 24.72617 -124.7007   073   186   153
#> 65      V      TRUE 2009-11-29 14:59:57 24.72617 -124.7007     3    04   180
#> 66      N      TRUE 2009-11-29 14:59:57 24.72617 -124.7007   252  10.0  <NA>
#> 67      W      TRUE 2009-11-29 14:59:57 24.72617 -124.7007     3  <NA>  <NA>
#> 68      *      TRUE 2009-11-29 15:05:54 24.72217 -124.7183  <NA>  <NA>  <NA>
#> 69      *      TRUE 2009-11-29 15:15:54 24.71467 -124.7470  <NA>  <NA>  <NA>
#> 70      *      TRUE 2009-11-29 15:25:54 24.70817 -124.7755  <NA>  <NA>  <NA>
#> 71      P      TRUE 2009-11-29 15:30:31 24.70467 -124.7888   185   073   186
#> 72      V      TRUE 2009-11-29 15:30:31 24.70467 -124.7888     3    04   180
#> 73      N      TRUE 2009-11-29 15:30:31 24.70467 -124.7888   252  10.0  <NA>
#> 74      W      TRUE 2009-11-29 15:30:31 24.70467 -124.7888     3  <NA>  <NA>
#> 75      *      TRUE 2009-11-29 15:35:54 24.70050 -124.8047  <NA>  <NA>  <NA>
#> 76      *      TRUE 2009-11-29 15:45:54 24.69367 -124.8338  <NA>  <NA>  <NA>
#> 77      *      TRUE 2009-11-29 15:55:54 24.68600 -124.8622  <NA>  <NA>  <NA>
#> 78      P      TRUE 2009-11-29 16:00:09 24.68350 -124.8753   126   185   073
#> 79      V      TRUE 2009-11-29 16:00:09 24.68350 -124.8753     3    04   180
#> 80      N      TRUE 2009-11-29 16:00:09 24.68350 -124.8753   252  10.0  <NA>
#> 81      W      TRUE 2009-11-29 16:00:09 24.68350 -124.8753     3  <NA>  <NA>
#> 82      *      TRUE 2009-11-29 16:05:54 24.67933 -124.8917  <NA>  <NA>  <NA>
#> 83      *      TRUE 2009-11-29 16:15:54 24.67333 -124.9208  <NA>  <NA>  <NA>
#> 84      W      TRUE 2009-11-29 16:17:31 24.67217 -124.9255     1  <NA>  <NA>
#> 85      *      TRUE 2009-11-29 16:25:54 24.66750 -124.9513  <NA>  <NA>  <NA>
#> 86      K      TRUE 2009-11-29 16:28:37 24.66583 -124.9593  1014   186     6
#> 87      A      TRUE 2009-11-29 16:28:37 24.66583 -124.9593  1014  <NA>     N
#> 88      1     FALSE                <NA>       NA        NA   186     4     4
#> 89      P      TRUE 2009-11-29 16:30:15 24.66500 -124.9640   184   126   185
#> 90      V      TRUE 2009-11-29 16:30:15 24.66500 -124.9640     3    04   180
#> 91      N      TRUE 2009-11-29 16:30:15 24.66500 -124.9640   252  10.0  <NA>
#> 92      W      TRUE 2009-11-29 16:30:15 24.66500 -124.9640     1  <NA>  <NA>
#> 93      #      TRUE 2009-11-29 16:33:57 24.66317 -124.9750   081   185     2
#> 94      #      TRUE 2009-11-29 16:33:57 24.66317 -124.9750   081  <NA>     N
#> 95      *      TRUE 2009-11-29 16:35:54 24.66150 -124.9803  <NA>  <NA>  <NA>
#> 96      E     FALSE 2009-11-29 16:36:34 24.66100 -124.9822     U  <NA>  <NA>
#> 97      k     FALSE 2009-11-29 16:37:44 24.66050 -124.9855  1014   315   1.0
#> 98      R      TRUE 2009-11-29 16:43:46 24.65867 -124.9965     S  <NA>  <NA>
#> 99      P      TRUE 2009-11-29 16:43:46 24.65867 -124.9965   184   126   185
#> 100     V      TRUE 2009-11-29 16:43:46 24.65867 -124.9965     3    04   180
#> 101     N      TRUE 2009-11-29 16:43:46 24.65867 -124.9965   252  10.0  <NA>
#> 102     W      TRUE 2009-11-29 16:43:46 24.65867 -124.9965     3  <NA>  <NA>
#> 103     *      TRUE 2009-11-29 16:45:54 24.65750 -125.0018  <NA>  <NA>  <NA>
#> 104     S      TRUE 2009-11-29 16:47:22 24.65667 -125.0063   081   126     6
#> 105     A      TRUE 2009-11-29 16:47:22 24.65667 -125.0063   081  <NA>     N
#> 106     1     FALSE                <NA>       NA        NA   001     3     5
#> 107     2     FALSE                <NA>       NA        NA   073     8    10
#> 108     3     FALSE                <NA>       NA        NA   126     2     2
#> 109     4     FALSE                <NA>       NA        NA   153     5     6
#> 110     5     FALSE                <NA>       NA        NA   184     4     6
#> 111     6     FALSE                <NA>       NA        NA   185     5     6
#> 112     E     FALSE 2009-11-29 16:48:44 24.65567 -125.0103     C  <NA>  <NA>
#> 113     C     FALSE 2009-11-29 16:51:19 24.65317 -125.0172  The  sight ing 8
#> 114     *     FALSE 2009-11-29 16:55:54 24.64467 -125.0180  <NA>  <NA>  <NA>
#> 115     s     FALSE 2009-11-29 16:58:43 24.64167 -125.0177   081   330   2.0
#> 116     N     FALSE 2009-11-29 17:04:50 24.63483 -125.0137   125  10.0  <NA>
#> 117     *     FALSE 2009-11-29 17:05:54 24.63383 -125.0122  <NA>  <NA>  <NA>
#> 118     *     FALSE 2009-11-29 17:15:54 24.62617 -125.0032  <NA>  <NA>  <NA>
#> 119     s     FALSE 2009-11-29 17:21:01 24.62033 -125.0017   081   329   6.0
#> 120     *     FALSE 2009-11-29 17:25:54 24.61433 -125.0007  <NA>  <NA>  <NA>
#> 121     s     FALSE 2009-11-29 17:29:44 24.60983 -125.0000   081   320   9.0
#> 122     *     FALSE 2009-11-29 17:35:54 24.60583 -124.9942  <NA>  <NA>  <NA>
#> 123     *     FALSE 2009-11-29 17:45:54 24.60967 -124.9880  <NA>  <NA>  <NA>
#> 124     s     FALSE 2009-11-29 17:50:15 24.61433 -124.9953   081   280  15.0
#> 125     N     FALSE 2009-11-29 17:51:13 24.61550 -124.9970   290  10.0  <NA>
#> 126     F     FALSE 2009-11-29 17:51:13 24.61550 -124.9970   126   000   2.7
#> 127     *     FALSE 2009-11-29 17:55:54 24.62000 -125.0048  <NA>  <NA>  <NA>
#> 128     *     FALSE 2009-11-29 18:05:54 24.62400 -125.0178  <NA>  <NA>  <NA>
#> 129     N     FALSE 2009-11-29 18:06:51 24.62350 -125.0167   175  10.0  <NA>
#> 130     s     FALSE 2009-11-29 18:10:25 24.62150 -125.0135   081   341   2.5
#> 131     *     FALSE 2009-11-29 18:15:54 24.62050 -125.0072  <NA>  <NA>  <NA>
#> 132     N     FALSE 2009-11-29 18:18:49 24.62033 -125.0037   105  10.0  <NA>
#> 133     s     FALSE 2009-11-29 18:22:24 24.61950 -124.9995   081   000  <NA>
#> 134     *     FALSE 2009-11-29 18:25:54 24.61667 -124.9972  <NA>  <NA>  <NA>
#> 135     *     FALSE 2009-11-29 18:35:54 24.61133 -124.9875  <NA>  <NA>  <NA>
#> 136     C     FALSE 2009-11-29 18:39:42 24.61100 -124.9868  clos ing d own o
#> 137     B      TRUE 2009-11-30 06:31:44 24.39583 -126.3413  1000     C     7
#> 138     R      TRUE 2009-11-30 06:31:44 24.39583 -126.3413     S  <NA>  <NA>
#> 139     P      TRUE 2009-11-30 06:31:44 24.39583 -126.3413   186   153   182
#> 140     V      TRUE 2009-11-30 06:31:44 24.39583 -126.3413     3    04   125
#> 141     N      TRUE 2009-11-30 06:31:44 24.39583 -126.3413   259  10.0  <NA>
#> 142     W      TRUE 2009-11-30 06:31:44 24.39583 -126.3413     1  <NA>  <NA>
#> 143     *      TRUE 2009-11-30 06:41:41 24.38967 -126.3698  <NA>  <NA>  <NA>
#> 144     V      TRUE 2009-11-30 06:47:57 24.38633 -126.3882     3    06   175
#> 145     *      TRUE 2009-11-30 06:51:41 24.38400 -126.3985  <NA>  <NA>  <NA>
#> 146     S      TRUE 2009-11-30 06:53:54 24.38233 -126.4047   082   153     6
#> 147     A      TRUE 2009-11-30 06:53:54 24.38233 -126.4047   082  <NA>     N
#> 148     1     FALSE                <NA>       NA        NA   001     3     4
#> 149     2     FALSE                <NA>       NA        NA   073     6    10
#> 150     3     FALSE                <NA>       NA        NA   126     4     4
#> 151     4     FALSE                <NA>       NA        NA   153     7     8
#> 152     5     FALSE                <NA>       NA        NA   182     4     6
#> 153     6     FALSE                <NA>       NA        NA   185     4     5
#> 154     7     FALSE                <NA>       NA        NA   186     5     5
#> 155     E     FALSE 2009-11-30 06:53:54 24.38233 -126.4047     C  <NA>  <NA>
#>     Data4 Data5 Data6 Data7 Data8
#> 1       Y  <NA>  <NA>  <NA>  <NA>
#> 2    <NA>  <NA>  <NA>  <NA>  <NA>
#> 3    <NA>  <NA>  <NA>  <NA>  <NA>
#> 4    <NA>  09.0  <NA>  <NA>  <NA>
#> 5    <NA>  <NA>  <NA>  <NA>  <NA>
#> 6     140   6.0  <NA>  <NA>  <NA>
#> 7    <NA>  <NA>  <NA>  <NA>  <NA>
#> 8    <NA>  <NA>  <NA>  <NA>  <NA>
#> 9    <NA>  09.0  <NA>  <NA>  <NA>
#> 10   <NA>  <NA>  <NA>  <NA>  <NA>
#> 11    140   6.0  <NA>  <NA>  <NA>
#> 12   <NA>  <NA>  <NA>  <NA>  <NA>
#> 13   <NA>  <NA>  <NA>  <NA>  <NA>
#> 14      4   012   0.2   3.6   210
#> 15      N   075  <NA>  <NA>  <NA>
#> 16   <NA>   074  <NA>  <NA>  <NA>
#> 17      3   100   000   000   000
#> 18   <NA>  <NA>  <NA>  <NA>  <NA>
#> 19   <NA>  <NA>  <NA>  <NA>  <NA>
#> 20   <NA>  09.0  <NA>  <NA>  <NA>
#> 21   <NA>  <NA>  <NA>  <NA>  <NA>
#> 22    140   6.0  <NA>  <NA>  <NA>
#> 23   <NA>  <NA>  <NA>  <NA>  <NA>
#> 24   <NA>  <NA>  <NA>  <NA>  <NA>
#> 25   <NA>  <NA>  <NA>  <NA>  <NA>
#> 26   <NA>  <NA>  <NA>  <NA>  <NA>
#> 27   <NA>  09.0  <NA>  <NA>  <NA>
#> 28   <NA>  <NA>  <NA>  <NA>  <NA>
#> 29    140   7.0  <NA>  <NA>  <NA>
#> 30   <NA>  <NA>  <NA>  <NA>  <NA>
#> 31      4   025   0.6   2.5   330
#> 32      N   018   016  <NA>  <NA>
#> 33     10    60    40   000   000
#> 34   <NA>  <NA>  <NA>  <NA>  <NA>
#> 35   <NA>  <NA>  <NA>  <NA>  <NA>
#> 36   <NA>  <NA>  <NA>  <NA>  <NA>
#> 37   <NA>  <NA>  <NA>  <NA>  <NA>
#> 38   <NA>  <NA>  <NA>  <NA>  <NA>
#> 39   <NA>  <NA>  <NA>  <NA>  <NA>
#> 40   <NA>  09.0  <NA>  <NA>  <NA>
#> 41   <NA>  <NA>  <NA>  <NA>  <NA>
#> 42    140   6.0  <NA>  <NA>  <NA>
#> 43   <NA>  <NA>  <NA>  <NA>  <NA>
#> 44   <NA>  <NA>  <NA>  <NA>  <NA>
#> 45   <NA>  <NA>  <NA>  <NA>  <NA>
#> 46    0.2     1     J  12.2     A
#> 47   <NA>  09.0  <NA>  <NA>  <NA>
#> 48    140   5.0  <NA>  <NA>  <NA>
#> 49   <NA>  <NA>  <NA>  <NA>  <NA>
#> 50   <NA>  09.0  <NA>  <NA>  <NA>
#> 51   <NA>  <NA>  <NA>  <NA>  <NA>
#> 52    140   4.0  <NA>  <NA>  <NA>
#> 53   <NA>  <NA>  <NA>  <NA>  <NA>
#> 54   <NA>  <NA>  <NA>  <NA>  <NA>
#> 55   <NA>  <NA>  <NA>  <NA>  <NA>
#> 56    140   5.0  <NA>  <NA>  <NA>
#> 57   <NA>  <NA>  <NA>  <NA>  <NA>
#> 58   <NA>  09.0  <NA>  <NA>  <NA>
#> 59   <NA>  <NA>  <NA>  <NA>  <NA>
#> 60    140   5.0  <NA>  <NA>  <NA>
#> 61   <NA>  <NA>  <NA>  <NA>  <NA>
#> 62   <NA>  <NA>  <NA>  <NA>  <NA>
#> 63   <NA>  <NA>  <NA>  <NA>  <NA>
#> 64   <NA>  <NA>  <NA>  <NA>  <NA>
#> 65   <NA>  09.0  <NA>  <NA>  <NA>
#> 66   <NA>  <NA>  <NA>  <NA>  <NA>
#> 67    140   5.0  <NA>  <NA>  <NA>
#> 68   <NA>  <NA>  <NA>  <NA>  <NA>
#> 69   <NA>  <NA>  <NA>  <NA>  <NA>
#> 70   <NA>  <NA>  <NA>  <NA>  <NA>
#> 71   <NA>  <NA>  <NA>  <NA>  <NA>
#> 72   <NA>  09.0  <NA>  <NA>  <NA>
#> 73   <NA>  <NA>  <NA>  <NA>  <NA>
#> 74    140   5.0  <NA>  <NA>  <NA>
#> 75   <NA>  <NA>  <NA>  <NA>  <NA>
#> 76   <NA>  <NA>  <NA>  <NA>  <NA>
#> 77   <NA>  <NA>  <NA>  <NA>  <NA>
#> 78   <NA>  <NA>  <NA>  <NA>  <NA>
#> 79   <NA>  09.0  <NA>  <NA>  <NA>
#> 80   <NA>  <NA>  <NA>  <NA>  <NA>
#> 81    140   5.0  <NA>  <NA>  <NA>
#> 82   <NA>  <NA>  <NA>  <NA>  <NA>
#> 83   <NA>  <NA>  <NA>  <NA>  <NA>
#> 84    140   6.0  <NA>  <NA>  <NA>
#> 85   <NA>  <NA>  <NA>  <NA>  <NA>
#> 86      4   340   1.5   2.2  <NA>
#> 87      N   079  <NA>  <NA>  <NA>
#> 88      4   100  <NA>  <NA>  <NA>
#> 89   <NA>  <NA>  <NA>  <NA>  <NA>
#> 90   <NA>  09.0  <NA>  <NA>  <NA>
#> 91   <NA>  <NA>  <NA>  <NA>  <NA>
#> 92    140   6.0  <NA>  <NA>  <NA>
#> 93      4   020   3.0   1.0  <NA>
#> 94      n  <NA>  <NA>  <NA>  <NA>
#> 95   <NA>  <NA>  <NA>  <NA>  <NA>
#> 96   <NA>  <NA>  <NA>  <NA>  <NA>
#> 97    2.7   257  <NA>  <NA>  <NA>
#> 98   <NA>  <NA>  <NA>  <NA>  <NA>
#> 99   <NA>  <NA>  <NA>  <NA>  <NA>
#> 100  <NA>  09.0  <NA>  <NA>  <NA>
#> 101  <NA>  <NA>  <NA>  <NA>  <NA>
#> 102   140   5.0  <NA>  <NA>  <NA>
#> 103  <NA>  <NA>  <NA>  <NA>  <NA>
#> 104     1   080   2.0   1.3   230
#> 105     N   079  <NA>  <NA>  <NA>
#> 106     3   100   000   000   000
#> 107     7   100   000   000   000
#> 108     1   100   000   000   000
#> 109     3   100   000   000   000
#> 110     3   100   000   000   000
#> 111     4   100   000   000   000
#> 112  <NA>  <NA>  <NA>  <NA>  <NA>
#> 113 1 is  a mat ch wi th th e sig
#> 114  <NA>  <NA>  <NA>  <NA>  <NA>
#> 115   1.3   252  <NA>  <NA>  <NA>
#> 116  <NA>  <NA>  <NA>  <NA>  <NA>
#> 117  <NA>  <NA>  <NA>  <NA>  <NA>
#> 118  <NA>  <NA>  <NA>  <NA>  <NA>
#> 119   0.6   125  <NA>  <NA>  <NA>
#> 120  <NA>  <NA>  <NA>  <NA>  <NA>
#> 121   0.4   125  <NA>  <NA>  <NA>
#> 122  <NA>  <NA>  <NA>  <NA>  <NA>
#> 123  <NA>  <NA>  <NA>  <NA>  <NA>
#> 124   0.3   125  <NA>  <NA>  <NA>
#> 125  <NA>  <NA>  <NA>  <NA>  <NA>
#> 126   0.5  <NA>  <NA>  <NA>  <NA>
#> 127  <NA>  <NA>  <NA>  <NA>  <NA>
#> 128  <NA>  <NA>  <NA>  <NA>  <NA>
#> 129  <NA>  <NA>  <NA>  <NA>  <NA>
#> 130   1.1   175  <NA>  <NA>  <NA>
#> 131  <NA>  <NA>  <NA>  <NA>  <NA>
#> 132  <NA>  <NA>  <NA>  <NA>  <NA>
#> 133   0.5   105  <NA>  <NA>  <NA>
#> 134  <NA>  <NA>  <NA>  <NA>  <NA>
#> 135  <NA>  <NA>  <NA>  <NA>  <NA>
#> 136 n fly ing b ridge .  be en tr
#> 137     Y  <NA>  <NA>  <NA>  <NA>
#> 138  <NA>  <NA>  <NA>  <NA>  <NA>
#> 139  <NA>  <NA>  <NA>  <NA>  <NA>
#> 140  <NA>  10.0  <NA>  <NA>  <NA>
#> 141  <NA>  <NA>  <NA>  <NA>  <NA>
#> 142   138   5.0  <NA>  <NA>  <NA>
#> 143  <NA>  <NA>  <NA>  <NA>  <NA>
#> 144  <NA>  10.0  <NA>  <NA>  <NA>
#> 145  <NA>  <NA>  <NA>  <NA>  <NA>
#> 146     1   070   2.2   1.2   090
#> 147     Y   096  <NA>  <NA>  <NA>
#> 148     2   100   000   000   000
#> 149     5   100   000   000   000
#> 150     3   100   000   000   000
#> 151     7   100   000   000   000
#> 152     1   100   000   000   000
#> 153     2   100   000   000   000
#> 154     4   100   000   000   000
#> 155  <NA>  <NA>  <NA>  <NA>  <NA>
#>                                               Data9 EventNum       file_das
#> 1                                              <NA>        1 das_sample.das
#> 2                                              <NA>        2 das_sample.das
#> 3                                              <NA>        3 das_sample.das
#> 4                                              <NA>        4 das_sample.das
#> 5                                              <NA>        5 das_sample.das
#> 6                                              <NA>        6 das_sample.das
#> 7                                              <NA>        7 das_sample.das
#> 8                                              <NA>        8 das_sample.das
#> 9                                              <NA>        9 das_sample.das
#> 10                                             <NA>       10 das_sample.das
#> 11                                             <NA>       11 das_sample.das
#> 12                                             <NA>       12 das_sample.das
#> 13                                             <NA>       13 das_sample.das
#> 14                                              4.0       14 das_sample.das
#> 15                                             <NA>       15 das_sample.das
#> 16                                             <NA>       NA das_sample.das
#> 17                                             <NA>       NA das_sample.das
#> 18                                             <NA>       16 das_sample.das
#> 19                                             <NA>       17 das_sample.das
#> 20                                             <NA>       18 das_sample.das
#> 21                                             <NA>       19 das_sample.das
#> 22                                             <NA>       20 das_sample.das
#> 23                                             <NA>       21 das_sample.das
#> 24                                             <NA>       22 das_sample.das
#> 25                                             <NA>       23 das_sample.das
#> 26                                             <NA>       24 das_sample.das
#> 27                                             <NA>       25 das_sample.das
#> 28                                             <NA>       26 das_sample.das
#> 29                                             <NA>       27 das_sample.das
#> 30                                             <NA>       28 das_sample.das
#> 31                                              2.0       29 das_sample.das
#> 32                                             <NA>       30 das_sample.das
#> 33                                             <NA>       NA das_sample.das
#> 34                                             <NA>       31 das_sample.das
#> 35                                             <NA>       32 das_sample.das
#> 36                                             <NA>       33 das_sample.das
#> 37                                             <NA>       34 das_sample.das
#> 38                                             <NA>       35 das_sample.das
#> 39                                             <NA>       36 das_sample.das
#> 40                                             <NA>       37 das_sample.das
#> 41                                             <NA>       38 das_sample.das
#> 42                                             <NA>       39 das_sample.das
#> 43                                             <NA>       40 das_sample.das
#> 44                                             <NA>       41 das_sample.das
#> 45                                             <NA>       42 das_sample.das
#> 46                                                N       43 das_sample.das
#> 47                                             <NA>       44 das_sample.das
#> 48                                             <NA>       45 das_sample.das
#> 49                                             <NA>       46 das_sample.das
#> 50                                             <NA>       47 das_sample.das
#> 51                                             <NA>       48 das_sample.das
#> 52                                             <NA>       49 das_sample.das
#> 53                                             <NA>       50 das_sample.das
#> 54                                             <NA>       51 das_sample.das
#> 55                                             <NA>       52 das_sample.das
#> 56                                             <NA>       53 das_sample.das
#> 57                                             <NA>       54 das_sample.das
#> 58                                             <NA>       55 das_sample.das
#> 59                                             <NA>       56 das_sample.das
#> 60                                             <NA>       57 das_sample.das
#> 61                                             <NA>       58 das_sample.das
#> 62                                             <NA>       59 das_sample.das
#> 63                                             <NA>       60 das_sample.das
#> 64                                             <NA>       61 das_sample.das
#> 65                                             <NA>       62 das_sample.das
#> 66                                             <NA>       63 das_sample.das
#> 67                                             <NA>       64 das_sample.das
#> 68                                             <NA>       65 das_sample.das
#> 69                                             <NA>       66 das_sample.das
#> 70                                             <NA>       67 das_sample.das
#> 71                                             <NA>       68 das_sample.das
#> 72                                             <NA>       69 das_sample.das
#> 73                                             <NA>       70 das_sample.das
#> 74                                             <NA>       71 das_sample.das
#> 75                                             <NA>       72 das_sample.das
#> 76                                             <NA>       73 das_sample.das
#> 77                                             <NA>       74 das_sample.das
#> 78                                             <NA>       75 das_sample.das
#> 79                                             <NA>       76 das_sample.das
#> 80                                             <NA>       77 das_sample.das
#> 81                                             <NA>       78 das_sample.das
#> 82                                             <NA>       79 das_sample.das
#> 83                                             <NA>       80 das_sample.das
#> 84                                             <NA>       81 das_sample.das
#> 85                                             <NA>       82 das_sample.das
#> 86                                             <NA>       83 das_sample.das
#> 87                                             <NA>       84 das_sample.das
#> 88                                             <NA>       NA das_sample.das
#> 89                                             <NA>       85 das_sample.das
#> 90                                             <NA>       86 das_sample.das
#> 91                                             <NA>       87 das_sample.das
#> 92                                             <NA>       88 das_sample.das
#> 93                                             <NA>       89 das_sample.das
#> 94                                             <NA>       90 das_sample.das
#> 95                                             <NA>       91 das_sample.das
#> 96                                             <NA>       92 das_sample.das
#> 97                                             <NA>       93 das_sample.das
#> 98                                             <NA>       94 das_sample.das
#> 99                                             <NA>       95 das_sample.das
#> 100                                            <NA>       96 das_sample.das
#> 101                                            <NA>       97 das_sample.das
#> 102                                            <NA>       98 das_sample.das
#> 103                                            <NA>       99 das_sample.das
#> 104                                             2.0      100 das_sample.das
#> 105                                            <NA>      101 das_sample.das
#> 106                                            <NA>       NA das_sample.das
#> 107                                            <NA>       NA das_sample.das
#> 108                                            <NA>       NA das_sample.das
#> 109                                            <NA>       NA das_sample.das
#> 110                                            <NA>       NA das_sample.das
#> 111                                            <NA>       NA das_sample.das
#> 112                                            <NA>      102 das_sample.das
#> 113 hting 1014                                           103 das_sample.das
#> 114                                            <NA>      104 das_sample.das
#> 115                                            <NA>      105 das_sample.das
#> 116                                            <NA>      106 das_sample.das
#> 117                                            <NA>      107 das_sample.das
#> 118                                            <NA>      108 das_sample.das
#> 119                                            <NA>      109 das_sample.das
#> 120                                            <NA>      110 das_sample.das
#> 121                                            <NA>      111 das_sample.das
#> 122                                            <NA>      112 das_sample.das
#> 123                                            <NA>      113 das_sample.das
#> 124                                            <NA>      114 das_sample.das
#> 125                                            <NA>      115 das_sample.das
#> 126                                            <NA>      116 das_sample.das
#> 127                                            <NA>      117 das_sample.das
#> 128                                            <NA>      118 das_sample.das
#> 129                                            <NA>      119 das_sample.das
#> 130                                            <NA>      120 das_sample.das
#> 131                                            <NA>      121 das_sample.das
#> 132                                            <NA>      122 das_sample.das
#> 133                                            <NA>      123 das_sample.das
#> 134                                            <NA>      124 das_sample.das
#> 135                                            <NA>      125 das_sample.das
#> 136 acking sighting 081 and prey patches until dark      126 das_sample.das
#> 137                                            <NA>        1 das_sample.das
#> 138                                            <NA>        2 das_sample.das
#> 139                                            <NA>        3 das_sample.das
#> 140                                            <NA>        4 das_sample.das
#> 141                                            <NA>        5 das_sample.das
#> 142                                            <NA>        6 das_sample.das
#> 143                                            <NA>        7 das_sample.das
#> 144                                            <NA>        8 das_sample.das
#> 145                                            <NA>        9 das_sample.das
#> 146                                             4.0       10 das_sample.das
#> 147                                            <NA>       11 das_sample.das
#> 148                                            <NA>       NA das_sample.das
#> 149                                            <NA>       NA das_sample.das
#> 150                                            <NA>       NA das_sample.das
#> 151                                            <NA>       NA das_sample.das
#> 152                                            <NA>       NA das_sample.das
#> 153                                            <NA>       NA das_sample.das
#> 154                                            <NA>       NA das_sample.das
#> 155                                            <NA>       12 das_sample.das
#>     line_num
#> 1          1
#> 2          2
#> 3          3
#> 4          4
#> 5          5
#> 6          6
#> 7          7
#> 8          8
#> 9          9
#> 10        10
#> 11        11
#> 12        12
#> 13        13
#> 14        14
#> 15        15
#> 16        16
#> 17        17
#> 18        18
#> 19        19
#> 20        20
#> 21        21
#> 22        22
#> 23        23
#> 24        24
#> 25        25
#> 26        26
#> 27        27
#> 28        28
#> 29        29
#> 30        30
#> 31        31
#> 32        32
#> 33        33
#> 34        34
#> 35        35
#> 36        36
#> 37        37
#> 38        38
#> 39        39
#> 40        40
#> 41        41
#> 42        42
#> 43        43
#> 44        44
#> 45        45
#> 46        46
#> 47        47
#> 48        48
#> 49        49
#> 50        50
#> 51        51
#> 52        52
#> 53        53
#> 54        54
#> 55        55
#> 56        56
#> 57        57
#> 58        58
#> 59        59
#> 60        60
#> 61        61
#> 62        62
#> 63        63
#> 64        64
#> 65        65
#> 66        66
#> 67        67
#> 68        68
#> 69        69
#> 70        70
#> 71        71
#> 72        72
#> 73        73
#> 74        74
#> 75        75
#> 76        76
#> 77        77
#> 78        78
#> 79        79
#> 80        80
#> 81        81
#> 82        82
#> 83        83
#> 84        84
#> 85        85
#> 86        86
#> 87        87
#> 88        88
#> 89        89
#> 90        90
#> 91        91
#> 92        92
#> 93        93
#> 94        94
#> 95        95
#> 96        96
#> 97        97
#> 98        98
#> 99        99
#> 100      100
#> 101      101
#> 102      102
#> 103      103
#> 104      104
#> 105      105
#> 106      106
#> 107      107
#> 108      108
#> 109      109
#> 110      110
#> 111      111
#> 112      112
#> 113      113
#> 114      114
#> 115      115
#> 116      116
#> 117      117
#> 118      118
#> 119      119
#> 120      120
#> 121      121
#> 122      122
#> 123      123
#> 124      124
#> 125      125
#> 126      126
#> 127      127
#> 128      128
#> 129      129
#> 130      130
#> 131      131
#> 132      132
#> 133      133
#> 134      134
#> 135      135
#> 136      136
#> 137      137
#> 138      138
#> 139      139
#> 140      140
#> 141      141
#> 142      142
#> 143      143
#> 144      144
#> 145      145
#> 146      146
#> 147      147
#> 148      148
#> 149      149
#> 150      150
#> 151      151
#> 152      152
#> 153      153
#> 154      154
#> 155      155