Read DAS file(s)

Read one or more fixed-width DAS text file(s) generated by WinCruz into a data frame, where each line is data for a specific event

das_read(file, skip = 0, tz = "UTC", ...)

Arguments

file
filename(s) of one or more DAS files

skip
integer; see read_fwf. Default is 0

tz
character; see strptime. Default is "UTC"

...
ignored

Value

A das_dfr object, which is also a data frame, with DAS data read into columns. The data are read into the data frame as characters as described in 'Details', with the following exceptions:

Name	Class	Details

EffortDot	logical	TRUE if "." was present, and FALSE otherwise

DateTime	POSIXct	combination of 'Date' and 'Time' columns with time zone tz

Lat	numeric	'Latitude' column converted to decimal degrees in range [-90, 90]

Lon	numeric	'Longitude' column converted to decimal degrees in range [-180, 180]

Data#	character	leading/trailing whitespace trimmed for non-comment events (i.e. where 'Event' is not "C")

EventNum	character	leading/trailing whitespace trimmed; left as character for some project-specific codes

file_das	character	base filename, extracted from the file argument

line_num	integer	line number of each data row

Warnings are printed if any unexpected events have NA DateTime/Lat/Lon values, or if any Lat/Lon values cannot be converted to numeric values. Events that are 'expected' to have NA DateTime/Lat/Lon values are: C, ?, 1, 2, 3, 4, 5, 6, 7, 8

Details

Reads/parses DAS data into columns of a data frame. If file contains multiple filenames, then the individual data frames will be concatenated.

The provided DAS file must adhere to the following column number and format specifications:

Item	Columns	Format

Event number	1-3	Event

4	Effort dot	5

Time	6-11	HHMMSS or HHMM

Date	13-18	MMDDYY

Latitude	20-28	NDD:MM.MM

Longitude	30-39	WDDD:MM.MM

Data1	40-44	Data2

45-49	Data3	50-54

Data4	55-59	Data5

60-64	Data6	65-69

Data7	70-74	Data8

75-79	Data9	80-84

Data10	85-89	Data11

90-94	Data12	95+

See das_format_pdf for more information about DAS format requirements, and note that 'Data#' columns may be referred to as 'Field#' columns in other documentation This function was inspired by das.read

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
das_read(y)

#>     Event EffortDot            DateTime      Lat       Lon Data1 Data2 Data3
#> 1       B      TRUE 2013-01-13 06:27:39 39.32033 -137.6043  1000     c     5
#> 2       R      TRUE 2013-01-13 06:27:39 39.32033 -137.6043     S  <NA>  <NA>
#> 3       P      TRUE 2013-01-13 06:27:39 39.32033 -137.6043   280   001   126
#> 4       V      TRUE 2013-01-13 06:27:39 39.32033 -137.6043     3    03   230
#> 5       N      TRUE 2013-01-13 06:27:39 39.32033 -137.6043   023  09.8  <NA>
#> 6       W      TRUE 2013-01-13 06:27:39 39.32033 -137.6043     1  <NA>  <NA>
#> 7       V      TRUE 2013-01-13 06:29:56 39.32583 -137.6018     3    03   230
#> 8       W      TRUE 2013-01-13 06:30:10 39.32650 -137.6015     1    02    03
#> 9       W      TRUE 2013-01-13 06:34:01 39.33600 -137.5970     1  <NA>  <NA>
#> 10      *      TRUE 2013-01-13 06:37:25 39.34450 -137.5927  <NA>  <NA>  <NA>
#> 11      P      TRUE 2013-01-13 06:41:08 39.35400 -137.5880   208   280   001
#> 12      V      TRUE 2013-01-13 06:41:08 39.35400 -137.5880     3    03   230
#> 13      N      TRUE 2013-01-13 06:41:08 39.35400 -137.5880   025  10.2  <NA>
#> 14      W      TRUE 2013-01-13 06:41:08 39.35400 -137.5880     1  <NA>  <NA>
#> 15      S      TRUE 2013-01-13 06:46:02 39.36617 -137.5820  1406   208     3
#> 16      A      TRUE 2013-01-13 06:46:02 39.36617 -137.5820  1406  <NA>     N
#> 17      1     FALSE                <NA>       NA        NA   280  <NA>  <NA>
#> 18      2     FALSE                <NA>       NA        NA   001  <NA>  <NA>
#> 19      3     FALSE                <NA>       NA        NA   208  <NA>  <NA>
#> 20      E     FALSE 2013-01-13 06:46:25 39.36717 -137.5817     C  <NA>  <NA>
#> 21      *     FALSE 2013-01-13 06:47:25 39.36967 -137.5807  <NA>  <NA>  <NA>
#> 22      *     FALSE 2013-01-13 06:57:25 39.37467 -137.5987  <NA>  <NA>  <NA>
#> 23      R      TRUE 2013-01-13 06:58:04 39.37617 -137.5978     S  <NA>  <NA>
#> 24      P      TRUE 2013-01-13 06:58:04 39.37617 -137.5978   208   280   001
#> 25      V      TRUE 2013-01-13 06:58:04 39.37617 -137.5978     3    03   230
#> 26      N      TRUE 2013-01-13 06:58:04 39.37617 -137.5978   029  09.1  <NA>
#> 27      W      TRUE 2013-01-13 06:58:04 39.37617 -137.5978     1  <NA>  <NA>
#> 28      *      TRUE 2013-01-13 07:07:25 39.39883 -137.5868  <NA>  <NA>  <NA>
#> 29      *      TRUE 2013-01-13 07:17:25 39.42317 -137.5747  <NA>  <NA>  <NA>
#> 30      P      TRUE 2013-01-13 07:20:02 39.42950 -137.5715   125   208   280
#> 31      V      TRUE 2013-01-13 07:20:02 39.42950 -137.5715     3    03   230
#> 32      N      TRUE 2013-01-13 07:20:02 39.42950 -137.5715   026  09.7  <NA>
#> 33      W      TRUE 2013-01-13 07:20:02 39.42950 -137.5715     3  <NA>  <NA>
#> 34      *      TRUE 2013-01-13 07:27:25 39.44733 -137.5627  <NA>  <NA>  <NA>
#> 35      *      TRUE 2013-01-13 07:37:25 39.47133 -137.5507  <NA>  <NA>  <NA>
#> 36      W      TRUE 2013-01-13 07:38:06 39.47300 -137.5500     3    02    02
#> 37      *      TRUE 2013-01-13 07:47:25 39.49567 -137.5390  <NA>  <NA>  <NA>
#> 38      S      TRUE 2013-01-13 07:56:22 39.51767 -137.5285  1407   125     3
#> 39      A      TRUE 2013-01-13 07:56:22 39.51767 -137.5285  1407  <NA>     Y
#> 40      1     FALSE                <NA>       NA        NA   280     6    10
#> 41      2     FALSE                <NA>       NA        NA   001     9    10
#> 42      3     FALSE                <NA>       NA        NA   125     9    22
#> 43      E     FALSE 2013-01-13 07:57:05 39.51933 -137.5277     C  <NA>  <NA>
#> 44      *     FALSE 2013-01-13 07:57:25 39.52017 -137.5272  <NA>  <NA>  <NA>
#> 45      s     FALSE 2013-01-13 08:06:00 39.54217 -137.5263  1407   011   2.0
#> 46      *     FALSE 2013-01-13 08:07:25 39.54583 -137.5262  <NA>  <NA>  <NA>
#> 47      s     FALSE 2013-01-13 08:08:58 39.55000 -137.5255  1407   005   3.5
#> 48      s     FALSE 2013-01-13 08:15:32 39.56000 -137.5218  1407   050  <NA>
#> 49      *     FALSE 2013-01-13 08:17:25 39.56233 -137.5210  <NA>  <NA>  <NA>
#> 50      s     FALSE 2013-01-13 08:17:28 39.56233 -137.5210  1407   071   4.5
#> 51      s     FALSE 2013-01-13 08:19:39 39.56483 -137.5197  1407   104   4.5
#> 52      s     FALSE 2013-01-13 08:26:45 39.56400 -137.5140  1407   002   2.2
#> 53      *     FALSE 2013-01-13 08:27:25 39.56333 -137.5133  <NA>  <NA>  <NA>
#> 54      *     FALSE 2013-01-13 08:37:25 39.54867 -137.5007  <NA>  <NA>  <NA>
#> 55      *     FALSE 2013-01-13 08:47:25 39.54533 -137.4807  <NA>  <NA>  <NA>
#> 56      *     FALSE 2013-01-13 08:57:25 39.54433 -137.4717  <NA>  <NA>  <NA>
#> 57      *     FALSE 2013-01-13 09:07:25 39.54333 -137.4627  <NA>  <NA>  <NA>
#> 58      *     FALSE 2013-01-13 09:17:25 39.55833 -137.4573  <NA>  <NA>  <NA>
#> 59      R      TRUE 2013-01-13 09:22:13 39.56800 -137.4530     S  <NA>  <NA>
#> 60      P      TRUE 2013-01-13 09:22:13 39.56800 -137.4530   001   126   149
#> 61      V      TRUE 2013-01-13 09:22:13 39.56800 -137.4530     3    03   230
#> 62      N      TRUE 2013-01-13 09:22:13 39.56800 -137.4530   027  09.0  <NA>
#> 63      W      TRUE 2013-01-13 09:22:13 39.56800 -137.4530     1    02    02
#> 64      *      TRUE 2013-01-13 09:27:25 39.57983 -137.4475  <NA>  <NA>  <NA>
#> 65      t      TRUE 2013-01-13 09:34:27 39.59733 -137.4400   280    LV   120
#> 66      *      TRUE 2013-01-13 09:37:25 39.60467 -137.4368  <NA>  <NA>  <NA>
#> 67      *      TRUE 2013-01-13 09:47:25 39.62983 -137.4262  <NA>  <NA>  <NA>
#> 68      *      TRUE 2013-01-13 09:57:25 39.65517 -137.4155  <NA>  <NA>  <NA>
#> 69      W      TRUE 2013-01-13 09:59:38 39.66083 -137.4132     1    02    02
#> 70      V      TRUE 2013-01-13 09:59:50 39.66133 -137.4130     2    03   230
#> 71      C      TRUE 2013-01-13 10:04:23 39.67300 -137.4083  Well ?      <NA>
#> 72      P      TRUE 2013-01-13 10:04:35 39.67350 -137.4080   280   001   126
#> 73      V      TRUE 2013-01-13 10:04:35 39.67350 -137.4080     2    03   230
#> 74      N      TRUE 2013-01-13 10:04:35 39.67350 -137.4080   023  10.0  <NA>
#> 75      W      TRUE 2013-01-13 10:04:35 39.67350 -137.4080     3  <NA>  <NA>
#> 76      *      TRUE 2013-01-13 10:07:25 39.68083 -137.4050  <NA>  <NA>  <NA>
#> 77      V      TRUE 2013-01-13 10:11:00 39.69000 -137.4012     2    03   230
#> 78      W      TRUE 2013-01-13 10:11:09 39.69050 -137.4010     3  <NA>  <NA>
#> 79      *      TRUE 2013-01-13 10:17:25 39.70650 -137.3943  <NA>  <NA>  <NA>
#> 80      N      TRUE 2013-01-13 10:20:38 39.71483 -137.3920   352  09.3  <NA>
#> 81      C      TRUE 2013-01-13 10:20:43 39.71483 -137.3920  goin g lef t to 
#> 82      #      TRUE 2013-01-13 10:21:00 39.71583 -137.3918  <NA>  <NA>  <NA>
#> 83      *      TRUE 2013-01-13 10:27:25 39.73150 -137.3993  <NA>  <NA>  <NA>
#> 84      W      TRUE 2013-01-13 10:30:28 39.73917 -137.4032     3  <NA>  <NA>
#> 85      W      TRUE 2013-01-13 10:35:14 39.75117 -137.4092     3  <NA>  <NA>
#> 86      P      TRUE 2013-01-13 10:36:06 39.75350 -137.4103   280   001   126
#> 87      V      TRUE 2013-01-13 10:36:06 39.75350 -137.4103     2    03   230
#> 88      N      TRUE 2013-01-13 10:36:06 39.75350 -137.4103   335  10.1  <NA>
#> 89      W      TRUE 2013-01-13 10:36:06 39.75350 -137.4103     3  <NA>  <NA>
#> 90      E     FALSE 2013-01-13 10:36:27 39.75433 -137.4107     W  <NA>  <NA>
#> 91      *     FALSE 2013-01-13 10:37:25 39.75683 -137.4118  <NA>  <NA>  <NA>
#> 92      *     FALSE 2013-01-13 10:47:25 39.78450 -137.4093  <NA>  <NA>  <NA>
#> 93      *     FALSE 2013-01-13 10:57:25 39.81250 -137.4087  <NA>  <NA>  <NA>
#> 94      *     FALSE 2013-01-13 11:07:25 39.83983 -137.4100  <NA>  <NA>  <NA>
#> 95      *     FALSE 2013-01-13 11:17:25 39.86700 -137.4110  <NA>  <NA>  <NA>
#> 96      *     FALSE 2013-01-13 11:27:25 39.89000 -137.4000  <NA>  <NA>  <NA>
#> 97      *     FALSE 2013-01-13 11:37:25 39.91233 -137.3875  <NA>  <NA>  <NA>
#> 98      *     FALSE 2013-01-13 11:47:25 39.93500 -137.3750  <NA>  <NA>  <NA>
#> 99      R      TRUE 2013-01-13 11:51:51 39.94517 -137.3692     S  <NA>  <NA>
#> 100     P      TRUE 2013-01-13 11:51:51 39.94517 -137.3692   125   208   280
#> 101     V      TRUE 2013-01-13 11:51:51 39.94517 -137.3692     3    03   230
#> 102     N      TRUE 2013-01-13 11:51:51 39.94517 -137.3692   032  09.5  <NA>
#> 103     W      TRUE 2013-01-13 11:51:51 39.94517 -137.3692     3  <NA>  <NA>
#> 104     *      TRUE 2013-01-13 11:57:25 39.95767 -137.3613  <NA>  <NA>  <NA>
#> 105     P      TRUE 2013-01-13 12:02:29 39.96900 -137.3542   149   125   208
#> 106     V      TRUE 2013-01-13 12:02:29 39.96900 -137.3542     3    03   230
#> 107     N      TRUE 2013-01-13 12:02:29 39.96900 -137.3542   035  09.6  <NA>
#> 108     W      TRUE 2013-01-13 12:02:29 39.96900 -137.3542     1  <NA>  <NA>
#> 109     *      TRUE 2013-01-13 12:07:25 39.97983 -137.3472  <NA>  <NA>  <NA>
#> 110     W      TRUE 2013-01-13 12:11:06 39.98767 -137.3418     1    12    12
#> 111     *      TRUE 2013-01-13 12:17:25 40.00117 -137.3332  <NA>  <NA>  <NA>
#> 112     *      TRUE 2013-01-13 12:27:25 40.02183 -137.3197  <NA>  <NA>  <NA>
#> 113     *      TRUE 2013-01-13 12:37:25 40.04317 -137.3060  <NA>  <NA>  <NA>
#> 114     P      TRUE 2013-01-13 12:43:14 40.05567 -137.2978   126   149   125
#> 115     V      TRUE 2013-01-13 12:43:14 40.05567 -137.2978     3    03   230
#> 116     N      TRUE 2013-01-13 12:43:14 40.05567 -137.2978   035  09.2  <NA>
#> 117     W      TRUE 2013-01-13 12:43:14 40.05567 -137.2978     1    12    12
#> 118     *      TRUE 2013-01-13 12:47:25 40.06450 -137.2920  <NA>  <NA>  <NA>
#> 119     *      TRUE 2013-01-13 12:57:25 40.08583 -137.2777  <NA>  <NA>  <NA>
#> 120     *      TRUE 2013-01-13 13:07:25 40.10750 -137.2627  <NA>  <NA>  <NA>
#> 121     E     FALSE 2013-01-13 13:16:38 40.12750 -137.2487     U  <NA>  <NA>
#> 122     *     FALSE 2013-01-13 13:17:25 40.12917 -137.2477  <NA>  <NA>  <NA>
#> 123     *     FALSE 2013-01-13 13:27:25 40.13817 -137.2248  <NA>  <NA>  <NA>
#> 124     t     FALSE 2013-01-13 13:35:18 40.14000 -137.2048   149    DC   270
#> 125     *     FALSE 2013-01-13 13:37:25 40.14100 -137.1993  <NA>  <NA>  <NA>
#> 126     *     FALSE 2013-01-13 13:47:25 40.14717 -137.1782  <NA>  <NA>  <NA>
#> 127     R      TRUE 2013-01-13 13:50:07 40.15217 -137.1737     S  <NA>  <NA>
#> 128     P      TRUE 2013-01-13 13:50:07 40.15217 -137.1737   001   126   149
#> 129     V      TRUE 2013-01-13 13:50:07 40.15217 -137.1737     3    03   230
#> 130     N      TRUE 2013-01-13 13:50:07 40.15217 -137.1737   022  09.5  <NA>
#> 131     W      TRUE 2013-01-13 13:50:07 40.15217 -137.1737     1    08    01
#> 132     *      TRUE 2013-01-13 13:57:25 40.16967 -137.1670  <NA>  <NA>  <NA>
#> 133     P      TRUE 2013-01-13 14:00:31 40.17700 -137.1642   280   001   126
#> 134     V      TRUE 2013-01-13 14:00:31 40.17700 -137.1642     3    03   230
#> 135     N      TRUE 2013-01-13 14:00:31 40.17700 -137.1642   020  09.3  <NA>
#> 136     W      TRUE 2013-01-13 14:00:31 40.17700 -137.1642     1    08    01
#> 137     t      TRUE 2013-01-13 14:02:55 40.18283 -137.1622   228    DC   300
#> 138     *      TRUE 2013-01-13 14:07:25 40.19333 -137.1583  <NA>  <NA>  <NA>
#> 139     *      TRUE 2013-01-13 14:17:25 40.21667 -137.1505  <NA>  <NA>  <NA>
#> 140     *      TRUE 2013-01-13 14:27:25 40.24017 -137.1432  <NA>  <NA>  <NA>
#> 141     *      TRUE 2013-01-13 14:37:25 40.26417 -137.1353  <NA>  <NA>  <NA>
#> 142     S      TRUE 2013-01-13 14:37:56 40.26567 -137.1350  1408   280     3
#> 143     A      TRUE 2013-01-13 14:37:56 40.26567 -137.1350  1408  <NA>     N
#> 144     1     FALSE                <NA>       NA        NA   280    11    24
#> 145     2     FALSE                <NA>       NA        NA   001    12    23
#> 146     3     FALSE                <NA>       NA        NA   126     9    13
#> 147     E     FALSE 2013-01-13 14:38:13 40.26617 -137.1348     C  <NA>  <NA>
#> 148     *     FALSE 2013-01-13 14:47:25 40.26317 -137.1353  <NA>  <NA>  <NA>
#> 149     *     FALSE 2013-01-13 14:57:25 40.26383 -137.1282  <NA>  <NA>  <NA>
#> 150     R      TRUE 2013-01-13 14:59:19 40.26867 -137.1268     S  <NA>  <NA>
#> 151     P      TRUE 2013-01-13 14:59:19 40.26867 -137.1268   208   280   001
#> 152     V      TRUE 2013-01-13 14:59:19 40.26867 -137.1268     3    03   230
#> 153     N      TRUE 2013-01-13 14:59:19 40.26867 -137.1268   017  09.3  <NA>
#> 154     W      TRUE 2013-01-13 14:59:19 40.26867 -137.1268     1    08    01
#> 155     *      TRUE 2013-01-13 15:07:25 40.28850 -137.1205  <NA>  <NA>  <NA>
#> 156     *      TRUE 2013-01-13 15:17:25 40.31300 -137.1132  <NA>  <NA>  <NA>
#> 157     P      TRUE 2013-01-13 15:20:26 40.32033 -137.1108   125   208   280
#> 158     V      TRUE 2013-01-13 15:20:26 40.32033 -137.1108     2    03   230
#> 159     N      TRUE 2013-01-13 15:20:26 40.32033 -137.1108   016  08.9  <NA>
#> 160     W      TRUE 2013-01-13 15:20:26 40.32033 -137.1108     1    09    01
#> 161     *      TRUE 2013-01-13 15:27:25 40.33700 -137.1057  <NA>  <NA>  <NA>
#> 162     t      TRUE 2013-01-13 15:36:47 40.36050 -137.0978   231    DC   045
#> 163     *      TRUE 2013-01-13 15:37:25 40.36217 -137.0973  <NA>  <NA>  <NA>
#> 164     E     FALSE 2013-01-13 15:43:08 40.37600 -137.0915     X  <NA>  <NA>
#> 165     *     FALSE 2013-01-13 15:47:25 40.37933 -137.0927  <NA>  <NA>  <NA>
#> 166     *     FALSE 2013-01-13 15:57:25 40.37950 -137.0990  <NA>  <NA>  <NA>
#> 167     R      TRUE 2013-01-13 15:58:41 40.38250 -137.0977     S  <NA>  <NA>
#> 168     P      TRUE 2013-01-13 15:58:41 40.38250 -137.0977   149   125   208
#> 169     V      TRUE 2013-01-13 15:58:41 40.38250 -137.0977     3    03   230
#> 170     N      TRUE 2013-01-13 15:58:41 40.38250 -137.0977   025  08.9  <NA>
#> 171     W      TRUE 2013-01-13 15:58:41 40.38250 -137.0977     1    08    02
#> 172     *      TRUE 2013-01-13 16:07:25 40.40200 -137.0885  <NA>  <NA>  <NA>
#> 173     *      TRUE 2013-01-13 16:17:25 40.42383 -137.0777  <NA>  <NA>  <NA>
#> 174     V      TRUE 2013-01-13 16:20:02 40.42967 -137.0745     2    03   230
#> 175     *      TRUE 2013-01-13 16:27:25 40.44600 -137.0658  <NA>  <NA>  <NA>
#> 176     S      TRUE 2013-01-13 16:29:50 40.45133 -137.0628  1409   149     3
#> 177     A      TRUE 2013-01-13 16:29:50 40.45133 -137.0628  1409  <NA>     Y
#> 178     1     FALSE                <NA>       NA        NA   125    46    90
#> 179     2     FALSE                <NA>       NA        NA   149    28    65
#> 180     3     FALSE                <NA>       NA        NA   208    66    82
#> 181     E     FALSE 2013-01-13 16:29:50 40.45133 -137.0628     C  <NA>  <NA>
#> 182     s     FALSE 2013-01-13 16:36:34 40.46667 -137.0570  1409   356   0.4
#> 183     *     FALSE 2013-01-13 16:37:25 40.46867 -137.0567  <NA>  <NA>  <NA>
#> 184     *     FALSE 2013-01-13 16:47:25 40.49250 -137.0563  <NA>  <NA>  <NA>
#> 185     #     FALSE 2013-01-13 16:55:21 40.51217 -137.0595  <NA>  <NA>  <NA>
#> 186     C     FALSE 2013-01-13 16:55:27 40.51250 -137.0593  off  effor t aft
#> 187     *     FALSE 2013-01-13 16:57:25 40.51650 -137.0567  <NA>  <NA>  <NA>
#> 188     R      TRUE 2013-01-13 16:59:54 40.52200 -137.0533     S  <NA>  <NA>
#> 189     P      TRUE 2013-01-13 16:59:54 40.52200 -137.0533   126   149   125
#> 190     V      TRUE 2013-01-13 16:59:54 40.52200 -137.0533     2    03   230
#> 191     N      TRUE 2013-01-13 16:59:54 40.52200 -137.0533   030  09.5  <NA>
#> 192     W      TRUE 2013-01-13 16:59:54 40.52200 -137.0533     1    08    02
#> 193     S      TRUE 2013-01-13 17:00:45 40.52400 -137.0522  1410   125     3
#> 194     A      TRUE 2013-01-13 17:00:45 40.52400 -137.0522  1410  <NA>     Y
#> 195     1     FALSE                <NA>       NA        NA   280    37    72
#> 196     2     FALSE                <NA>       NA        NA   125    35    74
#> 197     3     FALSE                <NA>       NA        NA   149    29    52
#> 198     4     FALSE                <NA>       NA        NA   126    66    93
#> 199     E     FALSE 2013-01-13 17:01:21 40.52533 -137.0515     C  <NA>  <NA>
#> 200     #     FALSE 2013-01-13 17:03:01 40.52733 -137.0482  <NA>  <NA>  <NA>
#> 201     *     FALSE 2013-01-13 17:07:25 40.52167 -137.0392  <NA>  <NA>  <NA>
#> 202     t     FALSE 2013-01-13 17:08:11 40.52083 -137.0383   280    DC   042
#> 203     *     FALSE 2013-01-13 17:17:25 40.51983 -137.0377  <NA>  <NA>  <NA>
#> 204     *     FALSE 2013-01-13 17:27:25 40.51750 -137.0427  <NA>  <NA>  <NA>
#> 205     *     FALSE 2013-01-13 17:37:25 40.50750 -137.0418  <NA>  <NA>  <NA>
#> 206     t     FALSE 2013-01-13 17:45:10 40.50450 -137.0437   099    LV   180
#> 207     *     FALSE 2013-01-13 17:47:32 40.50383 -137.0443  <NA>  <NA>  <NA>
#> 208     *     FALSE 2013-01-13 17:57:32 40.50033 -137.0472  <NA>  <NA>  <NA>
#> 209     *     FALSE 2013-01-13 18:07:32 40.50183 -137.0530  <NA>  <NA>  <NA>
#> 210     *     FALSE 2013-01-13 18:17:32 40.50567 -137.0432  <NA>  <NA>  <NA>
#> 211     *     FALSE 2013-01-14 08:45:06 40.70300 -135.8103  <NA>  <NA>  <NA>
#> 212     *     FALSE 2013-01-14 08:55:06 40.71917 -135.7988  <NA>  <NA>  <NA>
#> 213     *     FALSE 2013-01-14 09:05:06 40.73533 -135.7875  <NA>  <NA>  <NA>
#> 214     *     FALSE 2013-01-14 09:15:06 40.74917 -135.7770  <NA>  <NA>  <NA>
#> 215     S     FALSE 2013-01-14 09:17:21 40.75183 -135.7748  1411   280     3
#> 216     A     FALSE 2013-01-14 09:17:21 40.75183 -135.7748  1411  <NA>     N
#> 217     1     FALSE                <NA>       NA        NA   280  <NA>  <NA>
#> 218     C     FALSE 2013-01-14 09:18:04 40.75283 -135.7742  off  effor t, fi
#> 219     *     FALSE 2013-01-14 09:25:06 40.76233 -135.7655  <NA>  <NA>  <NA>
#> 220     *     FALSE 2013-01-14 09:35:06 40.76700 -135.7470  <NA>  <NA>  <NA>
#> 221     *     FALSE 2013-01-14 09:45:06 40.77450 -135.7313  <NA>  <NA>  <NA>
#> 222     *     FALSE 2013-01-14 09:55:06 40.78867 -135.7203  <NA>  <NA>  <NA>
#> 223     *     FALSE 2013-01-14 10:05:06 40.80750 -135.7092  <NA>  <NA>  <NA>
#> 224     *     FALSE 2013-01-14 10:15:06 40.82933 -135.6968  <NA>  <NA>  <NA>
#> 225     *     FALSE 2013-01-14 10:25:06 40.85200 -135.6843  <NA>  <NA>  <NA>
#> 226     *     FALSE 2013-01-14 10:35:06 40.87550 -135.6703  <NA>  <NA>  <NA>
#> 227     *     FALSE 2013-01-14 10:45:06 40.89883 -135.6565  <NA>  <NA>  <NA>
#> 228     *     FALSE 2013-01-14 10:55:06 40.92183 -135.6417  <NA>  <NA>  <NA>
#> 229     *     FALSE 2013-01-14 11:05:06 40.94400 -135.6273  <NA>  <NA>  <NA>
#> 230     *     FALSE 2013-01-14 11:15:06 40.96633 -135.6123  <NA>  <NA>  <NA>
#> 231     B      TRUE 2013-01-14 11:24:32 40.98717 -135.5980  1000     c     5
#> 232     R      TRUE 2013-01-14 11:24:32 40.98717 -135.5980     S  <NA>  <NA>
#> 233     P      TRUE 2013-01-14 11:24:32 40.98717 -135.5980   149   125   208
#> 234     V      TRUE 2013-01-14 11:24:32 40.98717 -135.5980     2    01   035
#> 235     N      TRUE 2013-01-14 11:24:32 40.98717 -135.5980   035  09.5  <NA>
#> 236     W      TRUE 2013-01-14 11:24:32 40.98717 -135.5980     3  <NA>  <NA>
#> 237     *      TRUE 2013-01-14 11:25:06 40.98833 -135.5972  <NA>  <NA>  <NA>
#> 238     F      TRUE 2013-01-14 11:25:32 40.98950 -135.5965   149   309  1.47
#> 239     *      TRUE 2013-01-14 11:35:06 41.01067 -135.5817  <NA>  <NA>  <NA>
#> 240     E     FALSE 2013-01-14 11:37:24 41.01583 -135.5782     S  <NA>  <NA>
#> 241     C     FALSE 2013-01-14 11:37:27 41.01600 -135.5780  cros sing  fishi
#> 242     R      TRUE 2013-01-14 11:40:38 41.02383 -135.5743     S  <NA>  <NA>
#> 243     P      TRUE 2013-01-14 11:40:38 41.02383 -135.5743   149   125   208
#> 244     V      TRUE 2013-01-14 11:40:38 41.02383 -135.5743     2    01   035
#> 245     N      TRUE 2013-01-14 11:40:38 41.02383 -135.5743   023  09.6  <NA>
#> 246     W      TRUE 2013-01-14 11:40:38 41.02383 -135.5743     3  <NA>  <NA>
#> 247     *      TRUE 2013-01-14 11:45:06 41.03400 -135.5678  <NA>  <NA>  <NA>
#> 248     S      TRUE 2013-01-14 11:47:51 41.04017 -135.5635  1412   149     2
#> 249     A      TRUE 2013-01-14 11:47:51 41.04017 -135.5635  1412  <NA>     Y
#> 250     1     FALSE                <NA>       NA        NA   149   183   328
#> 251     2     FALSE                <NA>       NA        NA   126   120   170
#> 252     S      TRUE 2013-01-14 11:49:14 41.04333 -135.5615  1413   208     3
#> 253     A      TRUE 2013-01-14 11:49:14 41.04333 -135.5615  1413  <NA>     Y
#> 254     ?     FALSE                <NA>       NA        NA  1413  <NA>  <NA>
#> 255     1     FALSE                <NA>       NA        NA   125    21    60
#> 256     2     FALSE                <NA>       NA        NA   208    16    20
#> 257     3     FALSE                <NA>       NA        NA   149    12    18
#> 258     4     FALSE                <NA>       NA        NA   126    36    53
#> 259     E     FALSE 2013-01-14 11:50:29 41.04600 -135.5595     C  <NA>  <NA>
#>     Data4 Data5 Data6 Data7 Data8      Data9 Data10 Data11 Data12 EventNum
#> 1       Y  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        1
#> 2    <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        2
#> 3    <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        3
#> 4    <NA>  10.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        4
#> 5    <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        5
#> 6     250   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        6
#> 7    <NA>  10.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        7
#> 8     257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        8
#> 9     257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        9
#> 10   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       10
#> 11   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       11
#> 12   <NA>  10.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       12
#> 13   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       13
#> 14    257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       14
#> 15      4   309   2.8  1.06   013       <NA>   <NA>   <NA>   <NA>       15
#> 16      N   018  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       16
#> 17     43   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 18     36   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 19     48   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 20   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       17
#> 21   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       18
#> 22   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       19
#> 23   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       20
#> 24   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       21
#> 25   <NA>  10.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       22
#> 26   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       23
#> 27    257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       24
#> 28   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       25
#> 29   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       26
#> 30   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       27
#> 31   <NA>  10.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       28
#> 32   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       29
#> 33    257   5.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       30
#> 34   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       31
#> 35   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       32
#> 36    257   5.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       33
#> 37   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       34
#> 38      4   326   0.4  2.97   037       <NA>   <NA>   <NA>   <NA>       35
#> 39      N   076  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       36
#> 40      6   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 41      2   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 42      9   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 43   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       37
#> 44   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       38
#> 45    1.3  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       39
#> 46   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       40
#> 47    0.9  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       41
#> 48    0.5  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       42
#> 49   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       43
#> 50    0.7   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       44
#> 51    0.7   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       45
#> 52    1.3  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       46
#> 53   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       47
#> 54   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       48
#> 55   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       49
#> 56   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       50
#> 57   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       51
#> 58   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       52
#> 59   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       53
#> 60   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       54
#> 61   <NA>  10.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       55
#> 62   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       56
#> 63    257   5.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       57
#> 64   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       58
#> 65   0.03     1  <NA>  <NA>     a          n   <NA>   <NA>   <NA>       59
#> 66   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       60
#> 67   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       61
#> 68   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       62
#> 69    257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       63
#> 70   <NA>  08.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       64
#> 71   <NA>  <NA>  <NA>  <NA>  <NA>              <NA>   <NA>   <NA>       65
#> 72   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       66
#> 73   <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       67
#> 74   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       68
#> 75    257   5.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       69
#> 76   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       70
#> 77   <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       71
#> 78    257   4.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       72
#> 79   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       73
#> 80   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       74
#> 81  avoid  <NA>  <NA>  <NA>  <NA>              <NA>   <NA>   <NA>       75
#> 82   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       76
#> 83   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       77
#> 84    257   3.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       78
#> 85    257   2.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       79
#> 86   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       80
#> 87   <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       81
#> 88   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       82
#> 89    257   2.5  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       83
#> 90   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       84
#> 91   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       85
#> 92   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       86
#> 93   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       87
#> 94   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       88
#> 95   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       89
#> 96   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       90
#> 97   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       91
#> 98   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       92
#> 99   <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       93
#> 100  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       94
#> 101  <NA>  09.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       95
#> 102  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       96
#> 103   257   5.8  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       97
#> 104  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       98
#> 105  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       99
#> 106  <NA>  09.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      100
#> 107  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      101
#> 108   257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      102
#> 109  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      103
#> 110   257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      104
#> 111  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      105
#> 112  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      106
#> 113  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      107
#> 114  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      108
#> 115  <NA>  09.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      109
#> 116  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      110
#> 117   257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      111
#> 118  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      112
#> 119  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      113
#> 120  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      114
#> 121  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      115
#> 122  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      116
#> 123  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      117
#> 124  0.03     2  <NA>  <NA>     a          n   <NA>   <NA>   <NA>      118
#> 125  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      119
#> 126  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      120
#> 127  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      121
#> 128  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      122
#> 129  <NA>  09.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      123
#> 130  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      124
#> 131   257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      125
#> 132  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      126
#> 133  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      127
#> 134  <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      128
#> 135  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      129
#> 136   257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      130
#> 137  0.02     1  <NA>  <NA>     j          n   <NA>   <NA>   <NA>      131
#> 138  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      132
#> 139  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      133
#> 140  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      134
#> 141  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      135
#> 142     4   270  14.0  0.28   015       <NA>   <NA>   <NA>   <NA>      136
#> 143     N   037  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      137
#> 144    11   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 145    12   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 146     9   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 147  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      138
#> 148  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      139
#> 149  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      140
#> 150  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      141
#> 151  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      142
#> 152  <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      143
#> 153  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      144
#> 154   257   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      145
#> 155  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      146
#> 156  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      147
#> 157  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      148
#> 158  <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      149
#> 159  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      150
#> 160   243   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      151
#> 161  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      152
#> 162  0.05     1  <NA>  <NA>     a       <NA>   <NA>   <NA>   <NA>      153
#> 163  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      154
#> 164  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      155
#> 165  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      156
#> 166  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      157
#> 167  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      158
#> 168  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      159
#> 169  <NA>  08.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      160
#> 170  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      161
#> 171   243   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      162
#> 172  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      163
#> 173  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      164
#> 174  <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      165
#> 175  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      166
#> 176     4   344   0.2  3.68   002       <NA>   <NA>   <NA>   <NA>      167
#> 177     Y   016  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      168
#> 178    46   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 179    28   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 180    66   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 181  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      169
#> 182   3.0  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      170
#> 183  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      171
#> 184  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      172
#> 185  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      173
#> 186 er th e sig hting  <NA>  <NA>              <NA>   <NA>   <NA>      174
#> 187  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      175
#> 188  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      176
#> 189  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      177
#> 190  <NA>  06.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      178
#> 191  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      179
#> 192   243   6.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      180
#> 193     4   070   1.4  1.66   036       <NA>   <NA>   <NA>   <NA>      181
#> 194     N   013   016  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      182
#> 195    37    68    32  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 196    35    75    25  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 197    29    65    35  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 198    66    80    20  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 199  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      183
#> 200  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      184
#> 201  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      185
#> 202  0.23     2     F  17.0     A          Y   <NA>   <NA>   <NA>      186
#> 203  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      187
#> 204  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      188
#> 205  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      189
#> 206  0.01     1  <NA>  <NA>     A          N   <NA>   <NA>   <NA>      190
#> 207  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      191
#> 208  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      192
#> 209  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      193
#> 210  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      194
#> 211  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      195
#> 212  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      196
#> 213  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      197
#> 214  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      198
#> 215     1   000  <NA>  0.00   018       <NA>   <NA>   <NA>   <NA>      199
#> 216     N   075  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      200
#> 217  <NA>   100  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 218 rst s een b y CO  while  ridi ng bow       <NA>   <NA>   <NA>      201
#> 219  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      202
#> 220  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      203
#> 221  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      204
#> 222  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      205
#> 223  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      206
#> 224  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      207
#> 225  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      208
#> 226  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      209
#> 227  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      210
#> 228  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      211
#> 229  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      212
#> 230  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>      213
#> 231     Y  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        1
#> 232  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        2
#> 233  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        3
#> 234  <NA>  05.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        4
#> 235  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        5
#> 236   040   4.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        6
#> 237  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        7
#> 238   1.7  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        8
#> 239  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>        9
#> 240  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       10
#> 241 ng ge ar     <NA>  <NA>  <NA>              <NA>   <NA>   <NA>       11
#> 242  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       12
#> 243  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       13
#> 244  <NA>  05.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       14
#> 245  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       15
#> 246   040   4.0  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       16
#> 247  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       17
#> 248     4   359   0.3  3.28   018       <NA>   <NA>   <NA>   <NA>       18
#> 249     N   018   277  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       19
#> 250   183    80    20  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 251   120    90    10  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 252     4   038   0.8  2.23   018       <NA>   <NA>   <NA>   <NA>       20
#> 253     N   016   277  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       21
#> 254  <NA>   016   016  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 255    21    60    40  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 256    16    56    44  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 257    12    70    30  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 258    36    98     2  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>     <NA>
#> 259  <NA>  <NA>  <NA>  <NA>  <NA>       <NA>   <NA>   <NA>   <NA>       22
#>           file_das line_num
#> 1   das_sample.das        1
#> 2   das_sample.das        2
#> 3   das_sample.das        3
#> 4   das_sample.das        4
#> 5   das_sample.das        5
#> 6   das_sample.das        6
#> 7   das_sample.das        7
#> 8   das_sample.das        8
#> 9   das_sample.das        9
#> 10  das_sample.das       10
#> 11  das_sample.das       11
#> 12  das_sample.das       12
#> 13  das_sample.das       13
#> 14  das_sample.das       14
#> 15  das_sample.das       15
#> 16  das_sample.das       16
#> 17  das_sample.das       17
#> 18  das_sample.das       18
#> 19  das_sample.das       19
#> 20  das_sample.das       20
#> 21  das_sample.das       21
#> 22  das_sample.das       22
#> 23  das_sample.das       23
#> 24  das_sample.das       24
#> 25  das_sample.das       25
#> 26  das_sample.das       26
#> 27  das_sample.das       27
#> 28  das_sample.das       28
#> 29  das_sample.das       29
#> 30  das_sample.das       30
#> 31  das_sample.das       31
#> 32  das_sample.das       32
#> 33  das_sample.das       33
#> 34  das_sample.das       34
#> 35  das_sample.das       35
#> 36  das_sample.das       36
#> 37  das_sample.das       37
#> 38  das_sample.das       38
#> 39  das_sample.das       39
#> 40  das_sample.das       40
#> 41  das_sample.das       41
#> 42  das_sample.das       42
#> 43  das_sample.das       43
#> 44  das_sample.das       44
#> 45  das_sample.das       45
#> 46  das_sample.das       46
#> 47  das_sample.das       47
#> 48  das_sample.das       48
#> 49  das_sample.das       49
#> 50  das_sample.das       50
#> 51  das_sample.das       51
#> 52  das_sample.das       52
#> 53  das_sample.das       53
#> 54  das_sample.das       54
#> 55  das_sample.das       55
#> 56  das_sample.das       56
#> 57  das_sample.das       57
#> 58  das_sample.das       58
#> 59  das_sample.das       59
#> 60  das_sample.das       60
#> 61  das_sample.das       61
#> 62  das_sample.das       62
#> 63  das_sample.das       63
#> 64  das_sample.das       64
#> 65  das_sample.das       65
#> 66  das_sample.das       66
#> 67  das_sample.das       67
#> 68  das_sample.das       68
#> 69  das_sample.das       69
#> 70  das_sample.das       70
#> 71  das_sample.das       71
#> 72  das_sample.das       72
#> 73  das_sample.das       73
#> 74  das_sample.das       74
#> 75  das_sample.das       75
#> 76  das_sample.das       76
#> 77  das_sample.das       77
#> 78  das_sample.das       78
#> 79  das_sample.das       79
#> 80  das_sample.das       80
#> 81  das_sample.das       81
#> 82  das_sample.das       82
#> 83  das_sample.das       83
#> 84  das_sample.das       84
#> 85  das_sample.das       85
#> 86  das_sample.das       86
#> 87  das_sample.das       87
#> 88  das_sample.das       88
#> 89  das_sample.das       89
#> 90  das_sample.das       90
#> 91  das_sample.das       91
#> 92  das_sample.das       92
#> 93  das_sample.das       93
#> 94  das_sample.das       94
#> 95  das_sample.das       95
#> 96  das_sample.das       96
#> 97  das_sample.das       97
#> 98  das_sample.das       98
#> 99  das_sample.das       99
#> 100 das_sample.das      100
#> 101 das_sample.das      101
#> 102 das_sample.das      102
#> 103 das_sample.das      103
#> 104 das_sample.das      104
#> 105 das_sample.das      105
#> 106 das_sample.das      106
#> 107 das_sample.das      107
#> 108 das_sample.das      108
#> 109 das_sample.das      109
#> 110 das_sample.das      110
#> 111 das_sample.das      111
#> 112 das_sample.das      112
#> 113 das_sample.das      113
#> 114 das_sample.das      114
#> 115 das_sample.das      115
#> 116 das_sample.das      116
#> 117 das_sample.das      117
#> 118 das_sample.das      118
#> 119 das_sample.das      119
#> 120 das_sample.das      120
#> 121 das_sample.das      121
#> 122 das_sample.das      122
#> 123 das_sample.das      123
#> 124 das_sample.das      124
#> 125 das_sample.das      125
#> 126 das_sample.das      126
#> 127 das_sample.das      127
#> 128 das_sample.das      128
#> 129 das_sample.das      129
#> 130 das_sample.das      130
#> 131 das_sample.das      131
#> 132 das_sample.das      132
#> 133 das_sample.das      133
#> 134 das_sample.das      134
#> 135 das_sample.das      135
#> 136 das_sample.das      136
#> 137 das_sample.das      137
#> 138 das_sample.das      138
#> 139 das_sample.das      139
#> 140 das_sample.das      140
#> 141 das_sample.das      141
#> 142 das_sample.das      142
#> 143 das_sample.das      143
#> 144 das_sample.das      144
#> 145 das_sample.das      145
#> 146 das_sample.das      146
#> 147 das_sample.das      147
#> 148 das_sample.das      148
#> 149 das_sample.das      149
#> 150 das_sample.das      150
#> 151 das_sample.das      151
#> 152 das_sample.das      152
#> 153 das_sample.das      153
#> 154 das_sample.das      154
#> 155 das_sample.das      155
#> 156 das_sample.das      156
#> 157 das_sample.das      157
#> 158 das_sample.das      158
#> 159 das_sample.das      159
#> 160 das_sample.das      160
#> 161 das_sample.das      161
#> 162 das_sample.das      162
#> 163 das_sample.das      163
#> 164 das_sample.das      164
#> 165 das_sample.das      165
#> 166 das_sample.das      166
#> 167 das_sample.das      167
#> 168 das_sample.das      168
#> 169 das_sample.das      169
#> 170 das_sample.das      170
#> 171 das_sample.das      171
#> 172 das_sample.das      172
#> 173 das_sample.das      173
#> 174 das_sample.das      174
#> 175 das_sample.das      175
#> 176 das_sample.das      176
#> 177 das_sample.das      177
#> 178 das_sample.das      178
#> 179 das_sample.das      179
#> 180 das_sample.das      180
#> 181 das_sample.das      181
#> 182 das_sample.das      182
#> 183 das_sample.das      183
#> 184 das_sample.das      184
#> 185 das_sample.das      185
#> 186 das_sample.das      186
#> 187 das_sample.das      187
#> 188 das_sample.das      188
#> 189 das_sample.das      189
#> 190 das_sample.das      190
#> 191 das_sample.das      191
#> 192 das_sample.das      192
#> 193 das_sample.das      193
#> 194 das_sample.das      194
#> 195 das_sample.das      195
#> 196 das_sample.das      196
#> 197 das_sample.das      197
#> 198 das_sample.das      198
#> 199 das_sample.das      199
#> 200 das_sample.das      200
#> 201 das_sample.das      201
#> 202 das_sample.das      202
#> 203 das_sample.das      203
#> 204 das_sample.das      204
#> 205 das_sample.das      205
#> 206 das_sample.das      206
#> 207 das_sample.das      207
#> 208 das_sample.das      208
#> 209 das_sample.das      209
#> 210 das_sample.das      210
#> 211 das_sample.das      211
#> 212 das_sample.das      212
#> 213 das_sample.das      213
#> 214 das_sample.das      214
#> 215 das_sample.das      215
#> 216 das_sample.das      216
#> 217 das_sample.das      217
#> 218 das_sample.das      218
#> 219 das_sample.das      219
#> 220 das_sample.das      220
#> 221 das_sample.das      221
#> 222 das_sample.das      222
#> 223 das_sample.das      223
#> 224 das_sample.das      224
#> 225 das_sample.das      225
#> 226 das_sample.das      226
#> 227 das_sample.das      227
#> 228 das_sample.das      228
#> 229 das_sample.das      229
#> 230 das_sample.das      230
#> 231 das_sample.das      231
#> 232 das_sample.das      232
#> 233 das_sample.das      233
#> 234 das_sample.das      234
#> 235 das_sample.das      235
#> 236 das_sample.das      236
#> 237 das_sample.das      237
#> 238 das_sample.das      238
#> 239 das_sample.das      239
#> 240 das_sample.das      240
#> 241 das_sample.das      241
#> 242 das_sample.das      242
#> 243 das_sample.das      243
#> 244 das_sample.das      244
#> 245 das_sample.das      245
#> 246 das_sample.das      246
#> 247 das_sample.das      247
#> 248 das_sample.das      248
#> 249 das_sample.das      249
#> 250 das_sample.das      250
#> 251 das_sample.das      251
#> 252 das_sample.das      252
#> 253 das_sample.das      253
#> 254 das_sample.das      254
#> 255 das_sample.das      255
#> 256 das_sample.das      256
#> 257 das_sample.das      257
#> 258 das_sample.das      258
#> 259 das_sample.das      259