Read DAS file(s)

Read one or more fixed-width DAS text file(s) generated by WinCruz into a data frame, where each line is data for a specific event

das_read(file, tz = "UTC")

Arguments

file
filename(s) of one or more DAS files

tz
character; see strptime. Default is UTC

Value

Data frame with DAS data read into columns. The data are read into the data frame as characters as described in 'Details', with the following exceptions:

Name	Class	Details

EffortDot	logical	TRUE if "." was present, and FALSE otherwise

DateTime	POSIXct	combination of 'Date' and 'Time' columns with time zone tz

Lat	numeric	'Latitude' column converted to decimal degrees in range [-90, 90]

Lon	numeric	'Longitude' column converted to decimal degrees in range [-180, 180]

Data#	character	leading/trailing whitespace trimmed for non-comment events (i.e. rows where 'Event' is not "C" or "c")

file_das	character	filename, extracted from the file argument using basename

event_num	integer	'Event number' coerced to an integer using as.integer

line_num	integer	line number of each data row

Details

Reads/parses DAS data into columns of a data frame. If file contains multiple filenames, then the individual data frames will be concatenated. The provided DAS file must adhere to the following column number and format specifications; note that 'Data#' columns may be referred to as 'Field#' columns in other documentation:

Item	Columns	Format

Sequence (event number)	1-3	Event

4	Effort dot	5

Time	6-11	HHMMSS or HHMM

Date	13-18	MMDDYY

Latitude	20-28	NDD:MM.MM

Longitude	30-39	WDDD:MM.MM

Data1	40-44	Data2

45-49	Data3	50-54

Data4	55-59	Data5

60-64	Data6	65-69

Data7	70-74	Data8

See das_format_pdf for more information about DAS format requirements

See also

Examples

# TODO
# das_read(system.file("das_sample.das", package = "swfscDAS"))