DAS sightings

Extract sightings and associated information from processed DAS data

das_sight(x, ...)

# S3 method for data.frame
das_sight(x, ...)

# S3 method for das_df
das_sight(
  x,
  return.format = c("default", "wide", "complete"),
  return.events = c("S", "K", "M", "G", "s", "k", "m", "g", "t", "p", "F"),
  ...
)

Arguments

x
an object of class das_df, or a data frame that can be coerced to class das_df

...
ignored

return.format
character; can be one of "default", "wide", "complete", or any partial match thereof (case sensitive). Formats described below

return.events
character; event codes included in the output. Must be one or more of: "S", "K", "M", "G", "s", "k", "m", "g", "t", "p", "F" (case-sensitive). The default is all of these event codes

Value

Data frame with 1) the columns from x, excluding the 'Data#' columns, and 2) columns with sighting information extracted from 'Data#' columns. See das_format_pdf for more information the sighting information. If return.format is "default", then there is one row for each species of each sighting event; if return.format is "wide", then there is one row for each sighting event; if return.format is "complete", then there is one row for every group size estimate for each sighting event.

The format-specific columns are described in their respective sections. The following sighting information columns are included in all return formats:

Sighting information	Column name	Notes

Sighting number	SightNo	Character

Subgroup code	Subgroup	Character

Daily sighting number	SightNoDaily	See below

Observer that made the sighting	Obs	Standard observer

ObsStd	Logical; TRUE if Obs is one of ObsL, Rec or ObsR, and FALSE otherwise (including if Obs is NA)	Bearing to the sighting

Bearing	Numeric; degrees, expected range 0 to 360	Number of reticle marks

Reticle	Numeric	Distance (nautical miles)

DistNm	Numeric	Sighting cue

Cue	Sighting method	Method

Photos of school?	Photos	Birds present with school?

Birds	Calibration school?	CalibSchool

Aerial photos taken?	PhotosAerial	Biopsy taken?

Biopsy	Probable sighting	Prob

Logical indicating if sighting has associated ? event; NA for non-S/K/M/G events	Number of species in sighting	nSp

NA for non-S/K/M/G events	Mixed species sighting	Mixed

Logical; TRUE if nSp > 1	Group size of school - best estimate	GsSchoolBest

See below	Group size of school - high estimate	GsSchoolHigh

See below	Group size of school - low estimate	GsSchoolLow

See below	Course (true heading) of school at resight	CourseSchool

NA for non-s/k/m events	Presence of associated JFR	TurtleJFR

NA for non-"t" events; JFR = jellyfish, floating debris, or red tide	Estimated turtle maturity	TurtleAge

NA for non-"t" events	Perpendicular distance (km) to sighting	PerpDistKm

SightNoDaily is a running count of the number of S/K/M/G sightings that occurred on each day. It is formatted as 'YYYYMMDD'_'running count', e.g. "20050101_1". The GsSchoolBest, GsSchoolHigh, and GsSchoolLow columns are either: 1) the arithmetic mean across observer estimates for the "default" and "wide" formats, or 2) the individual observer estimates for the "complete" format. Note that for non-"complete" formats, na.rm = TRUE is used when calculating the mean, and thus blank elements of estimates (but not the whole incomplete estimate) are ignored. To convert the perpendicular distance back to nautical miles, one would divide PerpDistKm by 1.852

Details

DAS events contain specific information in the 'Data#' columns, with the information depending on the event code for that row. The output data frame contains columns with this specific information extracted to dedicated columns as described below. This function recognizes the following types of sightings: marine mammal sightings (event codes "S", "K", or "M"), marine mammal resights (codes "s", "k", "m"), marine mammal subgroup sightings (code "G"), marine mammal subgroup resights (code "g"), turtle sightings (code "t"), pinniped sightings (code "p") and fishing vessel sightings (code "F"). Warnings are printed if all S, K, M, and G events (and only these events) are not followed by an A event and at least one numeric event. See das_format_pdf for more information about events and event formats,

The return.events argument simply provides a shortcut for filtering the output of das_sight by event codes

Abbreviations used in output column names: Gs = group size, Sp = species, Nm = nautical mile, Perc = percentage, Prob = probable, GsSchool = school-level group size info

This function makes the following assumptions, and alterations to the raw DAS data:

"A" events immediately following an S/K/M/G event have the same sighting number (Data1 value) as the S/K/M/G event

The 'nSp' column is equivalent to the number of non-NA values across the 'Data5', 'Data6', 'Data7', and 'Data8' columns for the pertinent "A" event

The following data are coerced to a numeric using as.numeric: Bearing, Reticle, DistNm, Cue, Method, species percentages, and group sizes (including for t, p, and F events). Note that if there are any formatting errors and these data are not numeric, the function will likely print a warning message

The values for the following columns are capitalized using toupper: 'Birds', 'Photos', 'CalibSchool', 'PhotosAerial', 'Biopsy', 'TurtleAge', and 'TurtleCapt'

The "default" format output

One row for each species of each sighting event, with group size estimates calculated via an arithmetic mean across observer estimates (using mean with na.rm = TRUE)

Sighting information columns/formats present specifically in the "default" format output:

Sighting information	Column name	Notes

Species code	SpCode	Boat type or mammal, turtle, or pinniped species codes

Probable species code	SpCodeProb	Probable mammal species codes; NA if none or not applicable

Group size of species - best estimate	GsSpBest	Arithmetic mean of the product of GsSchoolBest and the corresponding species percentage

Group size of species - high estimate	GsSpHigh	Arithmetic mean of the product of GsSchoolHigh and the corresponding species percentage

Group size of species - low estimate	GsSpLow	Arithmetic mean of the product of GsSchoolLow and the corresponding species percentage

The "wide" and "complete" format outputs

The "wide" and "complete" options have very similar columns in their output date frames. There are two main differences: 1) the "wide" format has one row for each sighting event, while the complete format has a row for every observer estimate for each sightings, and thus 2) in the "wide" format, all numeric information for which there are multiple observer estimates (school group size, species percentage, etc.) are averaged across estimated via an arithmetic mean (using mean with na.rm = TRUE)

With these formats, note that the species/type code and group size for turtle, pinniped, and boat sightings are in their own column

Sighting information columns present in the "wide" and "complete" format outputs:

Sighting information	Column name	Notes

Observer code - estimate	ObsEstimate	See below

Species 1 code	SpCode1	The species code for turtle and pinniped sightings

Species 2 code	SpCode2	Species 3 code

SpCode3	Species 4 code	SpCode4

Species 1 probable code	SpCodeProb1	Extracted from '?' event

Species 2 probable code	SpCodeProb2	Extracted from '?' event

Species 3 probable code	SpCodeProb3	Extracted from '?' event

Species 4 probable code	SpCodeProb4	Extracted from '?' event

Percentage of Sp 1 in school	SpPerc1	Percentage of Sp 2 in school

SpPerc2	Percentage of Sp 3 in school	SpPerc3

Percentage of Sp 4 in school	SpPerc4	Group size of species 1

GsSpBest1	Present in "wide" output only; see below	Group size of species 2

GsSpBest2	Present in "wide" output only; see below	Group size of species 3

GsSpBest3	Present in "wide" output only; see below	Group size of species 4

GsSpBest4	Present in "wide" output only; see below	Turtle species

TurtleSp	NA for non-"t" events	Sighting group size

TurtleGs	NA for non-"t" events; i.e. number of turtles	Was turtle captured?

TurtleCapt	NA for non-"t" events	Pinniped species

PinnipedSp	NA for non-"p" events	Sighting group size

PinnipedGs	NA for non-"p" events; i.e. number of pinnipeds	Boat or gear type

BoatType	NA for non-"F" events	Sighting group size

BoatGs	NA for non-"F" events; i.e. number of boats

ObsEstimate refers to the code of the observer that made the corresponding estimate. For the "wide" format, ObsEstimate is a list-column of all of the observer codes that provided an estimate. Also in the "wide" format, the GsSpBest# columns are the product of GsSchoolBest and the corresponding species percentage

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_sight(y.proc)

#>    Event            DateTime      Lat       Lon OnEffort Cruise Mode EffType
#> 1      S 2013-01-13 06:46:02 39.36617 -137.5820     TRUE   1000    C       S
#> 2      S 2013-01-13 07:56:22 39.51767 -137.5285     TRUE   1000    C       S
#> 3      s 2013-01-13 08:06:00 39.54217 -137.5263    FALSE   1000    C       S
#> 4      s 2013-01-13 08:08:58 39.55000 -137.5255    FALSE   1000    C       S
#> 5      s 2013-01-13 08:15:32 39.56000 -137.5218    FALSE   1000    C       S
#> 6      s 2013-01-13 08:17:28 39.56233 -137.5210    FALSE   1000    C       S
#> 7      s 2013-01-13 08:19:39 39.56483 -137.5197    FALSE   1000    C       S
#> 8      s 2013-01-13 08:26:45 39.56400 -137.5140    FALSE   1000    C       S
#> 9      t 2013-01-13 09:34:27 39.59733 -137.4400     TRUE   1000    C       S
#> 10     t 2013-01-13 13:35:18 40.14000 -137.2048    FALSE   1000    C       S
#> 11     t 2013-01-13 14:02:55 40.18283 -137.1622     TRUE   1000    C       S
#> 12     S 2013-01-13 14:37:56 40.26567 -137.1350     TRUE   1000    C       S
#> 13     t 2013-01-13 15:36:47 40.36050 -137.0978     TRUE   1000    C       S
#> 14     S 2013-01-13 16:29:50 40.45133 -137.0628     TRUE   1000    C       S
#> 15     s 2013-01-13 16:36:34 40.46667 -137.0570    FALSE   1000    C       S
#> 16     S 2013-01-13 17:00:45 40.52400 -137.0522     TRUE   1000    C       S
#> 17     S 2013-01-13 17:00:45 40.52400 -137.0522     TRUE   1000    C       S
#> 18     t 2013-01-13 17:08:11 40.52083 -137.0383    FALSE   1000    C       S
#> 19     t 2013-01-13 17:45:10 40.50450 -137.0437    FALSE   1000    C       S
#> 20     S 2013-01-14 09:17:21 40.75183 -135.7748    FALSE   1000 <NA>    <NA>
#> 21     F 2013-01-14 11:25:32 40.98950 -135.5965     TRUE   1000    C       S
#> 22     S 2013-01-14 11:47:51 41.04017 -135.5635     TRUE   1000    C       S
#> 23     S 2013-01-14 11:47:51 41.04017 -135.5635     TRUE   1000    C       S
#> 24     S 2013-01-14 11:49:14 41.04333 -135.5615     TRUE   1000    C       S
#> 25     S 2013-01-14 11:49:14 41.04333 -135.5615     TRUE   1000    C       S
#>    ESWsides Course SpdKt Bft SwellHght WindSpdKt RainFog HorizSun VertSun Glare
#> 1         2     25  10.2   3         3        10       1       NA      NA    NA
#> 2         2     26   9.7   3         3        10       3        2       2 FALSE
#> 3         2     26   9.7   3         3        10       3        2       2 FALSE
#> 4         2     26   9.7   3         3        10       3        2       2 FALSE
#> 5         2     26   9.7   3         3        10       3        2       2 FALSE
#> 6         2     26   9.7   3         3        10       3        2       2 FALSE
#> 7         2     26   9.7   3         3        10       3        2       2 FALSE
#> 8         2     26   9.7   3         3        10       3        2       2 FALSE
#> 9         2     27   9.0   3         3        10       1        2       2 FALSE
#> 10        2     35   9.2   3         3         9       1       12      12 FALSE
#> 11        2     20   9.3   3         3         6       1        8       1 FALSE
#> 12        2     20   9.3   3         3         6       1        8       1 FALSE
#> 13        2     16   8.9   2         3         6       1        9       1 FALSE
#> 14        2     25   8.9   2         3         6       1        8       2 FALSE
#> 15        2     25   8.9   2         3         6       1        8       2 FALSE
#> 16        2     30   9.5   2         3         6       1        8       2 FALSE
#> 17        2     30   9.5   2         3         6       1        8       2 FALSE
#> 18        2     30   9.5   2         3         6       1        8       2 FALSE
#> 19        2     30   9.5   2         3         6       1        8       2 FALSE
#> 20       NA     NA    NA  NA        NA        NA      NA       NA      NA    NA
#> 21        2     35   9.5   2         1         5       3       NA      NA    NA
#> 22        2     23   9.6   2         1         5       3       NA      NA    NA
#> 23        2     23   9.6   2         1         5       3       NA      NA    NA
#> 24        2     23   9.6   2         1         5       3       NA      NA    NA
#> 25        2     23   9.6   2         1         5       3       NA      NA    NA
#>    Vis ObsL  Rec ObsR ObsInd EffortDot EventNum       file_das line_num SightNo
#> 1  6.0  208  280  001   <NA>      TRUE       15 das_sample.das       15    1406
#> 2  5.5  125  208  280   <NA>      TRUE       35 das_sample.das       38    1407
#> 3  5.5  125  208  280   <NA>     FALSE       39 das_sample.das       45    1407
#> 4  5.5  125  208  280   <NA>     FALSE       41 das_sample.das       47    1407
#> 5  5.5  125  208  280   <NA>     FALSE       42 das_sample.das       48    1407
#> 6  5.5  125  208  280   <NA>     FALSE       44 das_sample.das       50    1407
#> 7  5.5  125  208  280   <NA>     FALSE       45 das_sample.das       51    1407
#> 8  5.5  125  208  280   <NA>     FALSE       46 das_sample.das       52    1407
#> 9  5.5  001  126  149   <NA>      TRUE       59 das_sample.das       65    <NA>
#> 10 6.0  126  149  125   <NA>     FALSE      118 das_sample.das      124    <NA>
#> 11 6.0  280  001  126   <NA>      TRUE      131 das_sample.das      137    <NA>
#> 12 6.0  280  001  126   <NA>      TRUE      136 das_sample.das      142    1408
#> 13 6.0  125  208  280   <NA>      TRUE      153 das_sample.das      162    <NA>
#> 14 6.0  149  125  208   <NA>      TRUE      167 das_sample.das      176    1409
#> 15 6.0  149  125  208   <NA>     FALSE      170 das_sample.das      182    1409
#> 16 6.0  126  149  125   <NA>      TRUE      181 das_sample.das      193    1410
#> 17 6.0  126  149  125   <NA>      TRUE      181 das_sample.das      193    1410
#> 18 6.0  126  149  125   <NA>     FALSE      186 das_sample.das      202    <NA>
#> 19 6.0  126  149  125   <NA>     FALSE      190 das_sample.das      206    <NA>
#> 20  NA <NA> <NA> <NA>   <NA>     FALSE      199 das_sample.das      215    1411
#> 21 4.0  149  125  208   <NA>      TRUE        8 das_sample.das      238    <NA>
#> 22 4.0  149  125  208   <NA>      TRUE       18 das_sample.das      248    1412
#> 23 4.0  149  125  208   <NA>      TRUE       18 das_sample.das      248    1412
#> 24 4.0  149  125  208   <NA>      TRUE       20 das_sample.das      252    1413
#> 25 4.0  149  125  208   <NA>      TRUE       20 das_sample.das      252    1413
#>    Subgroup SightNoDaily  Obs ObsStd Bearing Reticle DistNm Cue Method Photos
#> 1      <NA>   20130113_1  208   TRUE     309     2.8   1.06   3      4      N
#> 2      <NA>   20130113_2  125   TRUE     326     0.4   2.97   3      4      Y
#> 3      <NA>         <NA> <NA>  FALSE      11     2.0   1.30  NA     NA   <NA>
#> 4      <NA>         <NA> <NA>  FALSE       5     3.5   0.90  NA     NA   <NA>
#> 5      <NA>         <NA> <NA>  FALSE      50      NA   0.50  NA     NA   <NA>
#> 6      <NA>         <NA> <NA>  FALSE      71     4.5   0.70  NA     NA   <NA>
#> 7      <NA>         <NA> <NA>  FALSE     104     4.5   0.70  NA     NA   <NA>
#> 8      <NA>         <NA> <NA>  FALSE       2     2.2   1.30  NA     NA   <NA>
#> 9      <NA>         <NA>  280  FALSE     120      NA   0.03  NA     NA   <NA>
#> 10     <NA>         <NA>  149   TRUE     270      NA   0.03  NA     NA   <NA>
#> 11     <NA>         <NA>  228  FALSE     300      NA   0.02  NA     NA   <NA>
#> 12     <NA>   20130113_3  280   TRUE     270    14.0   0.28   3      4      N
#> 13     <NA>         <NA>  231  FALSE      45      NA   0.05  NA     NA   <NA>
#> 14     <NA>   20130113_4  149   TRUE     344     0.2   3.68   3      4      Y
#> 15     <NA>         <NA> <NA>  FALSE     356     0.4   3.00  NA     NA   <NA>
#> 16     <NA>   20130113_5  125   TRUE      70     1.4   1.66   3      4      Y
#> 17     <NA>   20130113_5  125   TRUE      70     1.4   1.66   3      4      Y
#> 18     <NA>         <NA>  280  FALSE      42    17.0   0.23  NA     NA   <NA>
#> 19     <NA>         <NA>  099  FALSE     180      NA   0.01  NA     NA   <NA>
#> 20     <NA>   20130114_1  280  FALSE       0      NA   0.00   3      1      N
#> 21     <NA>         <NA>  149   TRUE     309     1.7   1.47  NA     NA   <NA>
#> 22     <NA>   20130114_2  149   TRUE     359     0.3   3.28   2      4      Y
#> 23     <NA>   20130114_2  149   TRUE     359     0.3   3.28   2      4      Y
#> 24     <NA>   20130114_3  208   TRUE      38     0.8   2.23   3      4      Y
#> 25     <NA>   20130114_3  208   TRUE      38     0.8   2.23   3      4      Y
#>    Birds CalibSchool PhotosAerial Biopsy  Prob nSp Mixed SpCode SpCodeProb
#> 1      N        <NA>         <NA>   <NA> FALSE   1 FALSE    018       <NA>
#> 2      N        <NA>         <NA>   <NA> FALSE   1 FALSE    076       <NA>
#> 3   <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 4   <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 5   <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 6   <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 7   <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 8   <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 9   <NA>        <NA>         <NA>   <NA>    NA  NA    NA     LV       <NA>
#> 10  <NA>        <NA>         <NA>   <NA>    NA  NA    NA     DC       <NA>
#> 11  <NA>        <NA>         <NA>   <NA>    NA  NA    NA     DC       <NA>
#> 12     N        <NA>         <NA>   <NA> FALSE   1 FALSE    037       <NA>
#> 13  <NA>        <NA>         <NA>   <NA>    NA  NA    NA     DC       <NA>
#> 14     Y        <NA>         <NA>   <NA> FALSE   1 FALSE    016       <NA>
#> 15  <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 16     N        <NA>         <NA>   <NA> FALSE   2  TRUE    013       <NA>
#> 17     N        <NA>         <NA>   <NA> FALSE   2  TRUE    016       <NA>
#> 18  <NA>        <NA>         <NA>   <NA>    NA  NA    NA     DC       <NA>
#> 19  <NA>        <NA>         <NA>   <NA>    NA  NA    NA     LV       <NA>
#> 20     N        <NA>         <NA>   <NA> FALSE   1 FALSE    075       <NA>
#> 21  <NA>        <NA>         <NA>   <NA>    NA  NA    NA   <NA>       <NA>
#> 22     N        <NA>         <NA>   <NA> FALSE   2  TRUE    018       <NA>
#> 23     N        <NA>         <NA>   <NA> FALSE   2  TRUE    277       <NA>
#> 24     N        <NA>         <NA>   <NA>  TRUE   2  TRUE    016        016
#> 25     N        <NA>         <NA>   <NA>  TRUE   2  TRUE    277        016
#>    GsSchoolBest GsSchoolHigh GsSchoolLow  GsSpBest GsSpHigh    GsSpLow
#> 1            NA           NA   42.333333        NA       NA  42.333333
#> 2       8.00000        14.00    5.666667   8.00000   14.000   5.666667
#> 3            NA           NA          NA        NA       NA         NA
#> 4            NA           NA          NA        NA       NA         NA
#> 5            NA           NA          NA        NA       NA         NA
#> 6            NA           NA          NA        NA       NA         NA
#> 7            NA           NA          NA        NA       NA         NA
#> 8            NA           NA          NA        NA       NA         NA
#> 9       1.00000           NA          NA   1.00000       NA         NA
#> 10      2.00000           NA          NA   2.00000       NA         NA
#> 11      1.00000           NA          NA   1.00000       NA         NA
#> 12     10.66667        20.00   10.666667  10.66667   20.000  10.666667
#> 13      1.00000           NA          NA   1.00000       NA         NA
#> 14     46.66667        79.00   46.666667  46.66667   79.000  46.666667
#> 15           NA           NA          NA        NA       NA         NA
#> 16     41.75000        72.75   41.750000  30.76500   53.165  30.765000
#> 17     41.75000        72.75   41.750000  10.98500   19.585  10.985000
#> 18      2.00000           NA          NA   2.00000       NA         NA
#> 19      1.00000           NA          NA   1.00000       NA         NA
#> 20           NA           NA          NA        NA       NA         NA
#> 21           NA           NA          NA        NA       NA         NA
#> 22    151.50000       249.00  151.500000 127.20000  207.700 127.200000
#> 23    151.50000       249.00  151.500000  24.30000   41.300  24.300000
#> 24     21.25000        37.75   21.250000  16.31000   27.935  16.310000
#> 25     21.25000        37.75   21.250000   4.94000    9.815   4.940000
#>    CourseSchool TurtleJFR TurtleAge TurtleCapt   PerpDistKm
#> 1            NA      <NA>      <NA>       <NA> 1.525631e+00
#> 2            NA      <NA>      <NA>       <NA> 3.075807e+00
#> 3            NA      <NA>      <NA>       <NA> 4.593917e-01
#> 4            NA      <NA>      <NA>       <NA> 1.452712e-01
#> 5            NA      <NA>      <NA>       <NA> 7.093572e-01
#> 6           100      <NA>      <NA>       <NA> 1.225770e+00
#> 7           100      <NA>      <NA>       <NA> 1.257891e+00
#> 8            NA      <NA>      <NA>       <NA> 8.402403e-02
#> 9            NA      <NA>         A          N 4.811637e-02
#> 10           NA      <NA>         A          N 5.556000e-02
#> 11           NA      <NA>         J          N 3.207758e-02
#> 12           NA      <NA>      <NA>       <NA> 5.185600e-01
#> 13           NA      <NA>         A       <NA> 6.547809e-02
#> 14           NA      <NA>      <NA>       <NA> 1.878568e+00
#> 15           NA      <NA>      <NA>       <NA> 3.875670e-01
#> 16           NA      <NA>      <NA>       <NA> 2.888916e+00
#> 17           NA      <NA>      <NA>       <NA> 2.888916e+00
#> 18           NA         F         A          Y 2.850229e-01
#> 19           NA      <NA>         A          N 2.268046e-18
#> 20           NA      <NA>      <NA>       <NA> 0.000000e+00
#> 21           NA      <NA>      <NA>       <NA> 2.115733e+00
#> 22           NA      <NA>      <NA>       <NA> 1.060157e-01
#> 23           NA      <NA>      <NA>       <NA> 1.060157e-01
#> 24           NA      <NA>      <NA>       <NA> 2.542657e+00
#> 25           NA      <NA>      <NA>       <NA> 2.542657e+00

das_sight(y.proc, return.format = "complete")

#>    Event            DateTime      Lat       Lon OnEffort Cruise Mode EffType
#> 1      S 2013-01-13 06:46:02 39.36617 -137.5820     TRUE   1000    C       S
#> 2      S 2013-01-13 06:46:02 39.36617 -137.5820     TRUE   1000    C       S
#> 3      S 2013-01-13 06:46:02 39.36617 -137.5820     TRUE   1000    C       S
#> 4      S 2013-01-13 07:56:22 39.51767 -137.5285     TRUE   1000    C       S
#> 5      S 2013-01-13 07:56:22 39.51767 -137.5285     TRUE   1000    C       S
#> 6      S 2013-01-13 07:56:22 39.51767 -137.5285     TRUE   1000    C       S
#> 7      s 2013-01-13 08:06:00 39.54217 -137.5263    FALSE   1000    C       S
#> 8      s 2013-01-13 08:08:58 39.55000 -137.5255    FALSE   1000    C       S
#> 9      s 2013-01-13 08:15:32 39.56000 -137.5218    FALSE   1000    C       S
#> 10     s 2013-01-13 08:17:28 39.56233 -137.5210    FALSE   1000    C       S
#> 11     s 2013-01-13 08:19:39 39.56483 -137.5197    FALSE   1000    C       S
#> 12     s 2013-01-13 08:26:45 39.56400 -137.5140    FALSE   1000    C       S
#> 13     t 2013-01-13 09:34:27 39.59733 -137.4400     TRUE   1000    C       S
#> 14     t 2013-01-13 13:35:18 40.14000 -137.2048    FALSE   1000    C       S
#> 15     t 2013-01-13 14:02:55 40.18283 -137.1622     TRUE   1000    C       S
#> 16     S 2013-01-13 14:37:56 40.26567 -137.1350     TRUE   1000    C       S
#> 17     S 2013-01-13 14:37:56 40.26567 -137.1350     TRUE   1000    C       S
#> 18     S 2013-01-13 14:37:56 40.26567 -137.1350     TRUE   1000    C       S
#> 19     t 2013-01-13 15:36:47 40.36050 -137.0978     TRUE   1000    C       S
#> 20     S 2013-01-13 16:29:50 40.45133 -137.0628     TRUE   1000    C       S
#> 21     S 2013-01-13 16:29:50 40.45133 -137.0628     TRUE   1000    C       S
#> 22     S 2013-01-13 16:29:50 40.45133 -137.0628     TRUE   1000    C       S
#> 23     s 2013-01-13 16:36:34 40.46667 -137.0570    FALSE   1000    C       S
#> 24     S 2013-01-13 17:00:45 40.52400 -137.0522     TRUE   1000    C       S
#> 25     S 2013-01-13 17:00:45 40.52400 -137.0522     TRUE   1000    C       S
#> 26     S 2013-01-13 17:00:45 40.52400 -137.0522     TRUE   1000    C       S
#> 27     S 2013-01-13 17:00:45 40.52400 -137.0522     TRUE   1000    C       S
#> 28     t 2013-01-13 17:08:11 40.52083 -137.0383    FALSE   1000    C       S
#> 29     t 2013-01-13 17:45:10 40.50450 -137.0437    FALSE   1000    C       S
#> 30     S 2013-01-14 09:17:21 40.75183 -135.7748    FALSE   1000 <NA>    <NA>
#> 31     F 2013-01-14 11:25:32 40.98950 -135.5965     TRUE   1000    C       S
#> 32     S 2013-01-14 11:47:51 41.04017 -135.5635     TRUE   1000    C       S
#> 33     S 2013-01-14 11:47:51 41.04017 -135.5635     TRUE   1000    C       S
#> 34     S 2013-01-14 11:49:14 41.04333 -135.5615     TRUE   1000    C       S
#> 35     S 2013-01-14 11:49:14 41.04333 -135.5615     TRUE   1000    C       S
#> 36     S 2013-01-14 11:49:14 41.04333 -135.5615     TRUE   1000    C       S
#> 37     S 2013-01-14 11:49:14 41.04333 -135.5615     TRUE   1000    C       S
#>    ESWsides Course SpdKt Bft SwellHght WindSpdKt RainFog HorizSun VertSun Glare
#> 1         2     25  10.2   3         3        10       1       NA      NA    NA
#> 2         2     25  10.2   3         3        10       1       NA      NA    NA
#> 3         2     25  10.2   3         3        10       1       NA      NA    NA
#> 4         2     26   9.7   3         3        10       3        2       2 FALSE
#> 5         2     26   9.7   3         3        10       3        2       2 FALSE
#> 6         2     26   9.7   3         3        10       3        2       2 FALSE
#> 7         2     26   9.7   3         3        10       3        2       2 FALSE
#> 8         2     26   9.7   3         3        10       3        2       2 FALSE
#> 9         2     26   9.7   3         3        10       3        2       2 FALSE
#> 10        2     26   9.7   3         3        10       3        2       2 FALSE
#> 11        2     26   9.7   3         3        10       3        2       2 FALSE
#> 12        2     26   9.7   3         3        10       3        2       2 FALSE
#> 13        2     27   9.0   3         3        10       1        2       2 FALSE
#> 14        2     35   9.2   3         3         9       1       12      12 FALSE
#> 15        2     20   9.3   3         3         6       1        8       1 FALSE
#> 16        2     20   9.3   3         3         6       1        8       1 FALSE
#> 17        2     20   9.3   3         3         6       1        8       1 FALSE
#> 18        2     20   9.3   3         3         6       1        8       1 FALSE
#> 19        2     16   8.9   2         3         6       1        9       1 FALSE
#> 20        2     25   8.9   2         3         6       1        8       2 FALSE
#> 21        2     25   8.9   2         3         6       1        8       2 FALSE
#> 22        2     25   8.9   2         3         6       1        8       2 FALSE
#> 23        2     25   8.9   2         3         6       1        8       2 FALSE
#> 24        2     30   9.5   2         3         6       1        8       2 FALSE
#> 25        2     30   9.5   2         3         6       1        8       2 FALSE
#> 26        2     30   9.5   2         3         6       1        8       2 FALSE
#> 27        2     30   9.5   2         3         6       1        8       2 FALSE
#> 28        2     30   9.5   2         3         6       1        8       2 FALSE
#> 29        2     30   9.5   2         3         6       1        8       2 FALSE
#> 30       NA     NA    NA  NA        NA        NA      NA       NA      NA    NA
#> 31        2     35   9.5   2         1         5       3       NA      NA    NA
#> 32        2     23   9.6   2         1         5       3       NA      NA    NA
#> 33        2     23   9.6   2         1         5       3       NA      NA    NA
#> 34        2     23   9.6   2         1         5       3       NA      NA    NA
#> 35        2     23   9.6   2         1         5       3       NA      NA    NA
#> 36        2     23   9.6   2         1         5       3       NA      NA    NA
#> 37        2     23   9.6   2         1         5       3       NA      NA    NA
#>    Vis ObsL  Rec ObsR ObsInd EffortDot EventNum       file_das line_num SightNo
#> 1  6.0  208  280  001   <NA>      TRUE       15 das_sample.das       15    1406
#> 2  6.0  208  280  001   <NA>      TRUE       15 das_sample.das       15    1406
#> 3  6.0  208  280  001   <NA>      TRUE       15 das_sample.das       15    1406
#> 4  5.5  125  208  280   <NA>      TRUE       35 das_sample.das       38    1407
#> 5  5.5  125  208  280   <NA>      TRUE       35 das_sample.das       38    1407
#> 6  5.5  125  208  280   <NA>      TRUE       35 das_sample.das       38    1407
#> 7  5.5  125  208  280   <NA>     FALSE       39 das_sample.das       45    1407
#> 8  5.5  125  208  280   <NA>     FALSE       41 das_sample.das       47    1407
#> 9  5.5  125  208  280   <NA>     FALSE       42 das_sample.das       48    1407
#> 10 5.5  125  208  280   <NA>     FALSE       44 das_sample.das       50    1407
#> 11 5.5  125  208  280   <NA>     FALSE       45 das_sample.das       51    1407
#> 12 5.5  125  208  280   <NA>     FALSE       46 das_sample.das       52    1407
#> 13 5.5  001  126  149   <NA>      TRUE       59 das_sample.das       65    <NA>
#> 14 6.0  126  149  125   <NA>     FALSE      118 das_sample.das      124    <NA>
#> 15 6.0  280  001  126   <NA>      TRUE      131 das_sample.das      137    <NA>
#> 16 6.0  280  001  126   <NA>      TRUE      136 das_sample.das      142    1408
#> 17 6.0  280  001  126   <NA>      TRUE      136 das_sample.das      142    1408
#> 18 6.0  280  001  126   <NA>      TRUE      136 das_sample.das      142    1408
#> 19 6.0  125  208  280   <NA>      TRUE      153 das_sample.das      162    <NA>
#> 20 6.0  149  125  208   <NA>      TRUE      167 das_sample.das      176    1409
#> 21 6.0  149  125  208   <NA>      TRUE      167 das_sample.das      176    1409
#> 22 6.0  149  125  208   <NA>      TRUE      167 das_sample.das      176    1409
#> 23 6.0  149  125  208   <NA>     FALSE      170 das_sample.das      182    1409
#> 24 6.0  126  149  125   <NA>      TRUE      181 das_sample.das      193    1410
#> 25 6.0  126  149  125   <NA>      TRUE      181 das_sample.das      193    1410
#> 26 6.0  126  149  125   <NA>      TRUE      181 das_sample.das      193    1410
#> 27 6.0  126  149  125   <NA>      TRUE      181 das_sample.das      193    1410
#> 28 6.0  126  149  125   <NA>     FALSE      186 das_sample.das      202    <NA>
#> 29 6.0  126  149  125   <NA>     FALSE      190 das_sample.das      206    <NA>
#> 30  NA <NA> <NA> <NA>   <NA>     FALSE      199 das_sample.das      215    1411
#> 31 4.0  149  125  208   <NA>      TRUE        8 das_sample.das      238    <NA>
#> 32 4.0  149  125  208   <NA>      TRUE       18 das_sample.das      248    1412
#> 33 4.0  149  125  208   <NA>      TRUE       18 das_sample.das      248    1412
#> 34 4.0  149  125  208   <NA>      TRUE       20 das_sample.das      252    1413
#> 35 4.0  149  125  208   <NA>      TRUE       20 das_sample.das      252    1413
#> 36 4.0  149  125  208   <NA>      TRUE       20 das_sample.das      252    1413
#> 37 4.0  149  125  208   <NA>      TRUE       20 das_sample.das      252    1413
#>    Subgroup SightNoDaily  Obs ObsStd Bearing Reticle DistNm Cue Method Photos
#> 1      <NA>   20130113_1  208   TRUE     309     2.8   1.06   3      4      N
#> 2      <NA>   20130113_1  208   TRUE     309     2.8   1.06   3      4      N
#> 3      <NA>   20130113_1  208   TRUE     309     2.8   1.06   3      4      N
#> 4      <NA>   20130113_2  125   TRUE     326     0.4   2.97   3      4      Y
#> 5      <NA>   20130113_2  125   TRUE     326     0.4   2.97   3      4      Y
#> 6      <NA>   20130113_2  125   TRUE     326     0.4   2.97   3      4      Y
#> 7      <NA>         <NA> <NA>  FALSE      11     2.0   1.30  NA     NA   <NA>
#> 8      <NA>         <NA> <NA>  FALSE       5     3.5   0.90  NA     NA   <NA>
#> 9      <NA>         <NA> <NA>  FALSE      50      NA   0.50  NA     NA   <NA>
#> 10     <NA>         <NA> <NA>  FALSE      71     4.5   0.70  NA     NA   <NA>
#> 11     <NA>         <NA> <NA>  FALSE     104     4.5   0.70  NA     NA   <NA>
#> 12     <NA>         <NA> <NA>  FALSE       2     2.2   1.30  NA     NA   <NA>
#> 13     <NA>         <NA>  280  FALSE     120      NA   0.03  NA     NA   <NA>
#> 14     <NA>         <NA>  149   TRUE     270      NA   0.03  NA     NA   <NA>
#> 15     <NA>         <NA>  228  FALSE     300      NA   0.02  NA     NA   <NA>
#> 16     <NA>   20130113_3  280   TRUE     270    14.0   0.28   3      4      N
#> 17     <NA>   20130113_3  280   TRUE     270    14.0   0.28   3      4      N
#> 18     <NA>   20130113_3  280   TRUE     270    14.0   0.28   3      4      N
#> 19     <NA>         <NA>  231  FALSE      45      NA   0.05  NA     NA   <NA>
#> 20     <NA>   20130113_4  149   TRUE     344     0.2   3.68   3      4      Y
#> 21     <NA>   20130113_4  149   TRUE     344     0.2   3.68   3      4      Y
#> 22     <NA>   20130113_4  149   TRUE     344     0.2   3.68   3      4      Y
#> 23     <NA>         <NA> <NA>  FALSE     356     0.4   3.00  NA     NA   <NA>
#> 24     <NA>   20130113_5  125   TRUE      70     1.4   1.66   3      4      Y
#> 25     <NA>   20130113_5  125   TRUE      70     1.4   1.66   3      4      Y
#> 26     <NA>   20130113_5  125   TRUE      70     1.4   1.66   3      4      Y
#> 27     <NA>   20130113_5  125   TRUE      70     1.4   1.66   3      4      Y
#> 28     <NA>         <NA>  280  FALSE      42    17.0   0.23  NA     NA   <NA>
#> 29     <NA>         <NA>  099  FALSE     180      NA   0.01  NA     NA   <NA>
#> 30     <NA>   20130114_1  280  FALSE       0      NA   0.00   3      1      N
#> 31     <NA>         <NA>  149   TRUE     309     1.7   1.47  NA     NA   <NA>
#> 32     <NA>   20130114_2  149   TRUE     359     0.3   3.28   2      4      Y
#> 33     <NA>   20130114_2  149   TRUE     359     0.3   3.28   2      4      Y
#> 34     <NA>   20130114_3  208   TRUE      38     0.8   2.23   3      4      Y
#> 35     <NA>   20130114_3  208   TRUE      38     0.8   2.23   3      4      Y
#> 36     <NA>   20130114_3  208   TRUE      38     0.8   2.23   3      4      Y
#> 37     <NA>   20130114_3  208   TRUE      38     0.8   2.23   3      4      Y
#>    Birds CalibSchool PhotosAerial Biopsy  Prob nSp Mixed ObsEstimate SpCode1
#> 1      N        <NA>         <NA>   <NA> FALSE   1 FALSE         280     018
#> 2      N        <NA>         <NA>   <NA> FALSE   1 FALSE         001     018
#> 3      N        <NA>         <NA>   <NA> FALSE   1 FALSE         208     018
#> 4      N        <NA>         <NA>   <NA> FALSE   1 FALSE         280     076
#> 5      N        <NA>         <NA>   <NA> FALSE   1 FALSE         001     076
#> 6      N        <NA>         <NA>   <NA> FALSE   1 FALSE         125     076
#> 7   <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 8   <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 9   <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 10  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 11  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 12  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 13  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 14  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 15  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 16     N        <NA>         <NA>   <NA> FALSE   1 FALSE         280     037
#> 17     N        <NA>         <NA>   <NA> FALSE   1 FALSE         001     037
#> 18     N        <NA>         <NA>   <NA> FALSE   1 FALSE         126     037
#> 19  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 20     Y        <NA>         <NA>   <NA> FALSE   1 FALSE         125     016
#> 21     Y        <NA>         <NA>   <NA> FALSE   1 FALSE         149     016
#> 22     Y        <NA>         <NA>   <NA> FALSE   1 FALSE         208     016
#> 23  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 24     N        <NA>         <NA>   <NA> FALSE   2  TRUE         280     013
#> 25     N        <NA>         <NA>   <NA> FALSE   2  TRUE         125     013
#> 26     N        <NA>         <NA>   <NA> FALSE   2  TRUE         149     013
#> 27     N        <NA>         <NA>   <NA> FALSE   2  TRUE         126     013
#> 28  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 29  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 30     N        <NA>         <NA>   <NA> FALSE   1 FALSE         280     075
#> 31  <NA>        <NA>         <NA>   <NA>    NA  NA    NA        <NA>    <NA>
#> 32     N        <NA>         <NA>   <NA> FALSE   2  TRUE         149     018
#> 33     N        <NA>         <NA>   <NA> FALSE   2  TRUE         126     018
#> 34     N        <NA>         <NA>   <NA>  TRUE   2  TRUE         125     016
#> 35     N        <NA>         <NA>   <NA>  TRUE   2  TRUE         208     016
#> 36     N        <NA>         <NA>   <NA>  TRUE   2  TRUE         149     016
#> 37     N        <NA>         <NA>   <NA>  TRUE   2  TRUE         126     016
#>    SpCode2 SpCode3 SpCode4 SpCodeProb1 SpCodeProb2 SpCodeProb3 SpCodeProb4
#> 1     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 2     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 3     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 4     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 5     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 6     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 7     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 8     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 9     <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 10    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 11    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 12    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 13    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 14    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 15    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 16    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 17    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 18    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 19    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 20    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 21    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 22    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 23    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 24     016    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 25     016    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 26     016    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 27     016    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 28    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 29    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 30    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 31    <NA>    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 32     277    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 33     277    <NA>    <NA>        <NA>        <NA>        <NA>        <NA>
#> 34     277    <NA>    <NA>         016         016        <NA>        <NA>
#> 35     277    <NA>    <NA>         016         016        <NA>        <NA>
#> 36     277    <NA>    <NA>         016         016        <NA>        <NA>
#> 37     277    <NA>    <NA>         016         016        <NA>        <NA>
#>    SpPerc1 SpPerc2 SpPerc3 SpPerc4 GsSchoolBest GsSchoolHigh GsSchoolLow
#> 1      100      NA      NA      NA           NA           NA          43
#> 2      100      NA      NA      NA           NA           NA          36
#> 3      100      NA      NA      NA           NA           NA          48
#> 4      100      NA      NA      NA            6           10           6
#> 5      100      NA      NA      NA            9           10           2
#> 6      100      NA      NA      NA            9           22           9
#> 7       NA      NA      NA      NA           NA           NA          NA
#> 8       NA      NA      NA      NA           NA           NA          NA
#> 9       NA      NA      NA      NA           NA           NA          NA
#> 10      NA      NA      NA      NA           NA           NA          NA
#> 11      NA      NA      NA      NA           NA           NA          NA
#> 12      NA      NA      NA      NA           NA           NA          NA
#> 13      NA      NA      NA      NA           NA           NA          NA
#> 14      NA      NA      NA      NA           NA           NA          NA
#> 15      NA      NA      NA      NA           NA           NA          NA
#> 16     100      NA      NA      NA           11           24          11
#> 17     100      NA      NA      NA           12           23          12
#> 18     100      NA      NA      NA            9           13           9
#> 19      NA      NA      NA      NA           NA           NA          NA
#> 20     100      NA      NA      NA           46           90          46
#> 21     100      NA      NA      NA           28           65          28
#> 22     100      NA      NA      NA           66           82          66
#> 23      NA      NA      NA      NA           NA           NA          NA
#> 24      68      32      NA      NA           37           72          37
#> 25      75      25      NA      NA           35           74          35
#> 26      65      35      NA      NA           29           52          29
#> 27      80      20      NA      NA           66           93          66
#> 28      NA      NA      NA      NA           NA           NA          NA
#> 29      NA      NA      NA      NA           NA           NA          NA
#> 30     100      NA      NA      NA           NA           NA          NA
#> 31      NA      NA      NA      NA           NA           NA          NA
#> 32      80      20      NA      NA          183          328         183
#> 33      90      10      NA      NA          120          170         120
#> 34      60      40      NA      NA           21           60          21
#> 35      56      44      NA      NA           16           20          16
#> 36      70      30      NA      NA           12           18          12
#> 37      98       2      NA      NA           36           53          36
#>    CourseSchool TurtleSp TurtleGs TurtleJFR TurtleAge TurtleCapt PinnipedSp
#> 1            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 2            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 3            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 4            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 5            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 6            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 7            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 8            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 9            NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 10          100     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 11          100     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 12           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 13           NA       LV        1      <NA>         A          N       <NA>
#> 14           NA       DC        2      <NA>         A          N       <NA>
#> 15           NA       DC        1      <NA>         J          N       <NA>
#> 16           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 17           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 18           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 19           NA       DC        1      <NA>         A       <NA>       <NA>
#> 20           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 21           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 22           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 23           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 24           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 25           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 26           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 27           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 28           NA       DC        2         F         A          Y       <NA>
#> 29           NA       LV        1      <NA>         A          N       <NA>
#> 30           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 31           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 32           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 33           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 34           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 35           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 36           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#> 37           NA     <NA>       NA      <NA>      <NA>       <NA>       <NA>
#>    PinnipedGs BoatType BoatGs   PerpDistKm
#> 1          NA     <NA>     NA 1.525631e+00
#> 2          NA     <NA>     NA 1.525631e+00
#> 3          NA     <NA>     NA 1.525631e+00
#> 4          NA     <NA>     NA 3.075807e+00
#> 5          NA     <NA>     NA 3.075807e+00
#> 6          NA     <NA>     NA 3.075807e+00
#> 7          NA     <NA>     NA 4.593917e-01
#> 8          NA     <NA>     NA 1.452712e-01
#> 9          NA     <NA>     NA 7.093572e-01
#> 10         NA     <NA>     NA 1.225770e+00
#> 11         NA     <NA>     NA 1.257891e+00
#> 12         NA     <NA>     NA 8.402403e-02
#> 13         NA     <NA>     NA 4.811637e-02
#> 14         NA     <NA>     NA 5.556000e-02
#> 15         NA     <NA>     NA 3.207758e-02
#> 16         NA     <NA>     NA 5.185600e-01
#> 17         NA     <NA>     NA 5.185600e-01
#> 18         NA     <NA>     NA 5.185600e-01
#> 19         NA     <NA>     NA 6.547809e-02
#> 20         NA     <NA>     NA 1.878568e+00
#> 21         NA     <NA>     NA 1.878568e+00
#> 22         NA     <NA>     NA 1.878568e+00
#> 23         NA     <NA>     NA 3.875670e-01
#> 24         NA     <NA>     NA 2.888916e+00
#> 25         NA     <NA>     NA 2.888916e+00
#> 26         NA     <NA>     NA 2.888916e+00
#> 27         NA     <NA>     NA 2.888916e+00
#> 28         NA     <NA>     NA 2.850229e-01
#> 29         NA     <NA>     NA 2.268046e-18
#> 30         NA     <NA>     NA 0.000000e+00
#> 31         NA     <NA>     NA 2.115733e+00
#> 32         NA     <NA>     NA 1.060157e-01
#> 33         NA     <NA>     NA 1.060157e-01
#> 34         NA     <NA>     NA 2.542657e+00
#> 35         NA     <NA>     NA 2.542657e+00
#> 36         NA     <NA>     NA 2.542657e+00
#> 37         NA     <NA>     NA 2.542657e+00