DAS sightings

Extract sightings and associated information from processed DAS data

das_sight(x, mixed.multi)

# S3 method for data.frame
das_sight(x, mixed.multi)

# S3 method for das_df
das_sight(x, mixed.multi)

Arguments

x
das_df object; output from das_process

mixed.multi
logical; indicates if mixed-species sightings should be output in multiple rows

Value

Data frame with ... Associated information (observer, species, etc.) is extracted from Data# columns as specified in Details. Other columns from x will be included in the output

Details

This function requires that the following assumptions be true:

All (and only) "S", "K", and "M" events are immediately followed by an "A" event

TODO

Examples

y <- system.file("das_sample.das", package = "swfscDAS")
y.proc <- das_process(y)

das_sight(y.proc)

#>    Event            DateTime      Lat       Lon OnEffort Cruise Mode EffType
#> 1      S 2014-07-02 12:17:00 27.53617 -125.5988     TRUE   1000    C       S
#> 2      S 2014-07-02 13:07:16 27.51750 -125.7502     TRUE   1000    C       S
#> 3      K 2014-07-02 16:28:37 27.37233 -126.3248     TRUE   1000    C       S
#> 4      k 2014-07-02 16:37:44 27.36700 -126.3510    FALSE   1000    C       S
#> 5      S 2014-07-02 16:47:22 27.36317 -126.3718     TRUE   1000    C       S
#> 6      s 2014-07-02 16:58:43 27.34817 -126.3832    FALSE   1000    C       S
#> 7      s 2014-07-02 17:21:01 27.32683 -126.3672    FALSE   1000    C       S
#> 8      s 2014-07-02 17:29:44 27.31633 -126.3655    FALSE   1000    C       S
#> 9      s 2014-07-02 17:50:15 27.32083 -126.3608    FALSE   1000    C       S
#> 10     F 2014-07-02 17:51:13 27.32200 -126.3625    FALSE   1000    C       S
#> 11     s 2014-07-02 18:10:25 27.32800 -126.3790    FALSE   1000    C       S
#> 12     s 2014-07-02 18:22:24 27.32600 -126.3650    FALSE   1000    C       S
#> 13     S 2014-07-03 06:53:54 27.08883 -127.7702     TRUE   1000    C       S
#>    Course Bft SwellHght RainFog HorizSun VertSun Glare Vis Data1 Data2 Data3
#> 1     259   3         4       1        4       1 FALSE   6   079   073     3
#> 2     259   3         4       1        3       1 FALSE   7   080   001     3
#> 3     252   3         4       1       NA      NA    NA   6  1014   186     6
#> 4     252   3         4       1       NA      NA    NA   6  1014   315   1.0
#> 5     252   3         4       3       NA      NA    NA   5   081   126     6
#> 6     252   3         4       3       NA      NA    NA   5   081   330   2.0
#> 7     125   3         4       3       NA      NA    NA   5   081   329   6.0
#> 8     125   3         4       3       NA      NA    NA   5   081   320   9.0
#> 9     125   3         4       3       NA      NA    NA   5   081   280  15.0
#> 10    290   3         4       3       NA      NA    NA   5   126   000   2.7
#> 11    175   3         4       3       NA      NA    NA   5   081   341   2.5
#> 12    105   3         4       3       NA      NA    NA   5   081   000  <NA>
#> 13    259   3         6       1       NA      NA    NA   5   082   153     6
#>    Data4 Data5 Data6 Data7 Data8    Data9       file_das event_num line_num
#> 1      4   012   0.2   3.6   210      4.0 das_sample.das        14       14
#> 2      4   025   0.6   2.5   330      2.0 das_sample.das        29       31
#> 3      4   340   1.5   2.2  <NA>     <NA> das_sample.das        82       85
#> 4    2.7   257  <NA>  <NA>  <NA>     <NA> das_sample.das        92       96
#> 5      1   080   2.0   1.3   230 2.0 1014 das_sample.das        99      103
#> 6    1.3   252  <NA>  <NA>  <NA>     <NA> das_sample.das       104      114
#> 7    0.6   125  <NA>  <NA>  <NA>     <NA> das_sample.das       108      118
#> 8    0.4   125  <NA>  <NA>  <NA>     <NA> das_sample.das       110      120
#> 9    0.3   125  <NA>  <NA>  <NA>     <NA> das_sample.das       113      123
#> 10   0.5  <NA>  <NA>  <NA>  <NA>     <NA> das_sample.das       115      125
#> 11   1.1   175  <NA>  <NA>  <NA>     <NA> das_sample.das       119      129
#> 12   0.5   105  <NA>  <NA>  <NA>     <NA> das_sample.das       122      132
#> 13     1   070   2.2   1.2   090      4.0 das_sample.das        10      145
#>     Obs Bearing Reticle Dist_nm SightNo Cue Method Photos Birds Mixed Spp1 Spp2
#> 1   073     012     0.2     3.6     079   3      4      N     N FALSE  075 <NA>
#> 2   001     025     0.6     2.5     080   3      4      N     N  TRUE  018  016
#> 3   186     340     1.5     2.2    1014   6      4      N     N FALSE  079 <NA>
#> 4  <NA>     315     1.0     2.7    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 5   126     080     2.0     1.3     081   6      1      N     N FALSE  079 <NA>
#> 6  <NA>     330     2.0     1.3    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 7  <NA>     329     6.0     0.6    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 8  <NA>     320     9.0     0.4    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 9  <NA>     280    15.0     0.3    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 10  126     000     0.5     2.7    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 11 <NA>     341     2.5     1.1    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 12 <NA>     000    <NA>     0.5    <NA>  NA     NA   <NA>  <NA>    NA <NA> <NA>
#> 13  153     070     2.2     1.2     082   6      1      N     Y FALSE  096 <NA>
#>    Spp3 Spp4  Prob TotalGS_best Sp1_perc Sp2_perc Sp3_perc Sp4_perc
#> 1  <NA> <NA>  TRUE     3.000000      100        0        0        0
#> 2  <NA> <NA> FALSE    20.000000       60       40        0        0
#> 3  <NA> <NA> FALSE     3.000000      100      NaN      NaN      NaN
#> 4  <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 5  <NA> <NA> FALSE     4.833333      100        0        0        0
#> 6  <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 7  <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 8  <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 9  <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 10 <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 11 <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 12 <NA> <NA>    NA           NA       NA       NA       NA       NA
#> 13 <NA> <NA> FALSE     4.857143      100        0        0        0
#>    resight_course turtle_sp turtle_num turtle_jfr turtle_age turtle_captured
#> 1              NA      <NA>         NA       <NA>       <NA>              NA
#> 2              NA      <NA>         NA       <NA>       <NA>              NA
#> 3              NA      <NA>         NA       <NA>       <NA>              NA
#> 4             257      <NA>         NA       <NA>       <NA>              NA
#> 5              NA      <NA>         NA       <NA>       <NA>              NA
#> 6             252      <NA>         NA       <NA>       <NA>              NA
#> 7             125      <NA>         NA       <NA>       <NA>              NA
#> 8             125      <NA>         NA       <NA>       <NA>              NA
#> 9             125      <NA>         NA       <NA>       <NA>              NA
#> 10             NA      <NA>         NA       <NA>       <NA>              NA
#> 11            175      <NA>         NA       <NA>       <NA>              NA
#> 12            105      <NA>         NA       <NA>       <NA>              NA
#> 13             NA      <NA>         NA       <NA>       <NA>              NA
#>    boat_type boat_num
#> 1       <NA>       NA
#> 2       <NA>       NA
#> 3       <NA>       NA
#> 4       <NA>       NA
#> 5       <NA>       NA
#> 6       <NA>       NA
#> 7       <NA>       NA
#> 8       <NA>       NA
#> 9       <NA>       NA
#> 10      <NA>       NA
#> 11      <NA>       NA
#> 12      <NA>       NA
#> 13      <NA>       NA