Convert randpicks file

Convert randpicks file from segchopr format to swfscDAS format

Usage

randpicks_convert(x.randpicks, x.segdata, seg.km)

Arguments

x.randpicks

Data frame with two columns; randpick values formatted for segchopr that correspond to x.segdata

x.segdata

Data frame; segdata that corresponds to x.randpicks

seg.km

numeric; target segment length used when creating x.segdata

Value

Data frame with one line for each continuous effort section in x.segdata, and two columns: effort_section and randpicks

Details

Past DAS processing code (segchopr) only recorded the generated random values, whereas swfscDAS randpicks files contain one line for each continuous effort section. See das_chop_equallength for more details about the swfscDAS randpicks format. This function 'converts' a randpicks data frame generated by segchopr to a data frame that meets the swfscDAS randpicks format requirements